% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrisk_var.R
\name{attrisk_var}
\alias{attrisk_var}
\title{Compute the Variance Estimate for Attributable Risk}
\usage{
attrisk_var(
  response,
  stressor,
  response_levels,
  stressor_levels,
  wgt,
  x,
  y,
  stratum_ind,
  stratum_level,
  cluster_ind,
  cluster,
  wgt1,
  x1,
  y1,
  pcfactor_ind,
  fpcsize,
  Ncluster,
  stage1size,
  vartype,
  warn_ind,
  warn_df,
  warn_vec
)
}
\arguments{
\item{response}{Vector of the categorical response variable.}

\item{stressor}{Vector of the categorical stressor variable.}

\item{response_levels}{Vector of category values (levels) for the
categorical response variable.  If response_levels equals NULL, then values
"Poor" and "Good" are used for the first level and second level of the
response variable, respectively.  The default is NULL.}

\item{stressor_levels}{Vector of category values (levels) for the
categorical stressor variable.  If stressor_levels equals NULL, then values
"Poor" and "Good" are used for the first level and second level of the
stressor variable, respectively.  The default is NULL.}

\item{wgt}{Vector of the final adjusted weight (inverse of the sample
inclusion probability) for each site, which is either the weight for a
single-stage sample or the stage two weight for a two-stage sample.}

\item{x}{Vector of x-coordinate for location for each site, which is either
the x- coordinate for a single-stage sample or the stage two x-coordinate
for a two-stage sample.}

\item{y}{Vector of y-coordinate for location for each site, which is either
the y- coordinate for a single-stage sample or the stage two y-coordinate
for a two-stage sample.}

\item{stratum_ind}{Logical value that indicates whether the sample is
stratified, where TRUE = a stratified sample and FALSE = not a stratified
sample.}

\item{stratum_level}{Vector of the stratum levels for the sites.}

\item{cluster_ind}{Logical value that indicates whether the sample is a
two- stage sample, where TRUE = a two-stage sample and FALSE = not a
two-stage sample.}

\item{cluster}{Vector of the stage one sampling unit (primary sampling unit
or cluster) code for each site.}

\item{wgt1}{Vector of the final adjusted stage one weight for each site.}

\item{x1}{Vector of the stage one x-coordinate for location for each site.}

\item{y1}{Vector of the stage one y-coordinate for location for each site.}

\item{pcfactor_ind}{Logical value that indicates whether the finite
population correction factor is used during variance estimation, where TRUE
= use the population correction factor and FALSE = do not use the factor.
To employ the correction factor for a single-stage sample, a value must be
supplied for argument fpcsize.  To employ the correction factor for a
two-stage sample, values must be supplied for arguments Ncluster and
stage1size.}

\item{fpcsize}{Size of the resource, which is required for calculation of the
finite population correction factor for a single-stage sample.}

\item{Ncluster}{The number of stage one sampling units in the resource,
which is required for calculation of the finite population correction
factor for a two-stage sample.}

\item{stage1size}{Vector of the size of the stage one sampling units of a
two-stage sample, which is required for calculation of the finite
population correction factor for a two-stage sample.}

\item{vartype}{The choice of variance estimator, where "Local" = local mean
estimator and "SRS" = SRS estimator.}

\item{warn_ind}{Logical value that indicates whether warning messages were
generated, where TRUE = warning messages were generated and FALSE = warning
messages were not generated.}

\item{warn_df}{A data frame for storing warning messages.}

\item{warn_vec}{Vector that contains names of the population type, the
subpopulation, and an indicator.}
}
\value{
A list containing the following objects:
  \describe{
    \item{\code{varest}}{vector containing variance estimates}
    \item{\code{warn_ind}}{logical variable that indicates whether warning
      messages were generated}
    \item{\code{warn_df}}{data frame for storing warning messages}
  }
}
\description{
This function calculates the variance-covariance estimate for the cell totals
used to calculate the attributable risk estimate.  Either the simple random
sampling (SRS) variance estimator or the local mean variance estimator is
calculated, which is subject to user control.  The SRS variance estimator
uses the independent random sample approximation to calculate joint inclusion
probabilities.  The function can  accomodate single-stage and two-stage
samples.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{localmean_cov}}}{calculate the variance/covariance
      matrix using the local mean estimator}
    \item{\code{\link{localmean_weight}}}{calculate the weighting matrix for
       the local mean variance estimator}
  }
}

\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
