% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootfcn.R
\name{bootfcn}
\alias{bootfcn}
\title{Bootstrap Function for Trend Parameter Estimation}
\usage{
bootfcn(dframe, indices, ivar, siteID, yearID)
}
\arguments{
\item{dframe}{Data frame containing survey design variables, response
variables, and subpopulation (domain) variables.}

\item{indices}{Vector of indices referencing rows in the dframe data frame.}

\item{ivar}{Character value providing name of the response variable in the
dframe data frame.}

\item{siteID}{Character value providing name of the site ID variable in
the dframe data frame.}

\item{yearID}{Character value providing name of the year variable in
the dframe data frame.}
}
\description{
This function calculates trend parameter estimates using bootstrap replcates.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
