% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_localmean_total.R
\name{cat_localmean_total}
\alias{cat_localmean_total}
\title{Local Mean Variance Estimates of Estimated Totals for Categorical Data}
\usage{
cat_localmean_total(
  itype,
  lev_itype,
  nlev_itype,
  ivar,
  lev_ivar,
  nlev_ivar,
  design,
  design_names,
  rslt_U,
  popcorrect,
  vartype,
  mult,
  warn_ind,
  warn_df
)
}
\arguments{
\item{itype}{Character value that identifies a factor variable in the design
argument containing subpopulation (domain) values.}

\item{lev_itype}{Character vector that provides levels of the subpopulation
variable.}

\item{nlev_itype}{Numeric value that provides the number of levels of the
subpopulation variable.}

\item{ivar}{Character value that identifies a factor variable in the design
argument containing categorical response values.}

\item{lev_ivar}{Character vector that provides levels of the categorical
response variable.}

\item{nlev_ivar}{Numeric value that provides the number of levels of the
categorical response variable.}

\item{design}{Object of class \code{survey.design} that specifies a complex survey
design.}

\item{design_names}{Character vector that provides names of survey design
variables in the design argument.}

\item{rslt_U}{Data frame that provides size estimates for the categorical
response variable.}

\item{popcorrect}{Logical value that indicates whether the finite population
correction factor should be employed during variance estimation.}

\item{vartype}{The choice of variance estimator, where \code{"Local"} = local mean
estimator and \code{"SRS"} = SRS estimator.}

\item{mult}{Numeric value that provides the Normal distribution confidence
bound multiplier.}

\item{warn_ind}{Logical value that indicates whether warning messages were
generated.}

\item{warn_df}{Data frame for storing warning messages.}
}
\value{
A list containing the following objects:
  \describe{
    \item{\code{stderr_U}}{data frame containing standard error estimates}
    \item{\code{confval_U}}{data frame containing confidence bound estimates}
    \item{\code{warn_ind}}{logical variable that indicates whether warning
      messages were generated}
    \item{\code{warn_df}}{data frame for storing warning messages}
  }
}
\description{
This function organizes input and output for calculation of the local mean
variance estimator for estimated sizes (totals) for categorical data.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{catvar_total}}}{calculates variance estimates of the
      estimated size in each of a set of categories}
    \item{\code{\link{svytotal}}}{calculates the total for a complex survey
      design}
  }
}

\seealso{
\code{\link{cat_localmean_prop}}
  \code{\link{svytotal}}
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
