% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_est.R
\name{category_est}
\alias{category_est}
\title{Category Proportion and Total Estimates for Probability Survey Data}
\usage{
category_est(
  catsum,
  dframe,
  itype,
  lev_itype,
  nlev_itype,
  ivar,
  lev_ivar,
  nlev_ivar,
  design,
  design_names,
  popcorrect,
  vartype,
  conf,
  mult,
  warn_ind,
  warn_df
)
}
\arguments{
\item{catsum}{Data frame containing estimates.}

\item{dframe}{Data frame containing survey design variables, response
variables, and subpopulation (domain) variables.}

\item{itype}{Character value that identifies a factor variable in the design
argument containing subpopulation (domain) values.}

\item{lev_itype}{Character vector that provides levels of the subpopulation
variable.}

\item{nlev_itype}{Numeric value that provides the number of levels of the
subpopulation variable.}

\item{ivar}{Character value that identifies a factor variable in the design
argument containing categorical response values.}

\item{lev_ivar}{Character vector that provides levels of the categorical
response variable.}

\item{nlev_ivar}{Numeric value that provides the number of levels of the
categorical response variable.}

\item{design}{Object of class \code{survey.design} that specifies a complex survey
design.}

\item{design_names}{Character vector that provides names of survey design
variables in the design argument.}

\item{popcorrect}{Logical value that indicates whether the finite population
correction factor should be employed during variance estimation.}

\item{vartype}{The choice of variance estimator, where \code{"Local"} = local mean
estimator and \code{"SRS"} = SRS estimator.}

\item{conf}{Numeric value for the confidence level.}

\item{mult}{Numeric value that provides the Normal distribution confidence
bound multiplier.}

\item{warn_ind}{Logical value that indicates whether warning messages were
generated.}

\item{warn_df}{Data frame for storing warning messages.}
}
\value{
A list composed of the following objects:
  \describe{
    \item{\code{catsum}}{data frame containing the category estimates}
    \item{\code{warn_ind}}{logical variable that indicates whether warning
      messages were generated}
    \item{\code{warn_df}}{data frame for storing warning messages}
  }
}
\description{
This function calculates estimates of the proportion (expressed as percent)
and total (size) for each of the categories of a categorical response
variable.  Upper and lower confidence bounds also are estimated.  Proportion
estimates are calculated using the Horvitz-Thompson ratio estimator, i.e.,
the ratio of two Horvitz-Thompson estimators.  The numerator of the ratio
estimates the total of the category.  The denominator of the ratio estimates
the total of the resource. Variance estimates for the proportion estimates
are calculated using either the local mean variance estimator or the simple
random sampling (SRS) variance estimator.  The choice of variance estimator
is subject to user control. The local mean variance estimator requires the
x-coordinate and the y-coordinate of each site.  Confidence bounds are
calculated using a Normal distribution multiplier. For a finite resource,
total is the number of units in the resource.  For an extensive resource,
total is the measure (extent) of the resource, i.e., length, area, or volume.
Total estimates are calculated using the Horvitz-Thompson estimator.
Variance estimates for the total estimates are calculated using either the
local mean variance estimator or the SRS variance estimator.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{cat_localmean_prop}}}{organizes input and output for
      calculation of the local mean variance estimator for estimated
      proportions for categorical data}
    \item{\code{\link{cat_localmean_total}}}{organizes input and output for
      calculation of the local mean variance estimator for estimated sizes
      (totals) for categorical data}
    \item{\code{\link{SE}}}{extracts standard errors from a survey design
      object}
    \item{\code{\link{confint}}}{computes confidence intervals for a survey
      design object}
    \item{\code{\link{svyby}}}{Compute survey statistics on subsets of a
      survey defined by factors}
    \item{\code{\link{svymean}}}{calculates the mean for a complex survey
      design}
    \item{\code{\link{svytotal}}}{calculates the total for a complex survey
      design}
  }
}

\seealso{
\code{\link{cat_localmean_prop}}
  \code{\link{cat_localmean_total}}
  \code{\link{confint}}
  \code{\link{SE}}
  \code{\link{svyby}}
  \code{\link{svymean}}
  \code{\link{svytotal}}
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
\keyword{univar}
