% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catvar_prop.R
\name{catvar_prop}
\alias{catvar_prop}
\title{Variance Estimates of Estimated Proportions for Categorical Data}
\usage{
catvar_prop(
  z,
  wgt,
  x,
  y,
  prop,
  prop_names,
  stratum_ind,
  stratum_level,
  cluster_ind,
  clusterID,
  wgt1,
  x1,
  y1,
  pcfactor_ind,
  fpcsize,
  Ncluster,
  stage1size,
  vartype,
  warn_ind,
  warn_df,
  warn_vec
)
}
\arguments{
\item{z}{Vector of the value of the categorical response variable or the
site status for each site.}

\item{wgt}{Vector of the final adjusted weight (inverse of the sample
inclusion probability) for each site, which is either the weight for a
single-stage sample or the stage two weight for a two-stage sample.}

\item{x}{Vector of the x-coordinate for location for each site, which is
either the x-coordinate for a single-stage sample or the stage two
x-coordinate for a two-stage sample.}

\item{y}{Vector of the y-coordinate for location for each site, which is
either the y-coordinate for a single-stage sample or the stage two
y-coordinate for a two-stage sample.}

\item{prop}{Vector of the proportion estimates.}

\item{prop_names}{Vector of the category names for the response variable plus
\code{"Total"}.}

\item{stratum_ind}{Logical value that indicates whether the sample is
stratified, where \code{TRUE} = a stratified sample and \code{FALSE} = not a stratified
sample.}

\item{stratum_level}{The stratum level.}

\item{cluster_ind}{Logical value that indicates whether the sample is a
two-stage sample, where \code{TRUE} = a two-stage sample and \code{FALSE} = not a
two-stage sample.}

\item{clusterID}{Vector of the stage one sampling unit (primary sampling unit
or cluster) code for each site.}

\item{wgt1}{Vector of the final adjusted stage one weight for each site.}

\item{x1}{Vector of the stage one x-coordinate for location for each site.}

\item{y1}{Vector of the stage one y-coordinate for location for each site.}

\item{pcfactor_ind}{Logical value that indicates whether the finite
population correction factor is used during variance estimation, where \code{TRUE}
= use the population correction factor and \code{FALSE} = do not use the factor.
To employ the correction factor for a single-stage sample, a value must be
supplied for argument \code{fpcsize}.  To employ the correction factor for a
two-stage sample, values must be supplied for arguments \code{Ncluster} and
\code{stage1size}.}

\item{fpcsize}{Size of the resource, which is required for calculation of the
finite population correction factor for a single-stage sample.}

\item{Ncluster}{The number of stage one sampling units in the resource,
which is required for calculation of the finite population correction
factor for a two-stage sample.}

\item{stage1size}{Vector of the size of the stage one sampling units of a
two-stage sample, which is required for calculation of the finite
population correction factor for a two-stage sample.}

\item{vartype}{The choice of variance estimator, where \code{"Local"} = local mean
estimator and \code{"SRS"} = SRS estimator.}

\item{warn_ind}{Logical value that indicates whether warning messages were
generated, where \code{TRUE} = warning messages were generated and \code{FALSE} = warning
messages were not generated.}

\item{warn_df}{Data frame for storing warning messages.}

\item{warn_vec}{Vector that contains names of the subpopulation type, the
subpopulation, and the response variable.}
}
\value{
A list containing the following objects:
  \describe{
    \item{\code{vartype}}{character variable containing the type of variance
      estimator}
    \item{\code{varest}}{vector containing variance estimates}
    \item{\code{warn_ind}}{logical variable that indicates whether warning
      messages were generated}
    \item{\code{warn_df}}{data frame for storing warning messages}
  }
}
\description{
This function calculates variance estimates of the estimated proportion in
each of a set of categories.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{localmean_weight}}}{calculate the weighting matrix for
      the local mean variance estimator}
    \item{\code{\link{localmean_var}}}{calculate the local mean variance
      estimator}
  }
}

\author{
Tom Kincaid \email{kincaid.tom@epa.gov}
}
\keyword{survey}
