% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catvar_total.R
\name{catvar_total}
\alias{catvar_total}
\title{Variance Estimates of the Estimated Size for Categorical Data}
\usage{
catvar_total(
  z,
  wgt,
  x,
  y,
  size_names,
  stratum_ind,
  stratum_level,
  cluster_ind,
  clusterID,
  wgt1,
  x1,
  y1,
  pcfactor_ind,
  fpcsize,
  Ncluster,
  stage1size,
  vartype,
  warn_ind,
  warn_df,
  warn_vec
)
}
\arguments{
\item{z}{Vector of the values of the categorical response variable or the
site status for each site.}

\item{wgt}{Vector of the final adjusted weight (inverse of the sample
inclusion probability) for each site, which is either the weight for a
single-stage sample or the stage two weight for a two-stage sample.}

\item{x}{Vector of the x-coordinates for location for each site, which is
either the x-coordinate for a single-stage sample or the stage two
x-coordinate for a two-stage sample.}

\item{y}{Vector of the y-coordinates for location for each site, which is
either the y-coordinate for a single-stage sample or the stage two
y-coordinate for a two-stage sample.}

\item{size_names}{Vector of the category names for the response variable plus
"Total".}

\item{stratum_ind}{Logical value that indicates whether the sample is
stratified, where TRUE = a stratified sample and FALSE = not a stratified
sample.}

\item{stratum_level}{Vector of the stratum for each site.}

\item{cluster_ind}{Logical value that indicates whether the sample is a
two-stage sample, where TRUE = a two-stage sample and FALSE = not a
two-stage sample.}

\item{clusterID}{Vector of the stage one sampling units (primary sampling
unit or cluster) code for each site.}

\item{wgt1}{Vector of the final adjusted stage one weight for each site.}

\item{x1}{Vector of the stage one x-coordinates for location for each site.}

\item{y1}{Vector of the stage one y-coordinates for location for each site.}

\item{pcfactor_ind}{Logical value that indicates whether the finite
population correction factor is used during variance estimation, where TRUE
= use the population correction factor and FALSE = do not use the factor.
To employ the correction factor for a single-stage sample, a value must be
supplied for argument fpcsize.  To employ the correction factor for a
two-stage sample, values must be supplied for arguments Ncluster and
stage1size.}

\item{fpcsize}{Size of the resource, which is required for calculation of the
finite population correction factor for a single-stage sample.}

\item{Ncluster}{The number of stage one sampling units in the resource,
which is required for calculation of the finite population correction
factor for a two-stage sample.}

\item{stage1size}{Vector of the size of the stage one sampling units of a
two-stage sample, which is required for calculation of the finite
population correction factor for a two-stage sample.}

\item{vartype}{The choice of variance estimator, where "Local" = local mean
estimator and "SRS" = SRS estimator.}

\item{warn_ind}{Logical value that indicates whether warning messages were
generated, where TRUE = warning messages were generated and FALSE = warning
messages were not generated.}

\item{warn_df}{A data frame for storing warning messages.}

\item{warn_vec}{A vector that contains names of the population type, the
subpopulation, and an indicator.}
}
\value{
A list containing the following objects:
  \describe{
    \item{\code{vartype}}{character variable containing the type of variance
      estimator}
    \item{\code{varest}}{vector containing variance estimates}
    \item{\code{warn_ind}}{logical variable that indicates whether warning
      messages were generated}
    \item{\code{warn_df}}{data frame for storing warning messages}
  }
}
\description{
This function calculates variance estimates of the estimated size in each of
a set of categories.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{localmean_weight}}}{calculate the weighting matrix for
      the local mean variance estimator}
    \item{\code{\link{localmean_var}}}{calculate the local mean variance
      estimator}
  }
}

\author{
Tom Kincaid \email{kincaid.tom@epa.gov}
}
\keyword{survey}
