% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.nresp.R
\name{cdf.nresp}
\alias{cdf.nresp}
\title{Internal Function: Count Number of Responses Less Than a Set of Values}
\usage{
cdf.nresp(z, val)
}
\arguments{
\item{z}{Vector of the response values.}

\item{val}{Vector of the set of values at which the CDF is estimated.}
}
\value{
Output is the number of response values for each CDF estimation
  value.
}
\description{
This function calculates the number of response values less than or equal to
each of the set of values at which the cumulative distribution function (CDF)
is estimated.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
