% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_est.R
\name{cdf_est}
\alias{cdf_est}
\title{Cumulative Distribution Function Estimate for Probability Survey Data}
\usage{
cdf_est(
  cdfsum,
  dframe,
  itype,
  lev_itype,
  nlev_itype,
  ivar,
  design,
  design_names,
  var_nondetect,
  popcorrect,
  vartype,
  conf,
  mult,
  warn_ind,
  warn_df
)
}
\arguments{
\item{cdfsum}{Data frame containing estimates.}

\item{dframe}{Data frame containing survey design variables, response
variables, and subpopulation (domain) variables.}

\item{itype}{Character value that identifies a factor variable in the design
argument containing subpopulation (domain) values.}

\item{lev_itype}{Character vector that provides levels of the subpopulation
variable.}

\item{nlev_itype}{Numeric value that provides the number of levels of the
subpopulation variable.}

\item{ivar}{Character value that identifies a factor variable in the design
argument containing categorical response values.}

\item{design}{Object of class survey.design that specifies a complex survey
design.}

\item{design_names}{Character vector that provides names of survey design
variables in the design argument.}

\item{var_nondetect}{Character value that identifies the name of a logical
variable in the dframe data frame specifying the presence of not detected
(nondetect) values for the response variable.}

\item{popcorrect}{Logical value that indicates whether the finite population
correction factor should be employed during variance estimation.}

\item{vartype}{The choice of variance estimator, where "Local" = local mean
estimator and "SRS" = SRS estimator.}

\item{conf}{Numeric value for the confidence level.}

\item{mult}{Numeric value that provides the Normal distribution confidence
bound multiplier.}

\item{warn_ind}{Logical value that indicates whether warning messages were
generated.}

\item{warn_df}{Data frame for storing warning messages.}
}
\value{
A list composed of the following objects:
  \describe{
    \item{\code{cdfsum}}{data frame containing the CDF estimate}
    \item{\code{warn_ind}}{logical variable that indicates whether warning
      messages were generated}
    \item{\code{warn_df}}{data frame for storing warning messages}
  }
}
\description{
This function calculates an estimate of the cumulative distribution function
(CDF) for the proportion (expressed as percent) and the total of a
quantitative response variable, where the response variable may be defined
for either a finite or an extensive resource.  In addition, the standard
error of the estimated CDF and confidence bounds are calculated.  For the
CDF of a proportion, the Horvitz-Thompson ratio estimator, i.e., the ratio
of two Horvitz-Thompson estimators, is used to calculate the CDF estimate.
For the CDF of a total, the classic ratio estimator is used to calculate the
CDF estimate, where that estimator is the product of the known value and the
Horvitz-Thompson ratio estimator.  Variance estimates for the estimated CDF
are calculated using either the local mean variance estimator or the simple
random sampling (SRS) variance estimator.  The choice of variance estimator
is subject to user control. The local mean variance estimator requires the
x-coordinate and the y-coordinate of each site.  The SRS variance estimator
uses the independent random sampling approximation to calculate joint
inclusion probabilities.  Confidence bounds are calculated using a Normal
distribution multiplier.  The function can accommodate a stratified sample.
For a stratified sample, separate estimates and standard errors are
calculated for each stratum, which are used to produce estimates and standard
errors for all strata combined. The function can accommodate single-stage and
two-stage samples for both stratified and unstratified sampling designs.  The
finite population correction factor can be utilized in variance estimation.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{cdf_localmean_prop}}}{organizes input and output for
      calculation of the local mean variance estimator for the estimated CDF
      using the proportion scale}
    \item{\code{\link{cdf_localmean_total}}}{organizes input and output for
      calculation of the local mean variance estimator for the estimated CDF
      using the total scale}
    \item{\code{\link{cdf_nresp}}}{calculates number of response values less
      than a set of values}
    \item{\code{\link{SE}}}{extracts standard errors from a survey design
      object}
    \item{\code{\link{confint}}}{computes confidence intervals for a survey
      design object}
    \item{\code{\link{svyby}}}{Compute survey statistics on subsets of a
      survey defined by factors}
    \item{\code{\link{svymean}}}{calculates the mean for a complex survey
      design}
    \item{\code{\link{svytotal}}}{calculates the total for a complex survey
      design}
  }
}

\seealso{
\code{\link{cdf_localmean_prop}}
  \code{\link{cdf_localmean_total}}
  \code{\link{confint}}
  \code{\link{SE}}
  \code{\link{svyby}}
  \code{\link{svymean}}
  \code{\link{svytotal}}
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
\keyword{univar}
