% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_nresp.R
\name{cdf_nresp}
\alias{cdf_nresp}
\title{Internal Function: Calculate Number of Response Values Less Than a Set of Values}
\usage{
cdf_nresp(z, val)
}
\arguments{
\item{z}{Vector of the response values.}

\item{val}{Vector of the set of values at which the CDF is estimated.}
}
\value{
Vector containing the number of response values for each CDF estimation
  value.
}
\description{
This function calculates the number of response values less than or equal to
each of the set of values at which the cumulative distribution function (CDF)
is estimated.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
