% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdftest_localmean_total.R
\name{cdftest_localmean_total}
\alias{cdftest_localmean_total}
\title{Local Mean Variance/Covarince Estimates of Estimated Population Totals}
\usage{
cdftest_localmean_total(
  design,
  design_names,
  popcorrect,
  vartype,
  warn_ind,
  warn_df,
  warn_vec
)
}
\arguments{
\item{design}{Object of class survey.design that specifies a complex survey
design.}

\item{design_names}{Character vector that provides names of survey design
variables in the design argument.}

\item{popcorrect}{Logical value that indicates whether the finite population
correction factor should be employed during variance estimation.}

\item{vartype}{The choice of variance estimator, where "Local" = local mean
estimator and "SRS" = SRS estimator.}

\item{warn_ind}{Logical value that indicates whether warning messages were
generated.}

\item{warn_df}{Data frame for storing warning messages.}

\item{warn_vec}{Character vector that contains a subpopulation name, the
first subpopulation level, the second subpopulation level, and an
indicator name.}
}
\value{
A matrix containing the variance/covariance estimates for the
  contingency table proportion estimates.
}
\description{
This function organizes input and output for calculation of the local mean
variance/covariance estimator for estimated totals for categorical data.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{cdftestvar_total}}}{calculates variance-covariance
      estimates of the estimated totals (cell totals, row totals, column
      totals, and grand total) of a contingency table}
  }
}

\seealso{
\code{\link{cdftest_localmean_prop}}
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
