% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdftestvar_prop.R
\name{cdftestvar_prop}
\alias{cdftestvar_prop}
\title{Local Mean Variance/Covariance Estimates of Estimated Population Proportions}
\usage{
cdftestvar_prop(
  design,
  wgt,
  x,
  y,
  stratum_ind,
  stratum_level,
  cluster_ind,
  clusterID,
  wgt1,
  x1,
  y1,
  pcfactor_ind,
  fpcsize,
  Ncluster,
  stage1size,
  vartype,
  warn_ind,
  warn_df,
  warn_vec
)
}
\arguments{
\item{design}{Object of class \code{survey.design} that specifies a complex survey
design.}

\item{wgt}{Vector of the final adjusted weight (reciprocal of the sample
inclusion probability) for each site, which is either the weight for a
single-stage sample or the stage two weight for a two-stage sample.}

\item{x}{Vector of x-coordinate for location for each site, which is either
the x- coordinate for a single-stage sample or the stage two x-coordinate
for a two-stage sample.}

\item{y}{Vector of y-coordinate for location for each site, which is either
the y- coordinate for a single-stage sample or the stage two y-coordinate
for a two-stage sample.}

\item{stratum_ind}{Logical value that indicates whether the sample is
stratified, where \code{TRUE} = a stratified sample and \code{FALSE} = not a stratified
sample.}

\item{stratum_level}{Vector indicating the stratum level.}

\item{cluster_ind}{Logical value that indicates whether the sample is a
two- stage sample, where \code{TRUE} = a two-stage sample and \code{FALSE} = not a
two-stage sample.}

\item{clusterID}{Vector of the stage one sampling unit (primary sampling unit
or cluster) code for each site.}

\item{wgt1}{Vector of the final adjusted stage one weight for each site.}

\item{x1}{Vector of the stage one x-coordinate for location for each site.}

\item{y1}{Vector of the stage one y-coordinate for location for each site.}

\item{pcfactor_ind}{Logical value that indicates whether the finite
population correction factor is used during variance estimation, where \code{TRUE}
= use the population correction factor and \code{FALSE} = do not use the factor.
To employ the correction factor for a single-stage sample, a value must be
supplied for argument fpcsize.  To employ the correction factor for a
two-stage sample, values must be supplied for arguments Ncluster and
stage1size.}

\item{fpcsize}{Size of the resource, which is required for calculation of the
finite population correction factor for a single-stage sample.}

\item{Ncluster}{The number of stage one sampling units in the resource,
which is required for calculation of the finite population correction
factor for a two-stage sample.}

\item{stage1size}{Vector of the size of the stage one sampling units of a
two-stage sample, which is required for calculation of the finite
population correction factor for a two-stage sample.}

\item{vartype}{The choice of variance estimator, where \code{"Local"} = local mean
estimator and \code{"SRS"} = SRS estimator.}

\item{warn_ind}{Logical value that indicates whether warning messages were
generated, where \code{TRUE} = warning messages were generated and \code{FALSE} = warning
messages were not generated.}

\item{warn_df}{A data frame for storing warning messages.}

\item{warn_vec}{Character vector that contains a subpopulation name, the
first subpopulation level, the second subpopulation level, and an
indicator name.}
}
\value{
Object in list format composed of a matrix named varest, which
  contains variance/covariance estimates, a logical variable named warn,ind,
  which is the indicator for warning messges, and a data frame named \code{warn_df},
  which contains warning messages.
}
\description{
This function calculates the local mean estimate of the variance/covariance
matrix of the population proportions for a contingency table.  The function
can accomodate single-stage and two-stage samples.  The finite population
correction factor can be utilized in variance estimation.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{localmean_weight}}}{calculate the weighting matrix for
      the local mean variance estimator}
    \item{\code{\link{localmean_cov}}}{calculate the variance/covariance
      matrix using the local mean estimator}
    \item{\code{\link{svymean}}}{calculates means for a complex survey
      design}
  }
}

\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
