% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellWeight.R
\name{cellWeight}
\alias{cellWeight}
\title{Total Inclusion Probablity for a Grid Cell}
\usage{
cellWeight(xc, yc, dx, dy, sfobject)
}
\arguments{
\item{xc}{x-coordinates that define the cells.}

\item{yc}{y-coordinates that define the cells.}

\item{dx}{The x-axis grid cell dimension.}

\item{dy}{The y-axis grid cell dimension.}

\item{sfobject}{the sf object containing the survey frame.}
}
\value{
Vector containing the total inclusion probability for each cell.
}
\description{
Calculates the total inclusion probability for each grid cell from a GRTS
survey design.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
