% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_est.R
\name{change_est}
\alias{change_est}
\title{Estimate Change between Two Surveys}
\usage{
change_est(
  resp_ind,
  survey_names,
  changesum,
  dframe,
  survey_1,
  survey_2,
  itype,
  isubpop,
  ivar,
  lev_ivar,
  nlev_ivar,
  design_1,
  design_2,
  design_names,
  repeat_1,
  repeat_2,
  siteID,
  revisitwgt,
  test,
  var_nondetect,
  popcorrect,
  vartype,
  conf,
  mult,
  warn_ind,
  warn_df,
  warn_vec
)
}
\arguments{
\item{resp_ind}{A character value that indicates the type of response
variable, where "cat" indicates a categorical variable and "cont" indicates
a continuous variable.}

\item{survey_names}{Character vector of length two that provides the survey
names contained in the survey ID variable in the dframe data frame.  The
two values in the vector identify the first survey and second survey,
respectively.}

\item{changesum}{List containing estimates, which is composed of three data
frames: catsum, contsum_mean, and contsum_median.}

\item{dframe}{Data frame containing survey design variables, response
variables, and subpopulation (domain) variables for both surveys.}

\item{survey_1}{Logical vector that identifies survey one sites in the dframe
data frame.}

\item{survey_2}{Logical vector that identifies survey two sites in the dframe
data frame.}

\item{itype}{Character value that identifies the factor variable containing
subpopulation (domain) values.}

\item{isubpop}{Character value that identifies a level of the subpopulation
variable.}

\item{ivar}{Character value that identifies the response variable.}

\item{lev_ivar}{Character vector that provides levels of a categorical
response variable.}

\item{nlev_ivar}{Numeric value that provides the number of levels of a
categorical response variable.}

\item{design_1}{Object of class survey.design that specifies the complex
survey design for survey one.}

\item{design_2}{Object of class survey.design that specifies the complex
survey design for survey two}

\item{design_names}{Character vector that provides names of survey design
variables in the design argument.}

\item{repeat_1}{Logical vector that identifies repeat visit sites for
survey one.}

\item{repeat_2}{Logical vector that identifies repeat visit sites for
survey two.}

\item{siteID}{Character value providing name of the site ID variable in
the dframe data frame.}

\item{revisitwgt}{Logical value that indicates whether each repeat visit
site has the same survey design weight in the two surveys, where TRUE = the
weight for each repeat visit site is the same and FALSE = the weight for
each repeat visit site is not the same.  When this argument is FALSE, all
of the repeat visit sites are assigned equal weights when calculating the
covariance component of the change estimate standard error.  The default is
FALSE.}

\item{test}{Character string or character vector providing the location
measure(s) to use for change estimation for continuous variables.  The
choices are "mean", "median", or c("mean", "median").}

\item{var_nondetect}{Character value that identifies the name of a logical
variable in the dframe data frame specifying the presence of not detected
(nondetect) values for a continuous response variable.}

\item{popcorrect}{Logical value that indicates whether the finite population
correction factor is used during variance estimation.}

\item{vartype}{The choice of variance estimator, where "Local" = local mean
estimator and "SRS" = SRS estimator.}

\item{conf}{Numeric value for the confidence level.}

\item{mult}{Numeric value that provides the Normal distribution confidence
bound multiplier.}

\item{warn_ind}{Logical value that indicates whether warning messages were
generated, where TRUE = warning messages were generated and FALSE = warning
messages were not generated.}

\item{warn_df}{Data frame for storing warning messages.}

\item{warn_vec}{Vector that contains names of the population type, the
subpopulation, and an indicator.}
}
\value{
A list composed of the following objects:
  \describe{
    \item{\code{results}}{if resp_ind equals "cat", a data frame named
      catsum containing change estimates for categories; if resp_ind equals
      "cont", two data frames named contsum_mean and contsum_median
      containing mean and median change estimates, respectively, as specified
      by the argument named test.}
    \item{\code{warn_ind}}{logical variable that indicates whether warning
      messages were generated}
    \item{\code{warn_df}}{data frame for storing warning messages}
  }
}
\description{
This function estimates change between two probability surveys.  The function
can accommodate both categorical and continuous response variables. For a
categorical response variable, change is estimated by the difference in
category estimates for the two surveys, where a category estimate is the
estimated proportion of values in a category.  Note that a separate change
estimate is calculated for each category of a categorical response variable.
For a continuous response variable, change can be estimated for the mean, the
median, or for both the mean and median. For a continuous response variable
using the mean, change is estimated by the difference in estimated mean
values for the two surveys.  For change estimates using the median, the first
step is to calculate an estimate of the median for the first survey. The
estimated median from the first survey is then used to define two categories:
(1) values that are less than or equal to the estimated median and (2) values
that are greater than the estimated median.  Once the categories are defined,
change analysis for the median is identical to change analysis for a
categorical variable, i.e., change is estimated by the difference in category
estimates for the two surveys. In addition to change estimates, the standard
error of the change estimates and confidence bounds are calculated.  Variance
estimates are calculated using either the local mean variance estimator or
the simple random sampling (SRS) variance estimator.  The choice of variance
estimator is subject to user control. The local mean variance estimator
requires the x-coordinate and y-coordinate of each site.  The SRS variance
estimator uses the independent random sample approximation to calculate joint
inclusion probabilities.  Confidence bounds are calculated using a Normal
distribution multiplier.  The function can accommodate a stratified sample.
For a stratified sample, separate estimates and standard errors are
calculated for each stratum, which are used to produce estimates and standard
errors for all strata combined. Strata that contain a single value are
removed.  For a stratified sample, when either the size of the resource or
the sum of the size-weights of the resource is provided for each stratum,
those values are used as stratum weights for calculating the estimates and
standard errors for all strata combined.  For a stratified sample when
neither the size of the resource nor the sum of the size-weights of the
resource is provided for each stratum, estimated values are used as stratum
weights for calculating the estimates and standard errors for all strata
combined.  The function can accommodate single-stage and two-stage samples
for both stratified and unstratified sampling designs.  It is assumed that
both surveys employ the same type of survey design.  Finite population and
continuous population correction factors can be utilized in variance
estimation.  The function checks for compatibility of input values and
removes missing values.
}
\section{Other Functions Required}{

 \describe{
    \item{\code{\link{category_est}}}{calculate category proportion and total
      estimates}
   \item{\code{\link{changevar_mean}}}{calculate the covariance or
     correlation estimate of the estimated change in means between two
     probability surveys}
   \item{\code{\link{changevar_prop}}}{calculate covariance or correlation
     estimates of the estimated change in class proportion estimates between
    two probability surveys}
   \item{\code{\link{changevar_total}}}{calculate covariance or correlation
     estimates of the estimated change in class total estimates between two
     probability surveys}
    \item{\code{\link{percentile_est}}}{calculates percentile estimates}
 }
}

\seealso{
\code{\link{category_est}}
  \code{\link{changevar_mean}}
  \code{\link{changevar_prop}}
  \code{\link{changevar_total}}
  \code{\link{confint}}
  \code{\link{percentile_est}}
  \code{\link{SE}}
  \code{\link{svyby}}
  \code{\link{svymean}}
  \code{\link{svytotal}}
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
