% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changevar_mean.R
\name{changevar_mean}
\alias{changevar_mean}
\title{Covariance or Correlation Matrix Estimate of Change in Means between Two
Surveys}
\usage{
changevar_mean(
  z1,
  z2,
  wgt,
  x,
  y,
  revisitwgt,
  mean1,
  mean2,
  stratum_ind,
  stratum_level,
  cluster_ind,
  clusterID,
  wgt1,
  x1,
  y1,
  pcfactor_ind,
  fpcsize,
  Ncluster,
  stage1size,
  vartype,
  warn_ind,
  warn_df,
  warn_vec
)
}
\arguments{
\item{z1}{Vector of the response value for each site for survey one.}

\item{z2}{Vector of the response value for each site for survey two.}

\item{wgt}{Vector of the final adjusted weight (reciprocal of the sample
inclusion probability) for each site, which is either the weight for a
single-stage sample or the stage two weight for a two-stage sample.}

\item{x}{Vector of x-coordinate for location for each site, which is either
the x- coordinate for a single-stage sample or the stage two x-coordinate
for a two-stage sample.}

\item{y}{Vector of y-coordinate for location for each site, which is either
the y- coordinate for a single-stage sample or the stage two y-coordinate
for a two-stage sample.}

\item{revisitwgt}{Logical value that indicates whether each repeat visit site
has the same survey design weight in the two surveys, where \code{TRUE} = the
weight for each repeat visit site is the same and \code{FALSE} = the weight for
each repeat visit site is not the same.  When this argument is \code{FALSE}, all
of the repeat visit sites are assigned equal weights when calculating the
covariance component of the change estimate standard error.}

\item{mean1}{The estimated mean for survey one.}

\item{mean2}{The estimated mean for survey two.}

\item{stratum_ind}{Logical value that indicates whether the sample is
stratified, where \code{TRUE} = a stratified sample and \code{FALSE} = not a stratified
sample.}

\item{stratum_level}{The stratum level.}

\item{cluster_ind}{Logical value that indicates whether the sample is a
two- stage sample, where \code{TRUE} = a two-stage sample and \code{FALSE} = not a
two-stage sample.}

\item{clusterID}{Vector of the stage one sampling unit (primary sampling unit
or cluster) code for each site.}

\item{wgt1}{Vector of the final adjusted stage one weight for each site.}

\item{x1}{Vector of the stage one x-coordinate for location for each site.}

\item{y1}{Vector of the stage one y-coordinate for location for each site.}

\item{pcfactor_ind}{Logical value that indicates whether the finite
population correction factor is used during variance estimation, where \code{TRUE}
= use the population correction factor and \code{FALSE} = do not use the factor.
To employ the correction factor for a single-stage sample, a value must be
supplied for argument \code{fpcsize}.  To employ the correction factor for a
two-stage sample, values must be supplied for arguments \code{Ncluster} and
\code{stage1size}.}

\item{fpcsize}{Size of the resource, which is required for calculation of the
finite population correction factor for a single-stage sample.}

\item{Ncluster}{The number of stage one sampling units in the resource,
which is required for calculation of the finite population correction
factor for a two-stage sample.}

\item{stage1size}{Vector of the size of the stage one sampling units of a
two-stage sample, which is required for calculation of the finite
population correction factor for a two-stage sample.}

\item{vartype}{The choice of variance estimator, where \code{"Local"} = local mean
estimator and \code{"SRS"} = SRS estimator.}

\item{warn_ind}{Logical value that indicates whether warning messages were
generated, where \code{TRUE} = warning messages were generated and \code{FALSE} = warning
messages were not generated.}

\item{warn_df}{Data frame for storing warning messages.}

\item{warn_vec}{Vector that contains names of the population type, the
subpopulation, and an indicator.}
}
\value{
An object in list format composed of a vector named \code{rslt}, which
  contains the covariance or correlation estimate, a logical variable named
  \code{warn_ind}, which is the indicator for warning messges, and a data frame
  named \code{warn_df}, which contains warning messages.
}
\description{
This function calculates estimates of the variance-covariance matrix of the
population proportions in a set of intervals (classes).  The set of values
defining upper bounds for the classes is supplied to the function. Either the
simple random sampling (SRS) variance estimator or the local mean variance
estimator is calculated, which is subject to user control.  The simple random
sampling variance estimator uses the independent random sample approximation
to calculate joint inclusion probabilities.  The function can accomodate
single-stage and two-stage samples.  Finite population and continuous
population correction factors can be utilized in variance estimation.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{localmean_weight}}}{calculate the weighting matrix for
      the local mean variance estimator}
    \item{\code{\link{localmean_cov}}}{calculate the variance/covariance
      matrix using the local mean estimator}
  }
}

\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
