% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changevar_prop.R
\name{changevar_prop}
\alias{changevar_prop}
\title{Covariance or Correlation Matrix Estimate of Change in Class Proportions
between Two Surveys}
\usage{
changevar_prop(
  catvar_levels,
  catvar1,
  catvar2,
  wgt,
  x,
  y,
  revisitwgt,
  prop1,
  prop2,
  stratum_ind,
  stratum_level,
  cluster_ind,
  clusterID,
  wgt1,
  x1,
  y1,
  pcfactor_ind,
  fpcsize,
  Ncluster,
  stage1size,
  vartype,
  warn_ind,
  warn_df,
  warn_vec
)
}
\arguments{
\item{catvar_levels}{Vector of the set of categorical response values.}

\item{catvar1}{Vector of the response value for each site for survey one.}

\item{catvar2}{Vector of the response value for each site for survey two.}

\item{wgt}{Vector of the final adjusted weight (inverse of the sample
inclusion probability) for each site, which is either the weight for a
single-stage sample or the stage two weight for a two-stage sample.}

\item{x}{Vector of x-coordinate for location for each site, which is either
the x- coordinate for a single-stage sample or the stage two x-coordinate
for a two-stage sample.}

\item{y}{Vector of y-coordinate for location for each site, which is either
the y- coordinate for a single-stage sample or the stage two y-coordinate
for a two-stage sample.}

\item{revisitwgt}{Logical value that indicates whether each repeat visit
site has the same survey design weight in the two surveys, where TRUE = the
weight for each repeat visit site is the same and FALSE = the weight for
each repeat visit site is not the same.  When this argument is FALSE, all
of the repeat visit sites are assigned equal weights when calculating the
covariance component of the change estimate standard error.}

\item{prop1}{The set of category proportion estimates for survey one.}

\item{prop2}{The set of category proportion estimates for survey two.}

\item{stratum_ind}{Logical value that indicates whether the sample is
stratified, where TRUE = a stratified sample and FALSE = not a stratified
sample.}

\item{stratum_level}{The stratum level.}

\item{cluster_ind}{Logical value that indicates whether the sample is a
two- stage sample, where TRUE = a two-stage sample and FALSE = not a
two-stage sample.}

\item{clusterID}{Vector of the stage one sampling unit (primary sampling unit
or cluster) code for each site.}

\item{wgt1}{Vector of the final adjusted stage one weight for each site.}

\item{x1}{Vector of the stage one x-coordinate for location for each site.}

\item{y1}{Vector of the stage one y-coordinate for location for each site.}

\item{pcfactor_ind}{Logical value that indicates whether the finite
population correction factor is used during variance estimation, where TRUE
= use the population correction factor and FALSE = do not use the factor.
To employ the correction factor for a single-stage sample, a value must be
supplied for argument fpcsize.  To employ the correction factor for a
two-stage sample, values must be supplied for arguments Ncluster and
stage1size.}

\item{fpcsize}{Size of the resource, which is required for calculation of the
finite population correction factor for a single-stage sample.}

\item{Ncluster}{The number of stage one sampling units in the resource,
which is required for calculation of the finite population correction
factor for a two-stage sample.}

\item{stage1size}{Vector of the size of the stage one sampling units of a
two-stage sample, which is required for calculation of the finite
population correction factor for a two-stage sample.}

\item{vartype}{The choice of variance estimator, where "Local" = local mean
estimator and "SRS" = SRS estimator.}

\item{warn_ind}{Logical value that indicates whether warning messages were
generated, where TRUE = warning messages were generated and FALSE = warning
messages were not generated.}

\item{warn_df}{Data frame for storing warning messages.}

\item{warn_vec}{Vector that contains names of the population type, the
subpopulation, and an indicator.}
}
\value{
An object in list format composed of a vector named rslt, which
 contains the covariance or correlation estimates, a logical variable named
 warn,ind, which is the indicator for warning messges, and a data frame
 named warn_df, which contains warning messages.
}
\description{
This function uses the repeat visit sites for two probability surveys to
calculate either covariance or correlation estimates of estimated change in
the  proportion in each of a set of categories.  Covariance estimates are
calculated when the resivit sites have the same survey design weight in both
surveys.  Correlation estimates are calculated when the revisit sites do not
have the same weight in both surveys, in which case the sites are assigned
equal weights.  The revisitwgt argument controls whether covariance or
correlation estimates are calculated.  Either the simple random sampling
(SRS) variance/covariance estimator or the local mean variance/covariance
estimator is calculated, which is subject to user control.  The simple random
sampling variance/covariance estimator uses the independent random sample
approximation to calculate joint inclusion probabilities.  The function can
accomodate single-stage and two-stage samples.  Finite population and
continuous population correction factors can be utilized in variance
estimation.
}
\section{Other Functions Required}{

 \describe{
 \item{\code{\link{localmean_weight}}}{calculate the weighting matrix for
  the local mean variance estimator}
 \item{\code{\link{localmean_cov}}}{calculate the variance/covariance
  matrix using the local mean estimator}
 }
}

\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
