% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructAddr.R
\name{constructAddr}
\alias{constructAddr}
\title{Construct the Hierarchical Addresses for a Generalized Random-Tesselation
Stratified (GRTS) Survey Design}
\usage{
constructAddr(xc, yc, dx, dy, nlev)
}
\arguments{
\item{xc}{Vector of x-coordinates for the grid cells.}

\item{yc}{Vector of y-coordinates for the grid cells.}

\item{dx}{The x-axis grid cell dimension.}

\item{dy}{The y-axis grid cell dimension.}

\item{nlev}{Number of hierarchical levels for the grid.}
}
\value{
Vector of hierarchical addresses.
}
\description{
This function constructS the hierarchical addresses for a GRTS survey design.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
