% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon_data.R
\docType{data}
\name{decon_data}
\alias{decon_data}
\title{Small Lakes in Florida for Examining CDF Deconvolution}
\format{A data frame with 930 rows and 6 attributes:
\describe{
  \item{xcoord}{simulated x-coordinate value.}
  \item{ycoord}{simulated y-coordinate value.}
  \item{Richness}{simulated species richness value.}
  \item{Richness_25}{species richness value plus 25\% measurement error
    variance.}
  \item{Richness_50}{species richness value plus 50\% measurement error
    variance.}
  \item{Richness_100}{species richness value plus 100\% measurement error
    variance.}
}}
\usage{
decon_data
}
\description{
A dataset containing attributes for small lakes in Florida that can be used
to examine CDF deconvolution.
}
