% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dframe_to_sf.R
\name{dframe_to_sf}
\alias{dframe_to_sf}
\alias{dframe_to_df}
\title{Transform a dframe object into an sf object or a data frame object}
\usage{
dframe_to_sf(object, ...)

dframe_to_df(object)
}
\arguments{
\item{object}{A \code{dframe} object.}
}
\value{
An \code{sf} object (\code{dframe_to_sf()}) or a data frame object (\code{dframe_to_df()}).
}
\description{
\code{dframe_to_sf()} transforms an sframe object into an sf object,
which may be useful when calling generic functions with sf class (e.g. mapview).
\code{dframe_to_df()} transforms an sframe object into an data frame object,
while retaining the remaining structure of dframe (e.g. tibble).
}
\section{Functions}{
\itemize{
\item \code{dframe_to_df}: 
}}

\seealso{
dframe
}
