% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FL_lakes.R
\docType{data}
\name{FL_lakes}
\alias{FL_lakes}
\title{Small Lakes in Florida}
\format{
A data frame with 930 rows and 11 attributes:
\describe{
  \item{siteID}{site ID value.}
  \item{xcoord}{Albers projection x-coordinate.}
  \item{ycoord}{Albers projection y-coordinate.}
  \item{wgt}{survey design weight.}
  \item{Basin}{stream basin code.}
  \item{Status}{site evaluation status code.}
  \item{TNT}{target or nontarget category for the site evaluation status code.}
  \item{pH_Cat}{pH category.}
  \item{Coliform_Cat}{fecal coliform count category.}
  \item{Oxygen}{dissolved oxygen value.}
  \item{Turbidity}{turbidity value.}
}
}
\usage{
FL_lakes
}
\description{
A dataset containing attributes for small lakes in Florida.
}
