% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_address.R
\name{get_address}
\alias{get_address}
\title{Get Heirarchical Address}
\usage{
get_address(x, y, rand = TRUE)
}
\arguments{
\item{x}{The x-coordinate value in the sample frame of points.}

\item{y}{The y-coordinate value in the sample frame of points.}

\item{rand}{Whether or not to randomize the addressing.  Default is \code{TRUE}}
}
\value{
A sorted sample frame.
}
\description{
This function creates a tessellation of the two dimensional space, attributes
an address to each unit in the population and sorts the sampling frame in the 
order given by the addresses. This function was used directly from Guillaume Chauvet and Ronan Le Gleut 
as they describe in their paper Inference under pivotal sampling: Properties, 
variance estimation, and application to tesselation for spatial sampling,
Scand J Statist. 2020;1–24.
}
\details{
GET ADDRESS: function to attriubute an address to each unit and to sort the 
sampling frame. The parameter rand allows a randomization of the tessellation. 
The output is the order in which the sampling frame should be sorted.

BETWEEN: function to project the coordinates between two bounds low and up.
}
\author{
Marc Weber \email{Weber.Marc@epa.gov}
}
\keyword{survey}
