% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Great_Lakes}
\alias{Great_Lakes}
\title{Great Lakes}
\format{
A data frame with 4113 rows and 5 variables:
\describe{
  \item{COUNTRY}{.....}
  \item{WTBDY_NM}{.....}
  \item{RSRC_CLASS}{.....}
  \item{PSTL_CODE}{.....}
  \item{AREA_SQKM}{.....}
  \item{geometry}{simple features geometry list-column}
}
}
\usage{
Great_Lakes
}
\description{
A simple features (sf) multipolygon object of 4113 polygons created
from document.....  The variables are:
}
\keyword{datasets}
