% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grts_stratum.R
\name{grts_stratum}
\alias{grts_stratum}
\title{For a single stratum, select a spatially balanced sample using generalized random
tessalation stratified algorithm. For a point sample frame, the selection is a one-step
process. For linear and area sample frames, a two-step process is used. First a systematic
point sample is selected from either linear or area sample frame to create a dense point
sample. Then the grts algorithm is applied to the systematic point sample. The selection
in all cases is based on the survey design specification for the stratum.}
\usage{
grts_stratum(
  stratum,
  dsgn,
  sframe,
  sf_type,
  pt_density = NULL,
  legacy_option = NULL,
  legacy_sites = NULL,
  maxtry = 10,
  warn_ind = FALSE,
  warn_df = NULL
)
}
\value{
rslts A list consisting of an \code{sf} object for base sites, an \code{sf} object of
  \code{n_over} sites (\code{NULL} if none) an \code{sf} object of \code{n_near} sites (\code{NULL} if none) where
  the \code{sf} objects containing the sites selected that meet the survey design requirements
  and \code{warn_ind} - logical value for warning indicator and \code{warn_df} - a data frame
  for warning messages.
}
\description{
For a single stratum, select a spatially balanced sample using generalized random
tessalation stratified algorithm. For a point sample frame, the selection is a one-step
process. For linear and area sample frames, a two-step process is used. First a systematic
point sample is selected from either linear or area sample frame to create a dense point
sample. Then the grts algorithm is applied to the systematic point sample. The selection
in all cases is based on the survey design specification for the stratum.
}
\section{Other functions required}{

  \describe{
    \item{\code{\link{grtspts_ip}}}{Calculates inclusion probabilities}
    \item{\code{\link{grtspts_ipleg}}}{Calculates inclusion probabilities with legacy points}
    \item{\code{\link{get_address}}}{Creates hierarchical order for points}
    \item{\code{\link{UPpivotal}}}{Selects sample using pivotal method}
    \item{\code{\link{rho}}}{Orders sites in reverse hierarchical order}
    \item{\code{\link{grtspts_mindis}}}{selects a grts sample with minimum distance between
      points}
    }
}

\examples{
\dontrun{
test_sample <- grts(stratum, dsgn = test_design, sframe = "test_sf", sf_type = "point")
}

}
\author{
Tony Olsen \email{olsen.tony@epa.gov}
}
\keyword{survey}
