% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grtspts_ip.R
\name{grtspts_ip}
\alias{grtspts_ip}
\title{Calculate inclusion probabities for point sample frame for equal, unequal based
on categories or proportional to a variable. The inclusion probabilities are
calculated assuming a single stratum.}
\usage{
grtspts_ip(
  type = "equal",
  n_base,
  Nstratum = NULL,
  caty = NULL,
  aux = NULL,
  warn_ind = NULL,
  warn_df = NULL
)
}
\arguments{
\item{type}{Character string with values of equal, unequal or proportional. Default
is equal.}

\item{n_base}{Sample size for the stratum. If type is equal or proportional, then n_base
is numeric value for the total sample size for the stratum. If type is unequal,
then n_base is a named vector where names are the unequal probability categories with the
expected sample size for each category. Names in n_base must be a subset of the values in caty.
The sum of the category sample sizes is the total sample size for the stratum.}

\item{Nstratum}{Number of elements in the stratum. Each element will be assigned an
inclusion probabibility. Only required when type is "equal". Default is NULL}

\item{caty}{If type is unequal, a character variable that identifies the category
for each element of the sample frame that may be used for an unequal probability
design. All categories in n_base must be present in pt_caty but additional categories
are allowed but will not be included in the survey design for the statum. Default is NULL.}

\item{aux}{If type is proportional, a numeric vector that will be used to select
points proportional to the variable. Default is NULL.}

\item{warn_ind}{A logical value where TRUE indicates a warning message.
Used for internal collection of messages only.}

\item{warn_df}{A data frame containing messages warning of potential issues.
Used for internal collection of messages only.}
}
\value{
A numeric vector of inclusion probabilities with all values being greater
  than 0 and less than or equal 1. Values equal to 1 will be selected with certainty.
  Values equal to 0 will not be selected.
}
\description{
Calculate inclusion probabities for point sample frame for equal, unequal based
on categories or proportional to a variable. The inclusion probabilities are
calculated assuming a single stratum.
}
\examples{
tmp <- grtspts_ip(type = "equal", n_base = 10, Nstratum = 25)
tmp
tmp <- grtspts_ip(type = "unequal", n_base = c(small = 5, large = 5),
          caty = c(rep("small", 15), rep("large", 10)))
tmp
tmp <- grtspts_ip(type = "proportional", n_base = 10,
                  aux = c(rnorm(20, 3), -2, 30, 40, 0, 0))
tmp

}
\author{
Tony Olsen email{olsen.tony@epa.gov}
}
\keyword{inclusion}
\keyword{probability}
\keyword{survey,}
