% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Illinois_River_Legacy}
\alias{Illinois_River_Legacy}
\title{Illinois River legacy data}
\format{
5 rows and 2 variables:
\describe{
  \item{\code{STATE_NAME}}{ State name.}
  \item{\code{geometry}}{ POINT geometry using the NAD83 / Conus Albers
    coordinate reference system (EPSG: 5070).}
}
}
\usage{
Illinois_River_Legacy
}
\description{
An (\code{sf}) POINT object of legacy sites for  the Illinois
River data.
}
\keyword{datasets}
