% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{input_check}
\alias{input_check}
\title{Check Input Values for Analytical Functions}
\usage{
input_check(
  dframe,
  design_names,
  vars_cat,
  vars_cont,
  vars_stressor,
  vars_nondetect,
  subpops,
  sizeweight,
  popcorrect,
  popsize,
  vartype,
  jointprob,
  conf,
  cdfval = NULL,
  pctval = NULL,
  sigma = NULL,
  var_sigma = NULL,
  error_ind,
  error_vec
)
}
\arguments{
\item{dframe}{Data frame containing survey design variables, response
variables, and subpopulation (domain) variables.}

\item{design_names}{Vector composed of character values that identify the
names of the following survey design variables in the \code{dframe} data frame:
\code{siteID}, \code{weight}, \code{xcoord}, \code{ycoord}, \code{stratumID},
\code{clusterID}, \code{weight1}, \code{xcoord1},
\code{ycoord1}, \code{sweight}, \code{sweight1}, \code{fpcsize}, \code{Ncluster}, \code{stage1size}.}

\item{vars_cat}{Vector composed of character values that identify the
names of categorical response variables in the \code{dframe} data frame.}

\item{vars_cont}{Vector composed of character values that identify the
names of continuous response variables in the \code{dframe} data frame.}

\item{vars_stressor}{Vector composed of character values that identify the
names of stressor variables in the \code{dframe} data frame.}

\item{vars_nondetect}{Vector composed of character values that identify the
names of logical variables in the \code{dframe} data frame specifying the presence
of not detected (nondetect) values for response variables.}

\item{subpops}{Vector composed of character values that identify the
names of subpopulation variables in the \code{dframe} data frame.}

\item{sizeweight}{Logical value that indicates whether size weights should be
used during estimation.}

\item{popcorrect}{Logical value that indicates whether the finite population
correction factor is used during variance estimation.}

\item{popsize}{List providing known size of the resource.}

\item{vartype}{Character value providing choice of the variance estimator,
where \code{"Local"} = the local mean estimator, \code{"SRS"} = the simple random
sampling estimator, \code{"HT"} = the Horvitz-Thompson estimator, and \code{"YG"} = the
Yates-Grundy estimator.}

\item{jointprob}{Character value providing choice of joint inclusion
probability approximation for use with Horvitz-Thompson and Yates-Grundy
variance estimators, where \code{"overton"} indicates the Overton approximation,
\code{"hr"} indicates the Hartley_Rao approximation, and \code{"brewer"} equals the
Brewer approximation.}

\item{conf}{Numeric value providing the confidence level.}

\item{cdfval}{Vector of the set of values at which the CDF is estimated.}

\item{pctval}{Vector of the set of values at which percentiles are
estimated.}

\item{sigma}{Numeric value providing the measurement error variance.}

\item{var_sigma}{Numeric value providing the variance of the measurement
error variance.}

\item{error_ind}{Logical value that indicates whether error messages were
generated.}

\item{error_vec}{Vector for storing error messages.}
}
\value{
A list consisting of \code{dframe}, \code{vars_cat}, \code{vars_cont}, \code{vars_stressor},
  \code{subpops}, \code{popsize}, \code{vartype}, \code{jointprob}, \code{error_ind}, and \code{error_vec}.
}
\description{
This function checks input values for errors, consistency, and compatibility
with analytical functions.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{vecprint}}}{takes an input vector and outputs a
      character string with line breaks inserted}
  }
}

\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
