% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideAreaGridCell.R
\name{insideAreaGridCell}
\alias{insideAreaGridCell}
\title{Calculate Clipped Feature Areas in a Set of Grid Cells}
\usage{
insideAreaGridCell(sfobject, rdx.u, xc, yc, dx, dy)
}
\arguments{
\item{sfobject}{The sf polygon object.}

\item{rdx.u}{Vector of cell IDs.}

\item{xc}{Vector of x-coordinates for the grid cells.}

\item{yc}{Vector of y-coordinates for the grid cells.}

\item{dx}{The x-axis grid cell dimension.}

\item{dy}{The y-axis grid cell dimension.}
}
\value{
Data frame containing the following variables: cellID, featureArea,
  and featureID.
}
\description{
For each grid cell, this function calculates the clipped area of each polygon
feature contained in the cell.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
