% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideLinearGridCell.R
\name{insideLinearGridCell}
\alias{insideLinearGridCell}
\title{Calculate Clipped Feature Lengths in a Set of Grid Cells}
\usage{
insideLinearGridCell(sfobject, rdx.u, xc, yc, dx, dy)
}
\arguments{
\item{sfobject}{The sf linestring object.}

\item{rdx.u}{Vector of cell IDs.}

\item{xc}{Vector of x-coordinates for the grid cells.}

\item{yc}{Vector of y-coordinates for the grid cells.}

\item{dx}{The x-axis grid cell dimension.}

\item{dy}{The y-axis grid cell dimension.}
}
\value{
Data frame containing the following variables: cellID, featureLength,
  and featureID.
}
\description{
For each grid cell, this function calculates the clipped length of each
linestring feature contained in the cell.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
