% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.axis.R
\name{interp.axis}
\alias{interp.axis}
\title{Create Right-Side y-Axis Lables for a CDF Plot}
\usage{
interp.axis(yl.lab, cdfest.l, cdfest.r)
}
\arguments{
\item{yl.lab}{Vector of left side y-axis labels, which are the basis for
interpolating cdfest.r values.}

\item{cdfest.l}{Vector of CDF estimates corresponding to the left side
y-axis.}

\item{cdfest.r}{Vector of CDF estimates corresponding to the right side
y-axis.}
}
\value{
A numeric vector consisting of the right side y-axis labels.
}
\description{
This function creates right side y-axis labels for a CDF plot.  It assumes
that arguments cdfest.l and cdfest.r are strictly increasing.  If argument
yl.lab is less than the first cdfest.l vlaue, then the funciton assumes 0 for
both cdfest.l and cdfest.r.
}
\author{
Tony Olsen \email{Olsen.Tony@epa.gov}\cr
  Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
