% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irslin.R
\name{irslin}
\alias{irslin}
\title{Select an Independent Random Sample (IRS) of a Linear Resource}
\usage{
irslin(linframe, samplesize = 100, SiteBegin = 1)
}
\arguments{
\item{linframe}{The sf object containing attributes: id, mdcaty, length_mdm,
and mdm.}

\item{samplesize}{Number of points to select in the sample.  The default is
100.}

\item{SiteBegin}{First number to start siteID numbering.  The default is 1.}
}
\value{
An sf object of sample points containing attributes: siteID, id,
  mdcaty, and wgt.
}
\description{
This function selects an IRS of a linear resource.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
