% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irspts.R
\name{irspts}
\alias{irspts}
\title{Select an Independent Random Sample (IRS) of a Finite Resource}
\usage{
irspts(ptsframe, samplesize = 100, SiteBegin = 1)
}
\arguments{
\item{ptsframe}{The sf object containing attributes: id, mdcaty, and mdm.}

\item{samplesize}{Number of points to select in the sample.  The default is
100.}

\item{SiteBegin}{First number to start siteID numbering.  The default is 1.}
}
\value{
An sf object of sample points containing attributes: siteID, id,
  mdcaty, and wgt.
}
\description{
This function selects an IRS of a finite resource (discrete points).
}
\author{
Tom Kincaid   \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
