% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irspts_mindis.R
\name{irspts_mindis}
\alias{irspts_mindis}
\title{Select an irs sample with minimum distance between sites.}
\usage{
irspts_mindis(
  mindis,
  sframe,
  samplesize,
  stratum,
  maxtry = 10,
  legacy_option = NULL,
  legacy_var = NULL,
  warn_ind = NULL,
  warn_df = NULL
)
}
\arguments{
\item{mindis}{Minimum distance required between sites in sample.}

\item{sframe}{The sf object containing variables: id and ip.}

\item{samplesize}{The sample size required for the stratum.}

\item{stratum}{Character string that identifies stratum membership for each element 
in the frame. Cannot be NULL.}

\item{maxtry}{Number of maximum attempts to ensure minimum distance between sites.
Default is 10.}

\item{legacy_option}{Logical variable where if equal to TRUE, legacy sites are
to be incorporated into survey design.}

\item{legacy_var}{Character value for name of column for legacy site variable.
Default is NULL.}

\item{warn_ind}{A logical value where TRUE indicates a warning message.
Used for internal collection of messages only.}

\item{warn_df}{A data frame containing messages warning of potential issues.
Used for internal collection of messages only.}
}
\value{
sites A list of sf object of sample points, an sf object of over sample
  points if any, warning indicator and warning messages data.frame
}
\description{
Select an irs sample with minimum distance between sites.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{get_address}}{constructs the hierarchical address for
      sample points}
    \item{\code{grtspts_ipleg}}{Calculates inclusion probabilities with legacy points}
    \item{\code{\link{rho}}}{Orders sites in reverse hierarchical order}
  }
}

\author{
Tony Olsen \email{Olsen.Tony@epa.gov}
}
\keyword{survey}
