% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotonic.R
\name{isotonic}
\alias{isotonic}
\title{Internal Function: Isotonic Regression}
\usage{
isotonic(y, minval, maxval)
}
\arguments{
\item{y}{Vector of the set of values on which to perform isotonic
regression.}

\item{minval}{Minimum value for the output set of values.}

\item{maxval}{Maximum value for the output set of values.}
}
\value{
The revised set of input values

  Other Functions Required:
     sorted - determine whether a set of values is a nondecreasing sequence
}
\description{
This function performs isotonic regression of an input set of values so that
the output set of values is a nondecreasing sequence.  The output set of
values is truncated to the range: [minval, maxval].
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{sorted}}}{determines whether a set of values is a
      nondecreasing sequence}
  }
}

\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
