% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lake_Ontario}
\alias{Lake_Ontario}
\title{Lake Ontario}
\format{
An \code{sf} multipolygon object with 187 rows and 4 variables:
\itemize{
  \item{COUNTRY}{Country.}
  \item{RSRC_CLASS}{Bay class.}
  \item{PSTL_CODE}{Postal code.}
  \item{AREA_SQKM}{Area in square kilometers}
  \item{geometry}{simple features geometry list-column}
}
}
\usage{
Lake_Ontario
}
\description{
An \code{sf} multipolygon object of 187 polygons consisting
of shoreline segments in Lake Ontario.
}
\keyword{datasets}
