% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localmean.df.R
\name{localmean.df}
\alias{localmean.df}
\title{Internal Function: Degrees of Freedom for Local Mean Variance-Covariance
Estimator}
\usage{
localmean.df(weight.lst)
}
\arguments{
\item{weight.lst}{List from the local mean weight function containing two
elements: a matrix named ij composed of the index values of neighboring
points and a vector named gwt composed of weights.}
}
\value{
The degrees of freedom of the local mean variance-covariance
  estimator.
}
\description{
This function calculates the degrees of freedom of the local mean variance-
covariance estimator.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
