% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localmean.var.R
\name{localmean.var}
\alias{localmean.var}
\title{Internal Function: Local Mean Variance Estimator}
\usage{
localmean.var(z, weight.lst)
}
\arguments{
\item{z}{Vector of weighted response values or weighted residual values for
the sample points.}

\item{weight.lst}{List from the local mean weight function containing two
elements: a matrix named ij composed of the index values of neighboring
points and a vector named gwt composed of weights.}
}
\value{
The local mean estimator of the variance.
}
\description{
This function calculates the local mean variance estimator.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
