% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localmean_cov.R
\name{localmean_cov}
\alias{localmean_cov}
\title{Internal Function: Variance-Covariance Matrix Based on Local Mean Estimator}
\usage{
localmean_cov(zmat, weight_1st)
}
\arguments{
\item{zmat}{Matrix of weighted response values or weighted residual values
for the sample points.}

\item{weight_1st}{List from the local mean weight function containing two
elements: a matrix named \code{ij} composed of the index values of neighboring
points and a vector named \code{gwt} composed of weights.}
}
\value{
The local mean estimator of the variance-covariance matrix.
}
\description{
This function calculates the variance-covariance matrix using the local mean
estimator.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
