% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localmean_var.R
\name{localmean_var}
\alias{localmean_var}
\title{Internal Function: Local Mean Variance Estimator}
\usage{
localmean_var(z, weight_1st)
}
\arguments{
\item{z}{Vector of weighted response values or weighted residual values for
the sample points.}

\item{weight_1st}{List from the local mean weight function containing two
elements: a matrix named ij composed of the index values of neighboring
points and a vector named gwt composed of weights.}
}
\value{
The local mean estimator of the variance.
}
\description{
This function calculates the local mean variance estimator.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
