% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localmean_weight.R
\name{localmean_weight}
\alias{localmean_weight}
\title{Internal Function: Local Mean Variance Neighbors and Weights}
\usage{
localmean_weight(x, y, prb, nbh = 4)
}
\arguments{
\item{x}{Vector of x-coordinates for location of the sample points.}

\item{y}{Vector of y-coordinates for location of the sample points.}

\item{prb}{Vector of inclusion probabilities for the sample points.}

\item{nbh}{Number of neighboring points to use in the calculations.}
}
\value{
If ginv fails to return valid output, a NULL object.  Otherwise, a
  list containing two elements: a matrix named \code{ij} composed of the
  index values of neighboring points and a vector named \code{gwt}
  composed of weights.
}
\description{
This function calculates the index values of neighboring points and
associated weights required by the local mean variance estimator.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
