% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localmean_weight2.R
\name{localmean_weight2}
\alias{localmean_weight2}
\title{Internal Function: Recovery from a Singular Value Decomposition Error}
\usage{
localmean_weight2(x, y, prb, nbh)
}
\arguments{
\item{x}{Vector of x-coordinates for location of the sample points.}

\item{y}{Vector of y-coordinates for location of the sample points.}

\item{prb}{Vector of inclusion probabilities for the sample points.}

\item{nbh}{Number of neighboring points to use in the calculations.}
}
\value{
Either an object of class "try-error" when the ginv function
  terminates with an error or a generalized inverse matrix when the ginv
  function terminates normally.
}
\description{
This function calculates the initial section of the localmean_weight function
and serves to allow recovery from an error in the singular value
decomposition function (La.svd) that is called by the generalized inverse
function (ginv) in the MASS package.
}
\author{
Don Stevens  Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
