% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Create the Grid for a GRTS Survey Design}
\usage{
make_grid(xc, yc, dx, dy, sfobject)
}
\arguments{
\item{xc}{x-coordinates that define the cells.}

\item{yc}{y-coordinates that define the cells.}

\item{dx}{The x-axis grid cell dimension.}

\item{dy}{The y-axis grid cell dimension.}

\item{sfobject}{the sf object containing the survey frame.}
}
\value{
An sf object containing the grid.
}
\description{
Creates the grid for  a GRTS suvey design.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
