% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdmarea.R
\name{mdmarea}
\alias{mdmarea}
\title{Internal Function: GRTS Multipliers for Multi-Density Categories for Areas}
\usage{
mdmarea(area, mdcaty, n.desired)
}
\arguments{
\item{area}{Vector of polygon areas for each polygon in the sample frame.}

\item{mdcaty}{Vector of multi-density category names for each polygon in the
sample frame.}

\item{n.desired}{Expected sample size for each category.  Row names must
match category names in mdcaty.}
}
\value{
Numeric vector of multipliers that is same length as area and mdcaty.
}
\description{
GRTS Multiplieers for Multi-density Categories for areas
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
