% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdmlin.R
\name{mdmlin}
\alias{mdmlin}
\title{Internal Function: GRTS Multipliers for Multi-Density Categories for Linear Network}
\usage{
mdmlin(len, mdcaty, n.desired)
}
\arguments{
\item{len}{Vector of segment lengths for each segment in sample frame.}

\item{mdcaty}{Vector of multi-density category groups for each segment in
sample frame.}

\item{n.desired}{Expected sample size for each category.  Row names must
match category names in mdcaty.}
}
\value{
Numeric vector of multipliers that is same length as len and mdcaty.
}
\description{
Internal Function: GRTS Multipliers for Multi-Density Categories for Linear Network
}
\author{
Tony Olsen \email{Olsen.Tony@epa.gov}
}
