% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdmpts.R
\name{mdmpts}
\alias{mdmpts}
\title{Internal Function: GRTS Multipliers for Multi-Density Categories for Points}
\usage{
mdmpts(mdcaty, n.desired)
}
\arguments{
\item{mdcaty}{Vector of multi-density category groups for each element in
sample frame.}

\item{n.desired}{Expected sample size for each category.  Row names must
match category names in mdcaty.}
}
\value{
Numeric vector of multipliers that is same length as mdcaty.
}
\description{
Internal Function: GRTS Multipliers for Multi-Density Categories for Points
}
\author{
Tony Olsen \email{Olsen.Tony@epa.gov}
}
