% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_localmean.R
\name{mean_localmean}
\alias{mean_localmean}
\title{Local Mean Variance Estimates of the Estimated Mean}
\usage{
mean_localmean(
  itype,
  lev_itype,
  nlev_itype,
  levs,
  ivar,
  design,
  design_names,
  meanest,
  popcorrect,
  vartype,
  mult,
  warn_ind,
  warn_df
)
}
\arguments{
\item{itype}{Character value that identifies a factor variable in the design
argument containing subpopulation (domain) values.}

\item{lev_itype}{Character vector that provides levels of the subpopulation
variable.}

\item{nlev_itype}{Numeric value that provides the number of levels of the
subpopulation variable.}

\item{levs}{Numeric vector that is used to select subpopulation levels}

\item{ivar}{Character value that identifies a factor variable in the design
argument containing categorical response values.}

\item{design}{Object of class \code{survey.design} that specifies a complex survey
design.}

\item{design_names}{Character vector that provides names of survey design
variables in the \code{design} argument.}

\item{meanest}{Vector that provides estimates of the mean.}

\item{popcorrect}{Logical value that indicates whether the finite population
correction factor should be employed during variance estimation.}

\item{vartype}{The choice of variance estimator, where \code{"Local"} = local mean
estimator and \code{"SRS"} = SRS estimator.}

\item{mult}{Numeric value that provides the Normal distribution confidence
bound multiplier.}

\item{warn_ind}{Logical value that indicates whether warning messages were
generated.}

\item{warn_df}{Data frame for storing warning messages.}
}
\value{
A list containing the following objects:
  \describe{
    \item{\code{stderr}}{vector containing standard error estimates}
    \item{\code{confval}}{data frame containing confidence bound estimates}
    \item{\code{warn_ind}}{logical variable that indicates whether warning
      messages were generated}
    \item{\code{warn_df}}{data frame for storing warning messages}
  }
}
\description{
This function organizes input and output for calculation of the local mean
variance estimator for the estimated mean.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{mean_var}}}{calculates variance estimate of the
      estimated mean}
    \item{\code{\link{svymean}}}{calculates the mean for a complex survey
      design}
  }
}

\seealso{
\code{\link{cdf_localmean_total}}
  \code{\link{svymean}}
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
