% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_var.R
\name{mean_var}
\alias{mean_var}
\title{Local Mean Variance Estimate for the Mean}
\usage{
mean_var(
  z,
  wgt,
  x,
  y,
  mean_est,
  stratum_ind,
  stratum_level,
  cluster_ind,
  cluster,
  wgt1,
  x1,
  y1,
  pcfactor_ind,
  fpcsize,
  Ncluster,
  stage1size,
  vartype,
  warn_ind,
  warn_df,
  warn_vec
)
}
\arguments{
\item{z}{Vector of the response value for each site.}

\item{wgt}{Vector of the final adjusted weight (reciprocal of the sample
inclusion probability) for each site, which is either the weight for a
single-stage sample or the stage two weight for a two-stage sample.}

\item{x}{Vector of x-coordinate for location for each site, which is either
the x-coordinate for a single-stage sample or the stage two x-coordinate
for a two-stage sample.}

\item{y}{Vector of y-coordinate for location for each site, which is either
the y-coordinate for a single-stage sample or the stage two y-coordinate
for a two-stage sample.}

\item{mean_est}{The mean estimate.}

\item{stratum_ind}{Logical value that indicates whether the sample is
stratified, where \code{TRUE} = a stratified sample and \code{FALSE} = not a stratified
sample.}

\item{stratum_level}{The stratum level.}

\item{cluster_ind}{Logical value that indicates whether the sample is a
two-stage sample, where \code{TRUE} = a two-stage sample and \code{FALSE} = not a
two-stage sample.}

\item{cluster}{Vector of the stage one sampling unit (primary sampling unit
or cluster) code for each site.}

\item{wgt1}{Vector of the final adjusted stage one weight for each site.}

\item{x1}{Vector of the stage one x-coordinate for location for each site.}

\item{y1}{Vector of the stage one y-coordinate for location for each site.}

\item{pcfactor_ind}{Logical value that indicates whether the finite
population correction factor is used during variance estimation, where \code{TRUE}
= use the population correction factor and \code{FALSE} = do not use the factor.
To employ the correction factor for a single-stage sample, a value must be
supplied for argument \code{fpcsize}.  To employ the correction factor for a
two-stage sample, values must be supplied for arguments \code{Ncluster} and
\code{stage1size}.}

\item{fpcsize}{Size of the resource, which is required for calculation of the
finite population correction factor for a single-stage sample.}

\item{Ncluster}{The number of stage one sampling units in the resource,
which is required for calculation of finite and continuous population
correction factors for a two-stage sample.  For a stratified sample this
variable must be a vector containing a value for each stratum and must have
the names attribute set to identify the stratum codes.}

\item{stage1size}{Size of the stage one sampling units of a two-stage
sample, which is required for calculation of finite and continuous
population correction factors for a two-stage sample and must have the
names attribute set to identify the stage one sampling unit codes.  For a
stratified sample, the names attribute must be set to identify both stratum
codes and stage one sampling unit codes using a convention where the two
codes are separated by the & symbol, e.g., \code{"Stratum 1&Cluster 1"}.}

\item{vartype}{The choice of variance estimator, where \code{"Local"} = local mean
estimator and \code{"SRS"} = SRS estimator.}

\item{warn_ind}{LogicLal value that indicates whether warning messages were
generated, where \code{TRUE} = warning messages were generated and \code{FALSE} = warning
messages were not generated.}

\item{warn_df}{Data frame for storing warning messages.}

\item{warn_vec}{Vector that contains names of the population type, the
subpopulation, and an indicator.}
}
\value{
Object in list format composed of a vector named \code{varest}, which
  contains variance estimates, a logical variable named \code{warn_ind}, which is
  the indicator for warning messges, and a data frame named \code{warn_df}, which
  contains warning messages.
}
\description{
This function calculates variance estimates of the estimated population
mean of a response variable. Either the simple random sampling (SRS) variance
estimator or the local mean variance estimator is calculated.  The SRS
variance estimator uses the independent random sample approximation to
calculate joint inclusion probabilities.  The function can  accomodate
single-stage and two-stage samples.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{localmean_weight}}}{calculate the weighting matrix for
      the local mean variance estimator}
    \item{\code{\link{localmean_var}}}{calculate the local mean variance
      estimator}
  }
}

\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
