% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSize.R
\name{n.mean}
\alias{n.mean}
\title{Sample size for estimating a mean}
\usage{
n.mean(halfwidth, s, v = NULL, conf = 0.95, tolprb = 0.9, N = NULL,
  method = "Harris")
}
\arguments{
\item{halfwidth}{A single value or vector of one-half the width of the
confidence interval desired.}

\item{s}{Estimate of standard deviation for the population.}

\item{v}{Degrees of freedom for associated with the estimated standard
deviation. Only required for Harris method. Default is NULL.}

\item{conf}{A single value or vector for confidence level desired. Default is
0.95.}

\item{tolprb}{Tolerance probability for confidence interval coverage. Default
is 0.90.}

\item{N}{Finite population size.  If NULL (default), then infinite population
assumed.}

\item{method}{Which sample size approach to be used.  If 'Harris', then
Harris, Horvitz, & Moody (1948). If 'Guenther', then Guenther(1965). If
'Cochran', then Cochran(1977) simple formulation using t-distribution.
Default is 'Harris'.}
}
\value{
A single value or an array of the sample size required to meet
  desired precision.
}
\description{
Calculates sample size for estimating mean with 1-beta
 probability that the 100*conf confidence interval will have length equal to
 halfwidth. May specify multiple halfwidths for sample size calculation.
}
\details{
Sample size for confidence interval for a mean is calculated used
  three alternative methods.  Cochran method is the standard approach given
  in most books and assumes a known variance.  Guenther and Harris methods
  are a tolerance interval approach that require meeting a tolerance
  probability for the confidence interval.  Guenther method assumes a known
  variance while Harris method assumes an estimated variance.
}
\examples{
Example from Zar(1999)
n.mean (1.5, s = sqrt(18.0388), method = 'Cochran')
n.mean (1.5, s = sqrt(18.0388), method = 'Guenther')
n.mean (1.5, s = sqrt(18.0388), v = 24, method = 'Harris')
# n.mean test when given vector of halfwidth
n.mean (c(1.0, 1.5, 2.0, 2.5), s = sqrt(18.0388),
        conf=c(0.90, 0.95), method = 'Cochran')
n.mean(c(1.0, 1.5, 2.0, 2.5), s = sqrt(18.0388), method = 'Guenther')
n.mean(c(1.0, 1.5, 2.0, 2.5), s = sqrt(18.0388), v =24, method = 'Harris')

}
\references{
Kupper, L. L. and K. B. Hafner (1989). "How appropriate are popular sample size formulas?"
        The American Statistician 43(2): 101-105.

   Harris, Horvitz, & Moody (1948) On the Determination of Sample Sizes
       in Designing Experiments. American Statistical Association Journal 43:391-402.

   Cochran, W. G. (1987). Sampling Techniques. New York, John Wiley & Sons.
}
\seealso{
PracticalTools package has several functions for sample size
  calculations. nCont uses coefficient of variance as input for confidence
  interval for mean.  strAlloc allocates sample size to strata.
}
