% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSize.R
\name{n.mean.test}
\alias{n.mean.test}
\title{Sample size for testing a mean}
\usage{
n.mean.test(delta, s, alpha = 0.05, power = 0.9, alternative = "twoway",
  N = NULL, method = "Ttest")
}
\arguments{
\item{delta}{single value or vector of difference(s) to be detected}

\item{s}{Estimate of standard deviation for the population}

\item{alpha}{Significance level desired for test. Default is 0.05.}

\item{power}{Single value or vector of power (1 - beta) desired for test.
Default is 0.90.}

\item{alternative}{Equal to 'twoway' for two-sided test or 'oneway' for
one-sided test. Default is 'twoway'.}

\item{N}{Finite population size.  If NULL (default), then infinite
population assumed}
}
\value{
The estimated sample size required to meet desired power as a single
  value or as an array that is number of deltas by number of P values
  specified
}
\description{
Sample size for test of mean being equal to, less than/equal to
  or greater than/equal to a hypothesized value with alpha as significance
  probability (Type I error) and beta probability of type II error to detect
  a difference of delta.
}
\details{
Standard z-test or t-test based sample size formulas are used to
  compute the required sample size.
}
\examples{
n.mean.test(1, s=2, power=0.99, alpha=0.05, alternative='oneway', method='Ztest')
n.mean.test(1, s=2, power=0.99, alpha=0.05, alternative='oneway', method='Ttest')
n.mean.test( delta = c( 1.0, 1.5, 2.0, 2.5), s=sqrt(1.5682),
            power=c(0.90, 0.95), method='Ttest')
n.mean.test( delta = c( 1.0, 1.5, 2.0, 2.5), s=sqrt(1.5682),
             power=c(0.90, 0.95), method='Ztest')
}
\references{
Zar, J. H. (1999). Biostatistical Analysis. Upper Saddle River,
    New Jersey, Prentice-Hall, Inc.

 Guenther, W. C. (1973). "Determination of sample size for tests concerning means
   and variances of normal distributions." Statistical Neerlandica 27: 103-113.
}
\seealso{
PracticalTools package has functions for determining sample sizes
  for testing means. They include nDep2sam and nProp2sam
}
