% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleSize.R
\name{n.prop}
\alias{n.prop}
\title{Sample size for estimating proportions}
\usage{
n.prop(halfwidth, P = 0.5, v = NULL, conf = 0.95, tolprb = 0.9,
  N = NULL, method = "Wilson")
}
\arguments{
\item{halfwidth}{Single value or vector of one-half the width of the
confidence interval desired for the true proportion.}

\item{P}{Single value or vector of true proportion for the population desired}

\item{v}{Degrees of freedom for associated with the estimated standard
deviation of the assumed true proportion. Only required for Harris method}

\item{conf}{Confidence level desired}

\item{tolprb}{Tolerance probability for confidence interval coverage}

\item{N}{Finite population size.  If NULL, then infinite population assumed
Default is NULL.}

\item{method}{Which sample size approach to be used.  If 'Wilson', then the
score confidence interval is used as first discussed by Wilson (1927). If
'Wald', then Cochran(1977) simple formulation using t-distribution. Default
is 'Wilson'.}
}
\value{
The estimated sample size required to meet desired precision as a
 single value or as an array that is number of halfwidths by number of P
 values specified.
}
\description{
Calculates sample size for estimating proportion with 1-beta
  probability that the 100*conf confidence interval will have length equal to
  halfwidth.
}
\details{
Three methods for computing the sample size required for a confidence
 interval for a proportion are given. Wald method is the standard method
 given by Cochran (1977) with enhancement to use t-distribution with correct
 degrees of freedom which requires iteration to achieve. It assumes that true
 proportion is known.  Wilson method is based on the score confidence
 interval for a proportion and assumes true proportion is known. Harris
 method is a tolerance interval approach that requires meeting a tolerance
 probability for the confidence interval and assumes variance of p is
 estimated with v degrees of freedom.
}
\examples{
Example from Cochran for proportions
n.prop(0.05, P=.5, method = 'Wald')
n.prop(0.05, P=.5, method = 'Wilson')
n.prop(0.05, P=.5, v=25, method = 'Harris')
n.prop(0.05, P=.5, N = 3200, method = 'Wald')
n.prop(0.05, P=.5, N = 3200, v = 25, method = 'Harris')
Examples when given vector of halfwidth and of P
n.prop(halfwidth = c(0.025, 0.05, 0.075, 0.10, 0.20), P = c(0.3, 0.5), N=3200, method='Wald')
n.prop(c(0.025, 0.05, 0.075, 0.10, 0.20), P = c(0.3, 0.5), N=3200, method='Wilson')
n.prop(c(0.025, 0.05, 0.075, 0.10, 0.20), P = c(0.3, 0.5), v=25,  N=3200, method='Harris')
}
\references{
Agresti, A. and B. A. Coull (1998). "Approximate is better than `exact' for interval
    estimation of binomial proportions." The American Statistician 52: 119-126.

Cochran, W. G. (1977). Sampling Techniques. New York, John Wiley & Sons.

Wilson, E. B. (1927). "Probable inference, the law of succession, and statistical inference."
  Journal of the American Statistical Association 22: 209-212.
}
\seealso{
PracticalTools package has several functions for determining sample sizes for proportions
     including nProp, nPropMoe, nWilson.
}
