% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCCA_Gulf}
\alias{NCCA_Gulf}
\title{NCCA Gulf}
\format{
A data frame with 851 rows and 18 variables:
\describe{
  \item{UNIQUE_ID}{.....}
  \item{SITE_ID}{.....}
  \item{REVISIT}{.....}
  \item{DSGN_CYCLE}{.....}
  \item{LAT_DD83}{.....}
  \item{LON_DD83}{.....}
  \item{XCOORD}{.....}
  \item{YCOORD}{.....}
  \item{WGT_TP_EXTENT}{.....}
  \item{WGT_SP_CORE}{.....}
  \item{WGT_TP_CORE}{.....}
  \item{EVAL_CAT}{.....}
  \item{TNT_CAT}{.....}
  \item{SMALL_EST}{.....}
  \item{STATE_NM}{.....}
  \item{BENTHIC_MMI}{.....}
  \item{BENTHIC_MMI_COND}{.....}
  \item{FISH_TISSUE_COND}{.....}
  \item{PHOSPHORUS_COND}{.....}
}
}
\usage{
NCCA_Gulf
}
\description{
A data frame of gulf information from the "National Coastal Condition Assessment" ..... The variables:
}
\keyword{datasets}
