% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NE_Lakes}
\alias{NE_Lakes}
\title{New England Lakes}
\format{
An \code{sf} object with 5328 rows and 7 variables:
\describe{
  \item{AREA_HA}{Lake area in hectares.}
  \item{AREA_HA_CAT}{Lake area categories based on a hectare cutoff.}
  \item{ELEVATION}{Elevation in meters.}
  \item{ELEVATION_CAT}{Elevation categories based on a meter cutoff.}
  \item{LEGACY}{Legacy site indicator.}
  \item{geometry}{Point geometry using a USA Contiguous Albers Equal Area
    Conic projection.}
}
}
\usage{
NE_Lakes
}
\description{
An \code{sf} point object of 5328 lakes in the Northeastern
United States.
}
\keyword{datasets}
