% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NE_Lakes_Legacy}
\alias{NE_Lakes_Legacy}
\title{New England Lakes legacy data}
\format{
5 rows and 5 variables:
\describe{
  \item{\code{AREA}}{ Lake area in hectares.}
  \item{\code{AREA_CAT}}{ Lake area categories based on a hectare cutoff.}
  \item{\code{ELEV}}{ Elevation in meters.}
  \item{\code{ELEV_CAT}}{ Elevation categories based on a meter cutoff.}
  \item{\code{geometry}}{ POINT geometry using the NAD83 / Conus Albers
    coordinate reference system (EPSG: 5070).}
}
}
\usage{
NE_Lakes_Legacy
}
\description{
An \code{sf} POINT object of 5 legacy sites for the NE Lakes data
}
\keyword{datasets}
