% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLA_2007.R
\docType{data}
\name{NLA_2007}
\alias{NLA_2007}
\title{Western Mountains Ecoregion Lakes}
\format{
A data frame with 236 rows and 15 attributes:
\describe{
  \item{siteID}{site ID value.}
  \item{xcoord}{Albers projection x-coordinate.}
  \item{ycoord}{Albers projection y-coordinate.}
  \item{wgt}{survey design weight.}
  \item{Lake_Origin}{lake origin category.}
  \item{Chla}{chlorophyll-a concentration.}
  \item{OE5}{value of the index of macroinvertebrate taxa loss.}
  \item{PTL}{total phosphorus concentration.}
  \item{NTL}{total nitrogen concentration.}
  \item{Turbidity}{turbidity value.}
  \item{Chla_cond}{condition class category ("Good", "Fair", or "Poor") of
    the chlorophyll-a value.}
  \item{OE5_cond}{condition class category of the macroinvertebrate taxa loss
    index.}
  \item{PTL_cond}{condition class category of the total phosphorus value.}
  \item{NTL_cond}{condition class category of the total nitrogen value.}
  \item{Turbidity_cond}{condition class category of the turbidity value.}
}
}
\usage{
NLA_2007
}
\description{
A dataset containing attributes for Western Mountains Ecoregion lakes
sampled by the U.S. Environmental Protection Agency (EPA) during the National
Lakes Assessment (NLA) survey for 2007.
}
