% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NLA_PNW}
\alias{NLA_PNW}
\title{NLA PNW data}
\format{
An \code{sf} point object with 96 rows and 9 variables:
\describe{
  \item{\code{SITE_ID}}{ A unique lake identifier.}
  \item{\code{WEIGHT}}{ The sampling design weight.}
  \item{\code{URBAN}}{ Urban category.}
  \item{\code{STATE}}{ State name.}
  \item{\code{BMMI}}{ Benthic MMI value.}
  \item{\code{BMMI_COND}}{ Benthic MMI condition categories.}
  \item{\code{PHOS_COND}}{ Phosphorus condition categories.}
  \item{\code{NITR_COND}}{ Nitrogen condition categories.}
  \item{\code{geometry}}{ Point geometry using a NAD83 / Conus Albers
    projection (EPSG: 5070).}
}
}
\usage{
NLA_PNW
}
\description{
An \code{sf} point object of 96 lakes in the Pacific Northwest Region of the United
States during the year 2017, from a subset of the Environmental
Protection Agency's "National Lakes Assessment."
}
\keyword{datasets}
