% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NLA_West}
\alias{NLA_West}
\title{NLA West}
\format{
A data frame with 851 rows and 18 variables:
\describe{
  \item{UNIQUE_ID}{.....}
  \item{SITE_ID}{.....}
  \item{DSGN_CYCLE}{.....}
  \item{LAT_DD83}{latitude NAD83.....} consider reloading
  \item{LON_DD83}{.....}
  \item{XCOORD}{transformed latitude coordinate}
  \item{YCOORD}{transformed longitude coordinate}
  \item{WGT_TP_EXTENT}{weight target pop extent}
  \item{WGT_SP_CORE}{weight sampled population CORE means sampled}
  \item{WGT_TP_CORE}{.....}
  \item{EVAL_CAT}{.....}
  \item{TNT_CAT}{.....}
  \item{URBN_NLA17}{.....}
  \item{STATE_NM}{.....}
  \item{BENTHIC_MMI}{.....}
  \item{BENTHIC_MMI_COND}{.....}
  \item{PHOSPHORUS_COND}{.....}
  \item{NITROGEN_COND}{.....}
}
}
\usage{
NLA_West
}
\description{
A data frame of lake information from a "National Lakes Assessment" in the 
Western United States.
}
\keyword{datasets}
