% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NRSA_EPA7}
\alias{NRSA_EPA7}
\title{NRSA EPA7}
\format{
A data frame with 682 rows and 18 variables:
\describe{
  \item{SITE_ID}{A unique site identifier.}
  \item{YEAR}{Year of design cycle.}
  \item{WEIGHT}{Sampling design weights.}
  \item{ECOREGION}{Ecoregion.}
  \item{STATE}{State name.}
  \item{BMMI}{Benthic MMI value.}
  \item{BMMI_COND}{Benthic MMI categories.}
  \item{PHOS_COND}{Phosphorus condition categories.}
  \item{NITR_COND}{Nitrogen condition categories.}
}
}
\usage{
NRSA_EPA7
}
\description{
An sf object consisting of stream observations from a subset of the Environmental
Protection Agency's "National Rivers and Streams Assessment" in the Central
United States duing the years 2008 and 2013.
}
\keyword{datasets}
