% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_est.R
\name{percentile_est}
\alias{percentile_est}
\title{Percentile Estimates for Probability Survey Data}
\usage{
percentile_est(
  pctsum,
  dframe,
  itype,
  lev_itype,
  nlev_itype,
  ivar,
  design,
  design_names,
  var_nondetect,
  popcorrect,
  vartype,
  conf,
  mult,
  pctval,
  warn_ind,
  warn_df
)
}
\arguments{
\item{pctsum}{Data frame containing estimates.}

\item{dframe}{Data frame containing survey design variables, response
variables, and subpopulation (domain) variables.}

\item{itype}{Character value that identifies a factor variable in the design
argument containing subpopulation (domain) values.}

\item{lev_itype}{Character vector that provides levels of the subpopulation
variable.}

\item{nlev_itype}{Numeric value that provides the number of levels of the
subpopulation variable.}

\item{ivar}{Character value that identifies the response variable.}

\item{design}{Object of class \code{survey.design} that specifies a complex survey
design.}

\item{design_names}{Character vector that provides names of survey design
variables in the \code{design} argument.}

\item{var_nondetect}{Character value that identifies the name of a logical
variable in the \code{dframe} data frame specifying the presence of not detected
(nondetect) values for the response variable.}

\item{popcorrect}{Logical value that indicates whether the finite population
correction factor should be employed during variance estimation.}

\item{vartype}{The choice of variance estimator, where \code{"Local"} = local mean
estimator and \code{"SRS"} = SRS estimator.}

\item{conf}{Numeric value for the confidence level.}

\item{mult}{Numeric value that provides the Normal distribution confidence
bound multiplier.}

\item{pctval}{Vector of the set of values at which percentiles are
estimated.}

\item{warn_ind}{Logical value that indicates whether warning messages were
generated.}

\item{warn_df}{Data frame for storing warning messages.}
}
\value{
A list composed of the following objects:
  \describe{
    \item{\code{pctsum}}{data frame containing the percentile estimates}
    \item{\code{warn_ind}}{logical variable that indicates whether warning
      messages were generated}
    \item{\code{warn_df}}{data frame for storing warning messages}
  }
}
\description{
This function calculates percentile estimates using the svyquantile function
in the survey package.  In additon, the mean is estimated using the svymean
function in the survey package.  Upper and lower confidence bounds also are
estimated.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{cdf_nresp}}}{calculates number of response values less
      than a set of values}
    \item{\code{\link{SE}}}{extracts standard errors from a survey design
      object}
    \item{\code{\link{confint}}}{computes confidence intervals for a survey
      design object}
    \item{\code{\link{mean_localmean}}}{organizes input and output for
      calculation of the local mean variance estimator for the estimated
      mean}
    \item{\code{\link{svyby}}}{Compute survey statistics on subsets of a
      survey defined by factors}
    \item{\code{\link{svymean}}}{calculates the mean for a complex survey
      design}
    \item{\code{\link{svyquantile}}}{calculates percentile estimates for a
      complex survey design}
  }
}

\seealso{
\code{\link{confint}}
  \code{\link{mean_localmean}}
  \code{\link{SE}}
  \code{\link{svyby}}
  \code{\link{svymean}}
  \code{\link{svyquantile}}
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
\keyword{univar}
