% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickFiniteSamplePoints.R
\name{pickFiniteSamplePoints}
\alias{pickFiniteSamplePoints}
\title{Select Sample Points from a Simple Features Point Object}
\usage{
pickFiniteSamplePoints(rdx, xc, yc, dx, dy, sfobject)
}
\arguments{
\item{rdx}{Vector of cell IDs.}

\item{xc}{Vector of x-coordinates for the grid cells.}

\item{yc}{Vector of y-coordinates for the grid cells.}

\item{dx}{The x-axis grid cell dimension.}

\item{dy}{The y-axis grid cell dimension.}

\item{sfobject}{The sf point object.}
}
\value{
Vector containing feature IDs of the selected sample points.
}
\description{
This function selects sample points from an sf object of geometry type
point.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
