% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickGridCells.R
\name{pickGridCells}
\alias{pickGridCells}
\title{Identify Grid Cells from which Sample Points Will Be Selected}
\usage{
pickGridCells(samplesize, idx)
}
\arguments{
\item{samplesize}{The desired smaple size.}

\item{idx}{Vector of values for identifying cell IDs.}
}
\value{
Vector of grid cells IDs.
}
\description{
This function identifies grid cells from which sample points will be selected.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
