% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickSamplePoints.R
\name{pickSamplePoints}
\alias{pickSamplePoints}
\title{Select Sample Points from a Simple Features Object}
\usage{
pickSamplePoints(sfobject, featureID, xc, yc, dx, dy)
}
\arguments{
\item{sfobject}{The sf polygon object.}

\item{featureID}{Vector identifying features in sfobject from which a sample
point will be selected.}

\item{xc}{Vector of x-coordinates for the grid cells.}

\item{yc}{Vector of y-coordinates for the grid cells.}

\item{dx}{The x-axis grid cell dimension.}

\item{dy}{The y-axis grid cell dimension.}
}
\value{
Data frame containing x-coordinates and y-coordinates for sample
  points.
}
\description{
This function selects sample points from an sf object of geometry type
polygon or linestring.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
