% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.sframe}
\alias{plot.spdesign}
\alias{plot.dframe}
\title{Plot sample frame and design objects}
\usage{
\method{plot}{sframe}(
  x,
  y,
  formula = ~1,
  var_args = NULL,
  varlevel_args = NULL,
  geom = FALSE,
  onlyshow = NULL,
  fix_bbox = TRUE,
  ...
)

\method{plot}{spdesign}(
  x,
  y = NULL,
  formula = ~siteuse,
  siteuse = NULL,
  var_args = NULL,
  varlevel_args = NULL,
  geom = FALSE,
  onlyshow = NULL,
  fix_bbox = TRUE,
  ...
)

\method{plot}{dframe}(
  x,
  y = NULL,
  formula = ~1,
  var_args = NULL,
  varlevel_args = NULL,
  geom = FALSE,
  onlyshow = NULL,
  fix_bbox = TRUE,
  xcoord,
  ycoord,
  crs,
  ...
)
}
\arguments{
\item{x}{object of class \code{sframe}, \code{spdesign}, or \code{dframe}.}

\item{y}{ignored if \code{x} has class \code{sframe} or \code{dframe}; an 
object of class \code{sframe} if \code{x} has class \code{spdesign}.}

\item{formula}{A formula. Left hand side variables can be numeric or
categorical (or factor) and right hand side variables can be categorical
(or factor). Right hand side variables that are numeric will be coerced to
a categorical (or factor) variable. If an intercept is included in the right
hand side formula, the total will also be summarized.}

\item{var_args}{A named list. The name of each list corresponds to a
right hand side variable in \code{formula}. Values in the list are composed of
graphical arguments that are to be passed to \strong{every} level of the
variable.}

\item{varlevel_args}{A named list. The name of each list corresponds to a
right hand side variable in \code{formula}. The first name of each sublist
should be \code{"levels"} and contain all levels of the variable. Subsequent
names correspond to graphical arguments that are to be passed to
the specified levels (in order). Values must be specified
for every level of each graphical argument, but applicable \code{sf} defaults
will be matched by inputting the value \code{NA}.}

\item{geom}{Should separate geometries for each level of the right hand
side \code{formula} variables be plotted? Defaults to \code{FALSE}.}

\item{onlyshow}{A string indicating the level of the single right hand side
variable for which a summary is requested.}

\item{fix_bbox}{Should the geometry bounding box be fixed across plots?
Defaults to \code{TRUE}. If \code{TRUE}, the bounding box will be the largest
possible given \code{x} and \code{y}.}

\item{...}{Additional arguments to pass to \code{plot.sf()}.}

\item{siteuse}{A character vector of site types to include when plotting a
design object. Can only take on values \code{"sframe"} (sample frame),
\code{"Legacy"} (for legacy sites), \code{"Base"} (for base sites),
\code{"Over"} (for \code{n_over} replacement sites), and \code{"Near"}
(for \code{n_near} replacement sites). The order of sites represents the
layering in the plot (e.g. \code{siteuse = c("Base", "Legacy")} will plot
legacy sites on top of base sites. Defaults to all non-\code{NULL} elements
in \code{x} and \code{y} with plot order \code{"sframe"}, \code{"Near"},
\code{"Over"}, \code{"Base"}, \code{"Legacy"}.}

\item{xcoord}{Name of x (east-west)-coordinate in \code{x} (only required if
\code{x} is not an sf object)}

\item{ycoord}{Name of y (north-south)-coordinate in \code{x} (only required if
\code{x} is not an sf object)}

\item{crs}{Projection code for \code{xcoord} and \code{ycoord} (only
required if \code{x} is not an sf object)}
}
\description{
Plot sample frames and design objects.

This function is largely built on \code{plot.sf()}, and all spsurvey plotting 
methods can supply additional arguments to \code{plot.sf()}. For more information, 
run \code{vignette(plotting, package = "spsurvey"}. For more information on
plotting in \code{sf}, run \code{?plot.sf()}.
}
\examples{
\dontrun{
plot(NE_Lakes)
plot(NE_Lakes, formula = ~ ELEV_CAT)
plot(NE_Lakes, formula = AREA ~ ELEV_CAT)
sample <- grts(NE_Lakes, 30)
plot(sample, NE_Lakes)
}
###############################################################################
}
\author{
Michael Dumelle \email{Dumelle.Michael@epa.gov}
}
