% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranho.R
\name{ranho}
\alias{ranho}
\title{Construct Randomized Hierarchical Addresses for a Generalized
Random-Tesselation Stratified (GRTS) Survey Design}
\usage{
ranho(hadr)
}
\arguments{
\item{hadr}{Vector hierarchical addresses.}
}
\value{
Vector of randomized hierarchical addresses.
}
\description{
This function constructs randomized hierarchical addresses for a GRTS survey
design.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
