% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dbf.R
\name{read.dbf}
\alias{read.dbf}
\title{Read the dbf File of an ESRI Shapefile}
\usage{
read.dbf(filename)
}
\arguments{
\item{filename}{Character string containing the name of the shapefile.}
}
\value{
Data frame containing contents of the dbf file.
}
\description{
This function reads the dbf file of an ESRI shapefile and creates a data
  frame.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}

Marc Weber  \email{Weber.Marc@epa.gov}
}
