% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.shape.R
\name{read.shape}
\alias{read.shape}
\title{Read an ESRI Shapefile}
\usage{
read.shape(filename)
}
\arguments{
\item{filename}{Character string containing the name of the shapefile.  The
shapefile name should include the ".shp" extension.  If the name does not
include that extension, it will be added.}
}
\value{
An object belonging to class sf.
}
\description{
This function reads an ESRI shapefile and creates a simple features (sf)
  object.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}

Marc Weber  \email{Weber.Marc@epa.gov}
}
