% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho}
\alias{rho}
\title{Place the sites selected by the UPpivotal method in reverse hierarchical order}
\usage{
rho(sites)
}
\arguments{
\item{sites}{Sample object with sites selected by GRTS algorithm}
}
\value{
sites Sample object returned in reverse hierarchical order
}
\description{
Place the sites selected by the UPpivotal method in reverse hierarchical order
}
\examples{

}
\author{
Tony Olsen email{olsen.tony@epa.gov}
}
\keyword{survey}
