% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_estuaries.R
\docType{data}
\name{SC_estuaries}
\alias{SC_estuaries}
\title{Estuaries in South Carolina}
\format{A data frame with 135 rows and 10 attributes:
\describe{
  \item{siteID}{site ID value.}
  \item{xcoord}{Albers projection x-coordinate.}
  \item{ycoord}{Albers projection y-coordinate.}
  \item{wgt}{survey design weight.}
  \item{Stratum}{stratum code.}
  \item{Status}{site evaluation status code.}
  \item{IBI_score}{IBI (index of biotic integrity) score.}
  \item{IBI_status}{status category of the IBI score.}
  \item{WQ_score}{WQ (water quality) score.}
  \item{WQ_status}{status category of the WQ score.}
}}
\usage{
SC_estuaries
}
\description{
A dataset containing attributes for estuaries in South Carolina.
}
