% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeatureID.R
\name{selectFeatureID}
\alias{selectFeatureID}
\title{Identify a Feature in an sf Object for Selecting a Sample Point}
\usage{
selectFeatureID(rdx, cellID, featureMeasure, featureID, mdm, id)
}
\arguments{
\item{rdx}{Value of the randomized hierarchical address identifying a grid
cell that will get a sample point.}

\item{cellID}{Vector of grid cell IDs.}

\item{featureMeasure}{Vector of grid cell sf feature lengths for
linestring objects or sf feature areas for polygon objects.}

\item{featureID}{Vector of grid cell sf feature IDs.}

\item{mdm}{Vector of multidensity multipliers for the shapefile features.}

\item{id}{Vector of sf feature IDs.}
}
\value{
The ID of an sf feature.
}
\description{
This function determines the ID value of the feature in an sf object from
which a sample point will be selected.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
