% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectframe.R
\name{selectframe}
\alias{selectframe}
\title{Internal Function: Select All Points in Survey Frame}
\usage{
selectframe(rord, xc, yc, dx, dy, pts)
}
\arguments{
\item{rord}{Vector of the index value for all cells.}

\item{xc}{Vector of x-coordinates that define the cells.}

\item{yc}{Vector of y-coordinates that define the cells.}

\item{dx}{Width of the cells along the x-axis.}

\item{dy}{Width of the cells along the y-axis.}

\item{pts}{Data frame containing id values, x-coordinates, y-coordinates,
and mdm values.}
}
\value{
The id value for all points in the frame.
}
\description{
Internal Function: Select All Points in Frame
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
