% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectpts.R
\name{selectpts}
\alias{selectpts}
\title{Internal Function: Select Probability Sample from a Set of Cells}
\usage{
selectpts(rdx, xc, yc, dx, dy, pts)
}
\arguments{
\item{rdx}{Vector of the index value for selected cells.}

\item{xc}{Vector of x-coordinates that define the cells.}

\item{yc}{Vector of y-coordinates that define the cells.}

\item{dx}{Width of the cells along the x-axis.}

\item{dy}{Width of the cells along the y-axis.}

\item{pts}{Data frame containing id values, x-coordinates, y-coordinates,
and mdm values.}
}
\value{
The id value for the sample points.
}
\description{
Selects a sample of size one or larger from a set of cells based on inclusion
probabilities.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
        Tony Olsen \email{olsen.tony@epa.gov}
}
