% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectrecordID.R
\name{selectrecordID}
\alias{selectrecordID}
\title{Internal Function: Select Shapefile Record to Select Sample Point}
\usage{
selectrecordID(rdx, cellID, recordMeasure, recordID, mdm, id)
}
\arguments{
\item{rdx}{Vector of the randomized hierarchical address identifying a grid
cell that will get a sample point.}

\item{cellID}{Vector of grid cell IDs.}

\item{recordMeasure}{Vector of grid cell shapefile record length for
polyline-type shapefiles or shapefile record area for polygon-type
shapefiles.}

\item{recordID}{Vector of grid cell shapefile record IDs.}

\item{mdm}{Vector of multidensity multipliers for the shapefile records.}

\item{id}{Vector of shapefile record IDs.}
}
\value{
The ID of a shapefile record.
}
\description{
This function selects a shapefile record from which a sample point will be
selected.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
