% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sframe.R
\name{sframe}
\alias{sframe}
\title{Create an sframe object}
\usage{
sframe(object, ...)
}
\arguments{
\item{object}{For \code{sframe}, an \code{sp} or \code{sf} object.}

\item{...}{Additional arguments to \code{st_as_sf()} when using \code{sframe()}
to convert an \code{sp} object to an \code{sf} and \code{sframe} object.}
}
\value{
A \code{sframe} object that can be used by
spsurvey's \code{summary()} and \code{plot()} generics.
}
\description{
\code{sframe()} is used before summarizing or visualizing sample frames
(this function need not be run before calling \code{grts()} or \code{irs()}).
  \code{sframe()} gives \code{sp} and \code{sf} objects class \code{"sframe"}.
  \code{sp} objects are coerced to \code{sf}
  objects prior to giving class \code{sframe}.
}
\examples{
data("NE_Lakes")
NE_Lakes <- sframe(NE_Lakes)
summary(NE_Lakes, formula = ~ ELEV_CAT)
plot(NE_Lakes, formula = ~ ELEV_CAT)
}
\seealso{
\describe{
    \item{\code{\link{sframe_to_sf}}}{ to turn an \code{sframe} object into an \code{sf} object}
 }
}
\author{
Michael Dumelle \email{Dumelle.Michael@epa.gov}
}
