% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sframe_to_sf.R
\name{sframe_to_sf}
\alias{sframe_to_sf}
\title{Transform an sframe object into an sf object}
\usage{
sframe_to_sf(object)
}
\arguments{
\item{object}{An \code{sframe} object.}
}
\value{
An \code{sf} object.
}
\description{
\code{sframe_to_sf()} transforms an \code{sframe} object into an \code{sf} object,
which may be useful when calling generic functions with sf class (e.g. mapview)
}
\seealso{
sframe
}
