% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape2spList.R
\name{shape2spList}
\alias{shape2spList}
\title{Internal Function: Create Object of Class Lines or Class Polygons}
\usage{
shape2spList(shape, shp.type, ID)
}
\arguments{
\item{shape}{A single record from the .shp file of the shapefile.}

\item{shp.type}{The type of shapefile, which is either "arc" for a Polyline
shapefile or "poly" for a Polygon shapefile.}

\item{ID}{The shape ID value, i.e., the shapefile record number.}
}
\value{
Object of class Lines for a Polyline shapefile or class Polygons
  for a Polygon shapefile - see documentation for the sp package for further
  details.
}
\description{
This function creates an object of class Lines for a Polyline shapefile or
class Polygons for a Polygon shapefile.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{Line}}}{sp package function to create an object of
      class Line}
    \item{\code{\link{Lines}}}{sp package function to create an object of
      class Lines}
    \item{\code{\link{Polygon}}}{sp package function to create an object of
      class Polygon}
    \item{\code{\link{Polygons}}}{sp package function to create an object of
      class Polygons}
  }
}

\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
