% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorted.R
\name{sorted}
\alias{sorted}
\title{Internal Function: Determine if a Vector is Sorted}
\usage{
sorted(x)
}
\arguments{
\item{x}{Vector of values.}
}
\value{
Logical variable, where TRUE = sorted and FALSE = not sorted.
}
\description{
This function determines whether the input set of values is a nondecreasing
sequence.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
