% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp2shape.R
\name{sp2shape}
\alias{sp2shape}
\title{Create an ESRI shapefile from an sp package object}
\usage{
sp2shape(sp.obj, shpfilename)
}
\arguments{
\item{sp.obj}{the sp package object or object created by either the grts or
irs functions.}

\item{shpfilename}{Character string containing the name of the shapefile.
The shapefile name should include the ".shp" extension.  If the name does
not include that extension, it will be added.}
}
\value{
An ESRI shapefile.
}
\description{
This function creates an ESRI shapefile from an sp package object.  The
function can also accommodate an object created by the grts or irs functions
in spsurvey.  The type of shapefile, i.e., point, polyline, or polygon, is
determined by the class of the sp object.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
