% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDesign_class.R,
%   R/SpatialDesign_methods.R
\docType{class}
\name{SpatialDesign-class}
\alias{SpatialDesign-class}
\alias{SpatialDesign}
\alias{SpatialDesign-method}
\alias{summary.SpatialDesign}
\alias{summary,}
\alias{summary,SpatialDesign-method}
\alias{plot.SpatialDesign}
\alias{plot,}
\alias{plot,SpatialDesign,missing-method}
\title{Class SpatialDesign}
\usage{
SpatialDesign(design, sp_obj)

\method{summary}{SpatialDesign}(object, ..., auxvar = NULL,
  sfframe = NULL, tess_ind = TRUE, sbc_ind = FALSE, nrows = 5,
  dxdy = TRUE)

\S4method{summary}{SpatialDesign}(object, ..., auxvar = NULL,
  sfframe = NULL, tess_ind = TRUE, sbc_ind = FALSE, nrows = 5,
  dxdy = TRUE)

\method{plot}{SpatialDesign}(x, y, ..., sfframe = NULL, stratum = NULL,
  mdcaty = NULL, auxvar = NULL, pdffile = NULL, width = 8,
  height = 10)

\S4method{plot}{SpatialDesign,missing}(x, y, ..., sfframe = NULL,
  stratum = NULL, mdcaty = NULL, auxvar = NULL, pdffile = NULL,
  width = 8, height = 10)
}
\arguments{
\item{design}{Object of class list containing specifications for the
survey design.}

\item{sp_obj}{Object of class \code{SpatialPointsDataFrame} containing
spatial attributes that have spatial point locations.}

\item{object}{\code{SpatialDesign} object.}

\item{...}{Arguments passed through.}

\item{auxvar}{Vector containing the names of variables in the data slot of
the \code{SpatialDesign} object that identify auxiliary variables to be
used to summarize the survey design or create plots of the survey design.
 The default is NULL.}

\item{sfframe}{Object of class \code{sf} that contains the survey design
frame.  The default is NULL.}

\item{tess_ind}{a logical variable indicating whether spatial balance
metrics are calculated using proportions obtained from the intersection of
Dirichlet tesselation polygons for the sample points with the frame object.
TRUE means calculate the metrics.  FALSE means do not calculate the
metrics.  The default is TRUE.}

\item{sbc_ind}{a logical variable indicating whether spatial balance metrics
are calculated using proportions obtained from a rectangular grid
superimposed on the sample points and the frame.  TRUE means calculate the
metrics. FALSE means do not calculate the metrics. The default is FALSE.}

\item{nrows}{number of rows (and columns) for the grid of cells.  The default
is 5.}

\item{dxdy}{indicator for equal x-coordinate and y-coordinate grid cell
increments, where TRUE means the increments are equal and FALSE means the
increments are not equal.  The default is TRUE.}

\item{x}{\code{SpatialDesign} object.}

\item{y}{Missing - this argument is not used.}

\item{stratum}{name of the attribute from the sfframe object that identifies
stratum membership for each feature in the frame.  If stratum equals NULL,
the design is unstratified, and an attribute named "stratum" (with all its
elements equal to the stratum name specified in design) is added to the
sfframe object.  The default is NULL.}

\item{mdcaty}{name of the attribute from the sfframe object that identifies
the unequal probability category for each feature in the survey frame.  The
default is NULL.}

\item{pdffile}{a character variable containing the name of the pdf file to
which output is written. If a value is not provided, output is written to
the graphics window.  The default is NULL.}

\item{width}{width of the graphic region in inches.  The default is 8.}

\item{height}{height of the graphic region in inches.  The default is 10.}
}
\value{
Object of class \code{SpatialDesign}.

A summary or plot depending on the method called.
}
\description{
Create the definition for class \code{SpatialDesign}, a class for spatial survey designs.

Create the wrapper function for class \code{SpatialDesign}.

Define S3 and S4 methods for summary and plot for class \code{SpatialDesign}.
}
\section{Slots}{

\describe{
\item{\code{design}}{Object of class list containing specifications for the survey
design}

\item{\code{data}}{Object of class data.frame containing the attribute data}

\item{\code{coords.nrs}}{Numeric object that records the column positions in
\code{data} from which the coordinates were obtained}

\item{\code{coords}}{Object of class matrix containing the coordinates matrix,
where points are rows in the matrix}

\item{\code{bbox}}{Object of class matrix containing the bounding box}

\item{\code{proj4string}}{Object of class CRS containing the projection string}
}}

\section{Extends}{

Class \code{"SpatialPointsDataFrame"}, directly.\cr\cr
Class \code{"SpatialPoints"}, by class \code{"SpatialPointsDataFrame"}.\cr\cr
Class \code{"Spatial"}, by class \code{"SpatialPoints"}.
}

\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
