% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbalance.R
\name{spbalance}
\alias{spbalance}
\title{Calculate spatial balance metrics}
\usage{
spbalance(
  object,
  sframe,
  sites = "sites_base",
  metrics = "pielou",
  extents = FALSE
)
}
\arguments{
\item{object}{A design object output from `grts()` or `irs()`
having class "spsurvey"}

\item{sframe}{An `sframe or `sf` object}

\item{sites}{The sites in object for which spatial balance metrics
are desired. This argument must be "sites_base", "sites_over", or 
"sites_near". Defaults to "sites_base"}

\item{metrics}{A character vector of spatial balance metrics 
(pielou, simpsons, chisquare, absolute error). Defaults to 
"pielou".}

\item{extents}{Should the total extent within each dirichlet
tesselation be returned? Defaults to FALSE.}
}
\value{
A list having names equal to each strata
}
\description{
This function computes the spatial balance of a design with respect 
to the sample frame (sframe) using Dirichlet Tesselations, measuring the
extent to which a object is a miniature of sframe. This function is
applicable for unstratified or stratified designs with equal selection
probabilities.
}
