% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprbind.R
\name{sprbind}
\alias{sprbind}
\title{Combine rows from a GRTS or IRS design}
\usage{
sprbind(object, siteuse = NULL)
}
\arguments{
\item{object}{A design object from \code{grts()} or \code{irs()}}

\item{siteuse}{A character vector of site types to return. Can contain
\code{"Legacy"} (for legacy sites), \code{"Base"} (for base sites),
\code{"Over"} (for \code{n_over} replacement sites), and \code{"Near"}
(for \code{n_near} replacement sites). The default is \code{NULL}, which
returns all non-\code{NULL} output from \code{object$sites_legacy},
\code{object$sites_base}, \code{object$sites_over}, and \code{object$sites_near}.}
}
\value{
A single \code{sf} object containing rows for the requested sites.
}
\description{
Combine rows from a GRTS or IRS design
}
\examples{
\dontrun{
grts1 <- grts(NE_Lakes, 50, n_over = 10)
sprbind(grts1)
}
}
