% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopprnt.R
\name{stopprnt}
\alias{stopprnt}
\title{Internal Function: Print the stop data Frame that contains errors for the design.}
\usage{
stopprnt(stop_df = get("stop_df", envir = .GlobalEnv), m = 1:nrow(stop_df))
}
\arguments{
\item{stop_df}{Data frame that contains stop messages.  The default is
\code{stop_df}, which is the name given to the stop data frame created by
functions in the spsurvey package.}

\item{m}{Vector of indices for stop messages that are to be printed. The
default is a vector containing the integers from 1 through the number of
rows in \code{stop_df}, which will print all stop messages in the data frame.}
}
\value{
Invisible return.  Prints stop messages
}
\description{
Internal Function: Print the stop data Frame that contains errors for the design.
}
\author{
Tony Olsen \email{Olsen.Tony@epa.gov}
}
