% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.sframe}
\alias{summary.sframe}
\title{Calculate summaries of design objects}
\usage{
\method{summary}{sframe}(object, formula, onlyshow = NULL, ...)
}
\arguments{
\item{object}{A design object output from `grts()` or `irs()` having class
"spsurvey".}

\item{formula}{A formula. Left hand side variables can be numeric or
categorical (or factor) and right hand side variables can be categorical
(or factor). Right hand side variables that are numeric will be coerced
to a categorical (or factor) variable. If an intercept is included in the
right hand side formula, the total will also be summarized.}

\item{onlyshow}{A string indicating the level of the single right hand side
variable for which a summary is requested.}

\item{...}{Additional arguments to pass to `summary()`. If the left hand
side of the formula is empty, the appropriate generic arguments are passed
to summary.data.frame. If the left hand side of the formula is provided, 
the appropriate generic arguments are passed to summary.default}
}
\value{
If the left hand side of the formula is empty, a named list 
containing summaries of the count distribution for each right hand side
varaiable is returned. If the left hand side of the formula contains a
variable, a named list contianing five number
summaries (numeric left hand side) or tables (categorical or factor left
hand side) is returned for each right hand side variable.
}
\description{
`summary` summarizes sample frames or design objects, depending on 
which is provided. For design objects,sites in the base sample and
replacement sites (if they exist) are summarized. The right hand of the formula specifies
the categorical variables (or factors) you want to summarize by. If
the left hand side of the formula is empty, the summary will be of the
count distribution of the right hand side variables. If the left hand side
contains a variable, the summary will be of the left hand size variable
grouped by each of the right hand variables.
}
\examples{
\dontrun{
   summary(NE_lakes, ELEVATION ~ 1)
   sample <- grts(NE_lakes, 100)
   summary(sample, ELEVATION ~ 1)
}
}
