% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_design.R
\name{survey_design}
\alias{survey_design}
\title{Create a Survey Design Object}
\usage{
survey_design(
  dframe,
  siteID,
  weight,
  stratum_ind,
  stratumID,
  cluster_ind,
  clusterID,
  weight1,
  sizeweight,
  sweight,
  sweight1,
  popcorrect,
  fpcsize,
  Ncluster,
  stage1size,
  vartype,
  jointprob
)
}
\arguments{
\item{dframe}{Data frame containing survey design variables, response
variables, and subpopulation (domain) variables.}

\item{siteID}{Character value providing name of the site ID variable in
the \code{dframe} data frame.  For a two-stage sample, the site ID variable
identifies stage two site IDs.}

\item{weight}{Character value providing name of the survey design weight
variable in the \code{dframe} data frame.  For a two-stage sample, the weight
variable identifies stage two weights.}

\item{stratum_ind}{Logical value that indicates whether the sample is
stratified, where \code{TRUE} = a stratified sample and \code{FALSE} = an unstratified
sample.}

\item{stratumID}{Character value providing name of the stratum ID variable in
the \code{dframe} data frame.}

\item{cluster_ind}{Logical value that indicates whether the sample contains
two stages, where \code{TRUE} = a two-stage sample and \code{FALSE} = a single stage
sample.}

\item{clusterID}{Character value providing name of the cluster (stage one) ID
variable in the \code{dframe} data frame.  Note that cluster IDs are required for
a two-stage sample.}

\item{weight1}{Character value providing name of the stage one weight
variable in the \code{dframe} data frame.}

\item{sizeweight}{Logical value that indicates whether size weights should be
used during estimation, where \code{TRUE} = use size weights and \code{FALSE} = do not
use size weights.}

\item{sweight}{Character value providing name of the size weight variable in
the \code{dframe} data frame.  For a two-stage sample, the size weight variable
identifies stage two size weights.}

\item{sweight1}{Character value providing name of the stage one size weight
variable in the \code{dframe} data frame.}

\item{popcorrect}{Logical value that indicates whether the finite population
correction factor is used during variance estimation.}

\item{fpcsize}{Character value providing name of the variable in the \code{dframe}
data frame that identifies size of the resource, which is required for
calculation of the finite population correction factor for a single-stage
sample.}

\item{Ncluster}{Character value providing name of the variable in the \code{dframe}
data frame that identifies the number of clusters (stage one sampling
units) in the resource, which is required for calculation of the finite
population correction factor for a two-stage sample.  This argument is also
required for a two-stage sample when the popsize argument is not equal to
\code{NULL} and the \code{vartype} argument equals \code{"Local"}.}

\item{stage1size}{Character value providing name of the variable in the
\code{dframe} data frame that identifies cluster size, i.e. the number of the
stage two sampling units in the resource for a cluster.  Note that cluster
size is required for calculation of the finite population correction factor
for a two-stage sample.}

\item{vartype}{Character value providing choice of the variance estimator,
where \code{"Local"} = the local mean estimator, \code{"SRS"} = the simple random
sampling estimator, \code{"HT"} = the Horvitz-Thompson estimator, and \code{"YG"} = the
Yates-Grundy estimator.}

\item{jointprob}{Character value providing choice of joint inclusion
probability approximation for use with Horvitz-Thompson and Yates-Grundy
variance estimators, where \code{"overton"} indicates the Overton approximation,
\code{"hr"} indicates the Hartley_Rao approximation, and \code{"brewer"} equals the
Brewer approximation.}
}
\value{
An object of class \code{survey.design2}.
}
\description{
This function creates a survey design object.  The function organizes input
and then calls the svydesign function in the survey package to create an
object of class survey.design2.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{svydesign}}}{specifies a complex survey design}
  }
}

\seealso{
\code{\link{svydesign}}
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
