% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svychisq_localmean.R
\name{svychisq_localmean}
\alias{svychisq_localmean}
\title{Design-Based Contingency Table Tests}
\usage{
svychisq_localmean(
  formula,
  design,
  statistic = c("F", "Chisq", "Wald", "adjWald", "lincom", "saddlepoint"),
  vartype = "Local",
  var_totals = NULL,
  var_means = NULL
)
}
\arguments{
\item{formula}{Two-sided model formula specifying margins for the contingency
table.}

\item{design}{Survey design object created by the \code{svydesign} function.}

\item{statistic}{Character value specifying the test statistic.  See Details.}

\item{vartype}{Character value identifying the choice for variance
estimator, where \code{"Local"} = local mean estimator and \code{"SRS"} = simple random
sampling estimator.  The default is \code{"Local"}.}

\item{var_totals}{Vector containing variance of totals for the contingency
table, which is required for calculating the Wald statistics when vartype
equals \code{"Local"}.  The default is \code{NULL}.}

\item{var_means}{Vector containing variance of means for contingency table
cells, which is required for calculating the Rao-Scott statistics when
vartype equals \code{"Local"}.  The default is \code{NULL}.}
}
\value{
List belonging to class \code{htest}.
}
\description{
This function implements design-based contingency table tests.  It is a
modified version of the \code{svychisq} function in the survey package.  The
modification allows use of the local mean variance algorithms for calculating
variance of contingency table means and totals.
}
\section{Details}{

\code{svychisq_localmean} computes first and second-order Rao-Scott corrections to
the Pearson chi-squared test and two Wald-type tests.  For calculating the
variance of contingency table means and totals, the function can use either
the local mean variance estimator or the simple random sampling variance
estimator.

The default test (statistic equals \code{"F"}) is the Rao-Scott second-order
correction. The p-value is computed using a Satterthwaite approximation to
the distribution with denominator degrees of freedom as recommended by Thomas
and Rao (1990).

The test specified by statistic equals \code{"Chisq"} adjusts the Pearson
chi-squared statistic by a design effect estimate and then compares that
value to the chi-squared distribution it would have under simple random
sampling.

The test specified by statistic equals \code{"Wald"} is the procedure proposed by
Koch et al (1975) and used by the SUDAAN software package. It is a Wald test
based on the differences between the observed cells counts and those expected
under independence.

The test specified by statistic equals \code{"adjWald"} reduces the Wald statistic
when the number of sites is small compared to the number of degrees of
freedom of the test. Thomas and Rao (1990) compared The two Wald tests and
found the adjustment benefical.

The test specified by statistic equals \code{"lincom"} replaces the numerator of the
Rao-Scott F with the exact asymptotic distribution, which is a linear
combination of chi-squared variables (see \code{pchisqsum}).  The CompQuadForm
package is required when statistic equals \code{"lincom"} is specified.

The test specified by statistic equals \code{"saddlepoint"} uses a saddlepoint
approximation to the  exact asymptotic distribution mentioned previously.
The CompQuadForm package is not required when statistic equals \code{"lincom"} is
specified. The saddlepoint approximation is especially useful when the
p-value is very small (e.g., large-scale multiple testing problems).
}

\section{Other Functions Required}{

  \describe{
    \item{\code{\link{cat_localmean_prop}}}{organizes input and output for
      calculation of the local mean variance estimator for estimated
      proportions for categorical data}
    \item{\code{\link{cat_localmean_total}}}{organizes input and output for
      calculation of the local mean variance estimator for estimated sizes
      (totals) for categorical data}
    \item{\code{\link{chisq.test}}}{performs chi-squared contingency table
      tests and goodness-of-fit tests}
    \item{\code{\link{svymean}}}{calculates the mean for a complex survey
      design}
    \item{\code{\link{svytotal}}}{calculates the total for a complex survey
      design}
  }
}

\examples{
n <- 100
resp <- rnorm(n, 10, 1)
wgt <- runif(n, 10, 100)
xcoord <- runif(n)
ycoord <- runif(n)
sample1 <- data.frame(
  samp = "sample1", z = resp, wgt = wgt, x = xcoord,
  y = ycoord
)
sample2 <- data.frame(
  samp = "sample2", z = resp + 0.5, wgt = wgt, x = xcoord,
  y = ycoord
)
bounds <- sort(c(sample1$z, sample2$z))[floor(seq((2 * n) / 3, (2 * n),
  length = 3
))]
dframe <- rbind(sample1, sample2)
dframe$id <- paste0("Site", 1:(2 * n))
dframe$catvar <- cut(dframe$z, c(-Inf, bounds))
design <- survey::svydesign(id = ~id, weights = ~wgt, data = dframe)
svychisq_localmean(~ samp + catvar, design,
  statistic = "adjWald",
  vartype = "SRS"
)
}
\seealso{
\code{\link{cat_localmean_prop}}
  \code{\link{cat_localmean_total}}
  \code{\link{chisq.test}}
  \code{\link{svydesign}}
  \code{\link{svymean}}
  \code{\link{svytotal}}
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
