% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueID.R
\name{uniqueID}
\alias{uniqueID}
\title{Internal Function: Create Unique IDs for a Survey Design}
\usage{
uniqueID(siteID)
}
\arguments{
\item{siteID}{Vector of site IDs.}
}
\value{
Vector of unique site IDs.
}
\description{
This function creates unique site IDs by appending a unique number to each
occurrence of a site ID.  It is intended for survey designs that have repeat
visits to sites.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
