% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPpivotal.R
\name{UPpivotal}
\alias{UPpivotal}
\title{This function is from the R package "sampling". It selects an unequal probability
sample from a finite population.}
\usage{
UPpivotal(pik, eps = 1e-06)
}
\arguments{
\item{pik}{Vector of inclusion probabilities for the sample frame.}

\item{eps}{Control value. The default is 1e-6.}
}
\value{
A vector of \code{0} and \code{1} that indicates if point in the sample frame is
  in the sample.
}
\description{
This function is from the R package "sampling". It selects an unequal probability
sample from a finite population.
}
\references{
Tille, Y. and Matei, A. (2021). sampling: Survey Sampling. R package version 2.9.
  \url{https://CRAN.R-project.org/package=sampling}
}
\author{
From R package sampling
}
\keyword{survey}
