% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UT_ecoregions.R
\docType{data}
\name{UT_ecoregions}
\alias{UT_ecoregions}
\title{Ecoregions in Utah}
\format{
An object of class sf (simple features) containing 10 features and
  3 attributes:
\describe{
  \item{Level3}{Level 3 ecoregion code for the polygon.}
  \item{Level3_Nam}{Level 3 ecoregion name for the polygon.}
  \item{Area_ha}{area of the polygon in hectares.}
}
}
\usage{
UT_ecoregions
}
\description{
A dataset containing attributes for ecoregions in Utah.
}
