% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjust.Weights.R
\name{wgtadjNR}
\alias{wgtadjNR}
\title{Weight adjustment for Unit non response}
\usage{
wgtadjNR(wgt, EvalStatus, MARClass, ENRclass, ERClass)
}
\arguments{
\item{wgt}{Vector of weights for each sample unit that will be adjusted for
unknown eligibility. Weights may be design weights if weights not adjusted
for unknown eligibility. Otherwise they are the design weights.}

\item{EvalStatus}{Vector of evaluation status for each sample unit. Values
must include the values given in ENRclass or ERClass.  May include other
values not required for unit non-response adjustment.}

\item{MARClass}{Vector that identifies for each sample unit the class that
will be used in nonresponse weight adjustment for sample units that are
known to be eligible}

\item{ERClass}{Values in EvalStatus that identify sample units whose
eligibility status is known and that respond (i.e., are sampled)}

\item{ENRClass}{Values in EvalStatus that identify sample units whose
eligibility status is known and that do not respond (i.e., are not sampled)}
}
\value{
Vector of sample unit weights that are adjusted for nonresponse.
}
\description{
Class-based adjustment of weights for eligible sample units that
  do not respond and are missing at random. The missing at random assumption
  implies that their sample weight may be assigned to specific classes of
  units that are have responded (i.e., have been sampled).
}
\details{
Class-based weight adjustment is used to adjust the weights for
  sites that did not respond but are known to be eligible unit in the target
  population.  A ratio adjustment factor is used to proportionally allocate
  the weights associated with sample units that did not respond to units that
  did respond. Weights for sample units that did not response but were known
  to be eligible are set to zero. Weights for all other sample units are
  unchanged.
}
