% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjust.Weights.R
\name{wgtadjOS}
\alias{wgtadjOS}
\title{Weight adjustment of initial design weights}
\usage{
wgtadjOS(wgt, EvalStatus, OSclass, framesize, NN = "NN")
}
\arguments{
\item{wgt}{Vector of initial design weights for each sample unit that will
be adjusted to account for use of sample units greater than (or less than)
the base number of sample units.}

\item{EvalStatus}{Vector giving the evaluation status for each sample unit.}

\item{OSclass}{Vector that identifies for each sample unit the class that
will be used in weight adjustment}

\item{framesize}{Named vector containing the frame size (extent) for each
OSclass.  Names must match names in OSclass}

\item{NN}{A single value or vector of values that identify sample units that
were not evaluated. That is, sites not needed or not used. Values must be a
subset of the unique values in EvalStatus}
}
\value{
Vector of sample unit weights that are adjusted based on site
  evaluation. Length is same as length of wgt.
}
\description{
Adjusts initial survey design weights using a class-based
  method when implementation of the design does not use the planned sample
  size for the design or when it is desired to have final weights sum to a
  known frame size.
}
\details{
Adjusted weights are equal to initial weight times the frame size
  divided by the sum of the initial weights. The adjustment is done
  separately for each class specified in OSclass.  Weights for sample units
  that are not needed (NN) are set to zero.
}
