% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjust.Weights.R
\name{wgtadjU}
\alias{wgtadjU}
\title{Weight adjustment for unknown eligibility}
\usage{
wgtadjU(wgt, EvalStatus, UNKclass, ER = "ER", ENR = "ENR", IN = "IN",
  UNK = "UNK", NN = "NN")
}
\arguments{
\item{wgt}{Vector of design weights for each sample unit that will be
adjusted for unknown eligibility}

\item{EvalStatus}{Vector of evaluation status for each sample unit. Values
must be one of the values given in ER, ENR, IN, UNK, or NN.}

\item{UNKclass}{vector of character vector that identifies for each sample
unit the class that will be used in class-based weight adjustment for
unknown eligibility}

\item{ER}{vector of character values in EvalStatus that identify sample
units whose eligibility status is known and that responded}

\item{ENR}{vector of character values in EvalStatus that identify sample
units whose eligibility status is known and that do not response}

\item{IN}{vector of character values in EvalStatus that identify sample
units that are known to be ineligible}

\item{UNK}{vector of character values in EvalStatus that identify sample
units whose eligibility is unknown}

\item{NN}{vector of character values in EvalStatus that identify sample
units that are not needed or used in the weight adjustment.}
}
\value{
Vector of sample unit weights that are adjusted for uknown
  eligibility.
}
\description{
Adjust weights for sample units that have unknown eligibility
  using class-based method.  Assumes weights are either initial design
  weights when design is implemented as planned or design weights that are
  adjusted when sample size of evaluated sites does not match design sample
  size.
}
\details{
Class-based weight adjustment is used to adjust the weights for
  sites that have unknown eligibility as an unit in the target population.  A
  ratio adjustment factor is used to proportionally allocate the weights
  associated with sample units that have an unknown class to units that have
  known eligibility and known ineligibility. Sample units that are not needed
  (NN) or have unknown eligibility (UNK) have weights set to zero.
}
