% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.dbf.R
\name{write.dbf}
\alias{write.dbf}
\title{Write a Data Frame as the dbf File of an ESRI Shapefile}
\usage{
write.dbf(dframe, filename)
}
\arguments{
\item{dframe}{Data frame to be written to the dbf file}

\item{filename}{Name of the dbf file without any extension.}
}
\value{
Data frame composed of either the contents of the single dbf file,
  when filename is provided, or the contents of the dbf file(s) in the
  working directory, when filename is NULL.
}
\description{
This function writes a data frame to a dbf file.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{writeDbfFile}}{C function to write a single dbf file
      or multiple dbf files}
  }
}

\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
