% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.wt.R
\name{cell.wt}
\alias{cell.wt}
\title{Total Inclusion Probablity for a Cell in a Matrix}
\usage{
cell.wt(cel, xc, yc, dx, dy, pts)
}
\arguments{
\item{cel}{Index value for a cell.}

\item{xc}{x-coordinates that define the cells.}

\item{yc}{y-coordinates that define the cells.}

\item{dx}{Width of the cells along the x-axis.}

\item{dy}{Width of the cells along the y-axis.}

\item{pts}{Data frame containing x-coordinates, y-coordinates, and mdm
(inclusion probability) values.}
}
\value{
The total inclusion probability for the cell.
}
\description{
Calculates the total inclusion probability for a cell from a finite
population suvey design.
}
\author{
Tony Olsen \email{Olsen.Tony@epa.gov}
}
\keyword{survey}
