% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_analysis.R
\name{cont_analysis}
\alias{cont_analysis}
\title{Continuous Data Analysis for Probability Survey Data}
\usage{
cont_analysis(
  dframe,
  vars,
  vars_nondetect = NULL,
  subpops = NULL,
  siteID = "siteID",
  weight = "weight",
  xcoord = NULL,
  ycoord = NULL,
  stratumID = NULL,
  clusterID = NULL,
  weight1 = NULL,
  xcoord1 = NULL,
  ycoord1 = NULL,
  sizeweight = FALSE,
  sweight = NULL,
  sweight1 = NULL,
  fpc = NULL,
  popsize = NULL,
  vartype = "Local",
  jointprob = "overton",
  conf = 95,
  pctval = c(5, 10, 25, 50, 75, 90, 95),
  statistics = c("cdf", "pct", "mean")
)
}
\arguments{
\item{dframe}{Data frame containing survey design variables, response
variables, and subpopulation (domain) variables.}

\item{vars}{Vector composed of character values that identify the
names of response variables in the \code{dframe} data frame.}

\item{vars_nondetect}{Vector composed of character values that identify the
names of logical variables in the \code{dframe} data frame specifying the
presence of not detected (nondetect) values for response variables.  The
order of the values for this argument must match the order of the values
for the \code{vars} argument. Each logical variable specifies the detection
status for the corresponding values of a response variable, where
\code{TRUE} equals not detected, and \code{FALSE} equals detected.  If a
response variable does not include any notdetect values, then the name for
the corresponding logical variable may be set equal to \code{NULL}.  If
this argument equals \code{NULL}, then none of the response variables
contain nondetect values.  The default is \code{NULL}.}

\item{subpops}{Vector composed of character values that identify the
names of subpopulation (domain) variables in the \code{dframe} data frame.
If a value is not provided, the value \code{"All_Sites"} is assigned to the
subpops argument and a factor variable named \code{"All_Sites"} that takes
the value \code{"All Sites"} is added to the \code{dframe} data frame.  The
default value is \code{NULL}.}

\item{siteID}{Character value providing name of the site ID variable in
the \code{dframe} data frame.  For a two-stage sample, the site ID variable
identifies stage two site IDs.  The default value is \code{"siteID"}.}

\item{weight}{Character value providing name of the survey design weight
variable in the \code{dframe} data frame.  For a two-stage sample, the
weight variable identifies stage two weights.  The default value is
\code{"weight"}.}

\item{xcoord}{Character value providing name of the x-coordinate variable in
the \code{dframe} data frame.  For a two-stage sample, the x-coordinate
variable identifies stage two x-coordinates.  Note that x-coordinates are
required for calculation of the local mean variance estimator.  The default
value is \code{NULL}.}

\item{ycoord}{Character value providing name of the y-coordinate variable in
the \code{dframe} data frame.  For a two-stage sample, the y-coordinate
variable identifies stage two y-coordinates.  Note that y-coordinates are
required for calculation of the local mean variance estimator.  The default
value is \code{NULL}.}

\item{stratumID}{Character value providing name of the stratum ID variable in
the \code{dframe} data frame.  The default value is \code{NULL}.}

\item{clusterID}{Character value providing the name of the cluster
(stage one) ID variable in the \code{dframe} data frame.  Note that cluster
IDs are required for a two-stage sample.  The default value is \code{NULL}.}

\item{weight1}{Character value providing name of the stage one weight
variable in the \code{dframe} data frame.  The default value is \code{NULL}.}

\item{xcoord1}{Character value providing the name of the stage one
x-coordinate variable in the \code{dframe} data frame.  Note that x
coordinates are required for calculation of the local mean variance
estimator.  The default value is \code{NULL}.}

\item{ycoord1}{Character value providing the name of the stage one
y-coordinate variable in the \code{dframe} data frame.  Note that
y-coordinates are required for calculation of the local mean variance
estimator.  The default value is \code{NULL}.}

\item{sizeweight}{Logical value that indicates whether size weights should be
used during estimation, where \code{TRUE} = use size weights and
\code{FALSE} = do not use size weights. To employ size weights for a
single-stage sample, a value must be supplied for argument weight.  To
employ size weights for a two-stage sample, values must be supplied for
arguments \code{weight} and \code{weight1}. The default value is
\code{FALSE}.}

\item{sweight}{Character value providing the name of the size weight variable
in the \code{dframe} data frame.  For a two-stage sample, the size weight
variable identifies stage two size weights.  The default value is
\code{NULL}.}

\item{sweight1}{Character value providing name of the stage one size weight
variable in the \code{dframe} data frame.  The default value is
\code{NULL}.}

\item{fpc}{Object that specifies values required for calculation of the
finite population correction factor used during variance estimation. The
object must match the survey design in terms of stratification and whether
the design is single-stage or two-stage.  For an unstratified design, the
object is a vector.  The vector is composed of a single numeric value for a
single-stage design.  For a two-stage unstratified design, the object is a
named vector containing one more than the number of clusters in the sample,
where the first item in the vector specifies the number of clusters in the
population and each subsequent item specifies the number of stage two units
for the cluster.  The name for the first item in the vector is arbitrry.
Subsequent names in the vector identify clusters and must match the cluster
IDs.  For a stratified design, the object is a named list of vectors, where
names must match the strata IDs.  For each stratum, the format of the
vector is identical to the format described for unstratified single-stage
and two-stage designs.  Note that the finite population correction factor
is not used with the local mean variance estimator.\cr\cr
Example fpc for a single-stage unstratified survey design:\cr
\verb{fpc <- 15000}\cr\cr
Example fpc for a single-stage stratified survey design:\cr
\verb{fpc <- list(
  Stratum_1 = 9000,
  Stratum_2 = 6000)
 }\cr
Example fpc for a two-stage unstratified survey design:\cr
\verb{fpc <- c(
  Ncluster = 150,
  Cluster_1 = 150,
  Cluster_2 = 75,
  Cluster_3 = 75,
  Cluster_4 = 125,
  Cluster_5 = 75)
}\cr
Example fpc for a two-stage stratified survey design:\cr
\verb{fpc <- list(
  Stratum_1 = c(
    Ncluster_1 = 100,
    Cluster_1 = 125,
    Cluster_2 = 100,
    Cluster_3 = 100,
    Cluster_4 = 125,
    Cluster_5 = 50),
  Stratum_2 = c(
    Ncluster_2 = 50,
    Cluster_1 = 75,
    Cluster_2 = 150,
    Cluster_3 = 75,
    Cluster_4 = 75,
    Cluster_5 = 125))
}}

\item{popsize}{Object that provides values for the population argument of
the \code{calibrate} or \code{postStratify} functions.  For the
\code{calibrate} function, the object is a named list, where the names
identify factor variables in the \code{dframe} data frame.  Each element
of the list is a named vector containing the population total for each
level of the associated factor variable.  For the \code{postStratify}
function, the object is either a data frame, table, or xtabs
object that provides the population total for all combinations of selected
factor varaibles in the \code{dframe} data frame.  If a data frame is used
for \code{popsize}, the variable containing population totals must be the
last variable in the data frame.  If a table is used for \code{popsize},
the table must have named \code{dimnames} where the names identify factor
variables in the \code{dframe} data frame.  If the popsize argument is
equal to \code{NULL}, then neither calibration nor post-stratification is
performed.  The default value is
\code{NULL}.\cr\cr
Example popsize for calibration:\cr
\verb{popsize <- list(
  Ecoregion = c(
    East = 750,
    Central = 500,
    West = 250),
  Type = c(
    Streams = 1150,
    Rivers = 350))
}\cr
Example popsize for post-stratification using a data frame:\cr
\verb{popsize <- data.frame(
  Ecoregion = rep(c("East", "Central", "West"),
    rep(2, 3)),
  Type = rep(c("Streams", "Rivers"), 3),
  Total = c(575, 175, 400, 100, 175, 75))
}\cr
Example popsize for post-stratification using a table:\cr
\verb{popsize <- with(MySurveyFrame,
  table(Ecoregion, Type))}\cr\cr
Example popsize for post-stratification using an xtabs object:\cr
\verb{popsize <- xtabs(~Ecoregion + Type,
  data = MySurveyFrame)}}

\item{vartype}{Character value providing the choice of the variance
estimator, where "Local" = the local mean estimator, \code{"SRS"} = the
simple random sampling estimator, \code{"HT"} = the Horvitz-Thompson
estimator, and \code{"YG"} = the Yates-Grundy estimator.  The default value
is \code{"Local"}.}

\item{jointprob}{Character value providing the choice of joint inclusion
probability approximation for use with Horvitz-Thompson and Yates-Grundy
variance estimators, where \code{"overton"} indicates the Overton
approximation, \code{"hr"} indicates the Hartley_Rao approximation, and
\code{"brewer"} equals the Brewer approximation.  The default value is
\code{"overton"}.}

\item{conf}{Numeric value providing the confidence level.  The default value
is \code{95}.}

\item{pctval}{Vector of the set of values at which percentiles are
estimated.  The default set is: \code{c(5, 10, 25, 50, 75, 90, 95)}.}

\item{statistics}{Character vector specifying desired estimates, where
\code{"cdf"} specifies CDF estimates, \code{"pct"} specifies percentile
estimates, and \code{"mean"} specifies mean estimates.  Any combination of
the three choices may be provided by the user.  The defalt value is
\code{c("cdf", "pct", "mean")}.}
}
\value{
A list composed of one, two, or three data frames that contain
  population estimates for all combinations of subpopulations, categories
  within each subpopulation, and response variables, where the number of data
  frames is determined by argument statistics.  Estimates are provided for
  proportion and total (size) of the population plus standard error, margin
  of error, and confidence interval estimates.  The possible data frames in
  the output list are:
  \describe{
    \item{\code{CDF}}{data frame containing the CDF estimates}
    \item{\code{Pct}}{data frame containing the percentile estimates}
    \item{\code{Mean}}{data frame containing the mean estimates}
  }
}
\description{
This function organizes input and output for analysis of continuous data
generated by a probability survey.  The input data argument can be either a
data frame or a simple features (sf) object.  If an sf object is used,
coordinates are extracted from the geometry column in the object, arguments
\code{xcoord} and \code{ycoord} are assigned values \code{"xcoord"} and
\code{"ycoord"}, respectively, and the geometry column is dropped from the
object.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{calibrate}}}{conduct calibration for survey data}
    \item{\code{cdf_est}}{calculates an estimate of the cumulative
      distribution function}
    \item{\code{input_check}}{check input values for errors,
      consistency, and compatibility with analytical functions}
    \item{\code{percentile_est}}{calculates percentile estimates}
    \item{\code{\link{postStratify}}}{conduct post-stratification for survey
      data}
    \item{\code{survey_design}}{creates a survey design object}
    \item{\code{uniqueID}}{creates unique site IDs by appending a
      unique number to each occurrence of a site ID}
    \item{\code{vecprint}}{takes an input vector and outputs a
      character string with line breaks inserted}
  }
}

\examples{
dframe <- data.frame(
  siteID = paste0("Site", 1:100),
  wgt = runif(100, 10, 100),
  xcoord = runif(100),
  ycoord = runif(100),
  stratum = rep(c("Stratum1", "Stratum2"), 50),
  ContVar = rnorm(100, 10, 1),
  All_Sites = rep("All Sites", 100),
  Resource_Class = rep(c("Good", "Poor"), c(55, 45))
)
myvars <- c("ContVar")
mysubpops <- c("All_Sites", "Resource_Class")
mypopsize <- data.frame(
  Resource_Class = c("Good", "Poor"),
  Total = c(4000, 1500)
)
cont_analysis(dframe,
  vars = myvars, subpops = mysubpops, siteID = "siteID",
  weight = "wgt", xcoord = "xcoord", ycoord = "ycoord",
  stratumID = "stratum", popsize = mypopsize, statistics = "mean"
)
}
\seealso{
\code{\link{calibrate}}
  \code{\link{postStratify}}
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
\keyword{univar}
