% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_cdftest.R
\name{cont_cdftest}
\alias{cont_cdftest}
\title{Cumulative Distribution Function Inference for a Probability Survey}
\usage{
cont_cdftest(
  dframe,
  vars,
  subpops = NULL,
  surveyID = NULL,
  siteID = "siteID",
  weight = "weight",
  xcoord = NULL,
  ycoord = NULL,
  stratumID = NULL,
  clusterID = NULL,
  weight1 = NULL,
  xcoord1 = NULL,
  ycoord1 = NULL,
  sizeweight = FALSE,
  sweight = NULL,
  sweight1 = NULL,
  popcorrect = FALSE,
  fpcsize = NULL,
  Ncluster = NULL,
  stage1size = NULL,
  popsize = NULL,
  vartype = "Local",
  jointprob = "overton",
  testname = "adjWald",
  nclass = 3
)
}
\arguments{
\item{dframe}{Data frame containing survey design variables, response
variables, and subpopulation (domain) variables.}

\item{vars}{Vector composed of character values that identify the
names of response variables in the dframe data frame.}

\item{subpops}{Vector composed of character values that identify the
names of subpopulation (domain) variables in the dframe data frame.  If a
value is not provided, the value "All_Sites" is assigned to the subpops
argument and a factor variable named "All_Sites" that takes the value
"All Sites" is added to the dframe data frame.  The default value is NULL.}

\item{surveyID}{Character value providing name of the survey ID variable in
the dframe data frame.  If this argument equals NULL, then the dframe data
frame contains data for a single survey.  The default value is NULL.}

\item{siteID}{Character value providing name of the site ID variable in
the dframe data frame.  For a two-stage sample, the site ID variable
identifies stage two site IDs.  The default value is "siteID".}

\item{weight}{Character value providing name of the survey design weight
variable in the dframe data frame.  For a two-stage sample, the weight
variable identifies stage two weights.  The default value is "weight".}

\item{xcoord}{Character value providing name of the x-coordinate variable in
the dframe data frame.  For a two-stage sample, the x-coordinate variable
identifies stage two x-coordinates.  Note that x-coordinates are required
for calculation of the local mean variance estimator.  The default value is
NULL.}

\item{ycoord}{Character value providing name of the y-coordinate variable in
the dframe data frame.  For a two-stage sample, the y-coordinate variable
identifies stage two y-coordinates.  Note that y-coordinates are required
for calculation of the local mean variance estimator.  The default value is
NULL.}

\item{stratumID}{Character value providing name of the stratum ID variable in
the dframe data frame.  The default value is NULL.}

\item{clusterID}{Character value providing name of the cluster (stage one) ID
variable in the dframe data frame.  Note that cluster IDs are required for
a two-stage sample.  The default value is NULL.}

\item{weight1}{Character value providing name of the stage one weight
variable in the dframe data frame.  The default value is NULL.}

\item{xcoord1}{Character value providing name of the stage one x-coordinate
variable in the dframe data frame.  Note that x-coordinates are required
for calculation of the local mean variance estimator.  The default value is
NULL.}

\item{ycoord1}{Character value providing name of the stage one y-coordinate
variable in the dframe data frame.  Note that y-coordinates are required
for calculation of the local mean variance estimator.  The default value is
NULL.}

\item{sizeweight}{Logical value that indicates whether size weights should be
used during estimation, where TRUE = use size weights and FALSE = do not
use size weights. To employ size weights for a single-stage sample, a value
must be supplied for argument weight.  To employ size weights for a
two-stage sample, values must be supplied for arguments weight and weight1.
The default value is FALSE.}

\item{sweight}{Character value providing name of the size weight variable in
the dframe data frame.  For a two-stage sample, the size weight variable
identifies stage two size weights.  The default value is NULL.}

\item{sweight1}{Character value providing name of the stage one size weight
variable in the dframe data frame.  The default value is NULL.}

\item{popcorrect}{Logical value that indicates whether the finite population
correction factor is used during variance estimation. To employ the
correction factor for a single-stage sample, values must be supplied for
argument fpcsize.  To employ the correction factor for a two-stage sample,
values must be supplied for arguments Ncluster and stage1size.  The default
value is FALSE.}

\item{fpcsize}{Character value providing name of the variable in the dframe
data frame that identifies size of the resource, which is required for
calculation of the finite population correction factor for a single-stage
sample.  The default value is NULL.}

\item{Ncluster}{Character value providing name of the variable in the dframe
data frame that identifies the number of clusters (stage one sampling
units) in the resource, which is required for calculation of the finite
population correction factor for a two-stage sample.  This argument is also
required for a two-stage sample when the popsize argument is not equal to
NULL and the vartype argument equals "Local".  The default value is NULL.}

\item{stage1size}{Character value providing name of the variable in the
dframe data frame that identifies cluster size, i.e. the number of the
stage two sampling units in the resource for a cluster.  Note that cluster
size is required for calculation of the finite population correction factor
for a two-stage sample. The default value is NULL.}

\item{popsize}{Object that provides values for the population argument of the
calibrate or postStratify functions.  For the calibrate function, the object
is a named list, where the names identify factor variables in the dframe
data frame.  Each element of the list is a named vector containing the
population total for each level of the associated factor variable.  For the
postStratify function, the object is either a data frame, table, or xtabs
object that provides the population total for all combinations of selected
factor variables in the dframe data frame.  If a data frame is used for
popsize, the variable containing population totals must be the last
variable in the data frame.  If a table is used for popsize, the table must
have named dimnames where the names identify factor variables in the dframe
data frame.  If the popsize argument is equal to NULL, then neither
calibration nor post-stratification is performed.  The default value is
NULL.\cr\cr
Example popsize for calibration:\cr
  popsize <- list(Ecoregion = c(East = 750,\cr
                                Central = 500,\cr
                                West = 250),\cr
                  Type = c(Streams = 1150,\cr
                           Rivers = 350))\cr
Example popsize for post-stratification using a data frame:\cr
  popsize <- data.frame(
    Ecoregion = rep(c("East", "Central", "West"), rep(2, 3)),\cr
    Type = rep(c("Streams", "Rivers"), 3),\cr
    Total = c(575, 175, 400, 100, 175, 75))\cr
Example popsize for post-stratification using a table:\cr575
  popsize <- with(MySurveyFrame, table(Ecoregion, Type))\cr
Example popsize for post-stratification using an xtabs object:\cr
  popsize <- xtabs(~Ecoregion + Type, data = MySurveyFrame)\cr}

\item{vartype}{Character value providing choice of the variance estimator,
where "Local" = the local mean estimator, "SRS" = the simple random
sampling estimator, "HT" = the Horvitz-Thompson estimator, and "YG" = the
Yates-Grundy estimator.  The default value is "Local".}

\item{jointprob}{Character value providing choice of joint inclusion
probability approximation for use with Horvitz-Thompson and Yates-Grundy
variance estimators, where "overton" indicates the Overton approximation,
"hr" indicates the Hartley_Rao approximation, and "brewer" equals the
Brewer approximation.  The default value is "overton".}

\item{testname}{Name of the test statistic to be reported in the output
data frame.  Choices for the name are: "Wald", "adjWald", "RaoScott_First",
and "RaoScott_Second", which correspond to the Wald statistic, adjusted
Wald statistic, Rao-Scott first-order corrected statistic, and Rao-Scott
second-order corrected statistic, respectively.  The default is "adjWald".}

\item{nclass}{Number of classes into which the CDFs will be divided
(binned), which must equal at least two.  The default is 3.}
}
\value{
Data frame of CDF test results for all pairs of subpopulations
  within each population Type for every response variable.  The data frame
  includes the test statistic specified by argument testname plus its degrees
  of freedom and p-value.
}
\description{
This function organizes input and output for conducting inference regarding
cumulative distribution functions (CDFs) generated by a probability survey.
For every response variable and every subpopulation (domain) variable,
differences between CDFs are tested for every pair of subpopulations within
the domain.  Data input to the function can be either a single surey or
multiple surveys (two or more).  If the data contain multile surveys, then
the domain variables will reference those surveys and (potentially)
subpopulations within those surveys.  The inferential procedures divide the
CDFs into a discrete set of intervals (classes) and then utilize procedures
that have been developed for analysis of categorical data from probability
surveys.  Choices for inference are the Wald, adjusted Wald, Rao-Scott first
order corrected (mean eigenvalue corrected), and Rao-Scott second order
corrected (Satterthwaite corrected) test statistics. The default test
statistic is the adjusted Wald statistic.  The input data argument can be
either a data frame or a simple features (sf) object.  If an sf object is
used, coordinates are extracted from the geometry column in the object,
arguments xcoord and ycoord are assigned values "xcoord" and "ycoord",
respectively, and the geometry column is dropped from the object.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{calibrate}}}{conduct calibration for survey data}
    \item{\code{\link{cdftest_localmean_prop}}}{calculates local mean
      variance/covarince estimates of estimated population proportions}
    \item{\code{\link{cdftest_localmean_total}}}{calculates local mean
      variance/covarince estimates of estimated population totals}
    \item{\code{\link{input_check}}}{check input values for errors,
      consistency, and compatibility with analytical functions}
    \item{\code{\link{postStratify}}}{conduct post-stratification for survey
      data}
    \item{\code{\link{svychisq_localmean}}}{performs design-based contingency
      table tests}
    \item{\code{\link{survey_design}}}{creates a survey design object}
    \item{\code{\link{uniqueID}}}{creates unique site IDs by appending a
      unique number to each occurrence of a site ID}
    \item{\code{\link{vecprint}}}{takes an input vector and outputs a
      character string with line breaks inserted}
  }
}

\examples{
n <- 200
mysiteID <- paste("Site", 1:n, sep="")
dframe <- data.frame(
  siteID = mysiteID,
  wgt = runif(n, 10, 100),
  xcoord = runif(n),
  ycoord = runif(n),
  stratum = rep(c("Stratum1", "Stratum2"), n/2),
  Resource_Class = sample(c("Agr", "Forest", "Urban"), n, replace=TRUE))
ContVar <- numeric(n)
tst <- dframe$Resource_Class == "Agr"
ContVar[tst] <- rnorm(sum(tst), 10, 1)
tst <- dframe$Resource_Class == "Forest"
ContVar[tst] <- rnorm(sum(tst), 10.1, 1)
tst <- dframe$Resource_Class == "Urban"
ContVar[tst] <- rnorm(sum(tst), 10.5, 1)
dframe$ContVar <- ContVar
myvars <- c("ContVar")
mysubpops <- c("Resource_Class")
mypopsize <- data.frame(
  Resource_Class = rep(c("Agr", "Forest", "Urban"), rep(2, 3)),
  stratum = rep(c("Stratum1", "Stratum2"), 3),
  Total = c(2500, 1500, 1000, 500, 600, 450))
cont_cdftest(dframe, vars = myvars, subpops = mysubpops, siteID = "siteID",
  weight = "wgt", xcoord = "xcoord", ycoord = "ycoord",
  stratumID = "stratum", popsize = mypopsize, testname="RaoScott_First")

}
\seealso{
\code{\link{calibrate}}
  \code{\link{postStratify}}
  \code{\link{svychisq_localmean}}
  \code{\link{survey_design}}
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
\keyword{survey}
