% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dframe.R
\name{dframe}
\alias{dframe}
\title{Create an dframe object}
\usage{
dframe(object)
}
\arguments{
\item{object}{An \code{sp} or \code{sf} object}
}
\value{
An \code{dframe} object
}
\description{
dframe gives data frames, \code{sp}, and \code{sf} objects class \code{dframe} to be used
in \code{summary()} generics. \code{sp} objects are coerced to \code{sf}
objects prior to giving class \code{dframe}.
}
\examples{
\dontrun{
data("NE_lakes")
NE_lakes <- dframe(NE_lakes)
}
}
\author{
Michael Dumelle
}
