% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dframe.R
\name{dframe}
\alias{dframe}
\title{Create an dframe object}
\usage{
dframe(object)
}
\arguments{
\item{object}{An `sp` or `sf` object}
}
\value{
An `dframe` object
}
\description{
dframe gives data frames, `sp`, and `sf` objects class `"dframe"` to be used
in `summary()` generics. `sp` objects are coerced to `sf`
objects prior to giving class `dframe`.
}
\examples{
\dontrun{
data("NE_lakes")
NE_lakes <- dframe(NE_lakes)
}
}
