% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dframe.R
\name{dframe}
\alias{dframe}
\title{Create an dframe object}
\usage{
dframe(object, ...)
}
\arguments{
\item{object}{An \code{sp} or \code{sf} object.}

\item{...}{Additional arguments to \code{st_as_sf()} if converting an \code{sp}
object to an \code{sf} and \code{dframe} object.}
}
\value{
A \code{dframe} object that can be used by
spsurvey's \code{summary()} and \code{plot()} generics.
}
\description{
\code{dframe()} is used before summarizing or visualizing
  data that has been collected (this function need not be run before calling any analysis functions).
  \code{dframe()} gives data frames, \code{sp} objects, and \code{sf}
  objects class \code{dframe}. \code{sp} objects are coerced to \code{sf}
  objects prior to giving class \code{dframe}.
}
\examples{
data("NLA_PNW")
NLA_PNW <- dframe(NLA_PNW)
summary(NLA_PNW, formula = ~ BMMI_COND)
plot(NLA_PNW, formula = ~ BMMI_COND)
}
\seealso{
\describe{
    \item{\code{\link{dframe_to_sf}}}{ to turn a \code{dframe} object into an \code{sf} object}
    \item{\code{\link{dframe_to_df}}}{ to turn a \code{dframe} object into a data frame}
  }
}
\author{
Michael Dumelle \email{Dumelle.Michael@epa.gov}
}
