% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dframe_to_df.R
\name{dframe_to_df}
\alias{dframe_to_df}
\title{Transform a dframe object into a data frame object}
\usage{
dframe_to_df(object)
}
\arguments{
\item{object}{A \code{dframe} object.}
}
\value{
A  data frame object (\code{dframe_to_df()}), potentially with coordinates
added as extra columns.
}
\description{
\code{dframe_to_df()} transforms a \code{dframe} object into an \code{data.frame} object. If
the \code{dframe} object is also an \code{sf} object, \code{dframe_to_df()} stores the
coordinates as columns in the new \code{data.frame} object.
The remaining structure of \code{dframe} stays the same (e.g. if \code{dframe} is also a tibble,
it will remain a tibble after \code{dframe_to_df()}).
}
\examples{
data("NLA_PNW")
NLA_PNW <- dframe(NLA_PNW)
NLA_PNW <- dframe_to_df(NLA_PNW)
}
\seealso{
\describe{
    \item{\code{\link{dframe}}}{ to create a \code{dframe} object}
    \item{\code{\link{dframe_to_sf}}}{ to turn a \code{dframe} object into an \code{sf} object}
  }
}
\author{
Michael Dumelle \email{Dumelle.Michael@epa.gov}
}
