% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dframe_to_sf.R
\name{dframe_to_sf}
\alias{dframe_to_sf}
\title{Transform a dframe object into an sf object or a data frame object}
\usage{
dframe_to_sf(object, ...)
}
\arguments{
\item{object}{A \code{dframe} object.}

\item{...}{Additional arguments to \code{st_as_sf()} if spatial information is not already
present in \code{dframe}}
}
\value{
An \code{sf} object.
}
\description{
\code{dframe_to_sf()} transforms an sframe object into an \code{sf} object,
which may be useful when calling generic functions with sf class (e.g. mapview).
}
\seealso{
dframe dframe_to_df
}
