% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsgn_check.R
\name{dsgn_check}
\alias{dsgn_check}
\title{Check the input associated with the survey design list object and the sample frame}
\usage{
dsgn_check(
  sframe,
  sf_type,
  legacy_sites,
  legacy_option,
  stratum,
  seltype,
  n_base,
  caty_n,
  n_over,
  n_near,
  stratum_var,
  caty_var,
  aux_var,
  legacy_var,
  mindis,
  DesignID,
  SiteBegin,
  maxtry
)
}
\arguments{
\item{sframe}{Sample frame as an sf object.}

\item{sf_type}{Geometry type for sample frame: point, linear, or area.}

\item{legacy_sites}{An sf object of any legacy sites to be included in the survey
design.}

\item{legacy_option}{Logical variable where if equal to TRUE, legacy sites are
to be incorporated into survey design.}

\item{stratum}{Single character value or character vector that identifies the strata for
the design.  Default is NULL which means no stratification and all elements in the sample
frame are assumed to be included.}

\item{seltype}{Single character value or character vector that identifies the type of 
random selection, which must be one of following: "equal" for equal probability selection, 
"unequal" for unequal probability selection by the categories specified in caty_n or 
"proportional" for unequal probability selection proportional to the auxiliary variable
aux_var. If single character, then seltype applies to all strata. If vector, then each 
stratum may have different selection type. Default is single character value of "equal".}

\item{n_base}{The sample size required. If single stratum, then single numeric value.
If sample is stratified, then numeric vector with same length as "stratum" and sample sizes
required in same order as strata in "stratum". Must be specified.}

\item{caty_n}{If design is not stratified and seltype is "unequal", a named character vector
with the expected sample size for each category specified in variable caty_var. If design
is stratified, then either a named character vector with the expected sample size for each
category for all strata or if the expected sample size for each category may differ, then
a list of named character vectors with the expected sample size for each category in the
stratum. The list must be in same order as the "stratum" variable. For each stratum, 
the sum of caty_n values must equal n_base for that stratum. Default is NULL.}

\item{n_over}{If seltype is "equal" and is not stratified, a numeric value specifying the 
over sample size requested. If seltype is "equal" and is stratified either a numeric value
specifying the over sample size that will be applied to each stratum or a numeric vector
specifying the over sample size for each stratum listed in same order as "strata".  
If seltype is "unequal" and is not stratified, a named character vector with the over sample size 
for each category where names are the same as "caty_n". If seltype is "unequal" and is 
stratified,  either a numeric vector specifying the over sample size for each category in
"caty_n" that will be applied to each stratum or a list of named numeric vectors with 
the over sample size for each "caty_n" category for each stratum. List must be in same order
as the "stratum" variable order. Default is NULL.}

\item{n_near}{Numeric value specifying the number of nearby points to select as
possible replacement sites if a site cannot be sampled. Default is NULL. If specified,
must be integer from 1 to 10.}

\item{stratum_var}{Character string containing the name of the column from
sframe that identifies stratum membership for each element in the frame.
If stratum equals NULL, the design is unstratified and all elements in sample frame
are eligible to be selected in the sample. The default is NULL.}

\item{caty_var}{Character string containing the name of the column from
sframe that identifies the unequal probability category for each element
in the frame.  Default is NULL.}

\item{aux_var}{Character string that is the name of the column from sframe that
identifies the auxiliary variable value for each element in the sample frame
that will be used to calculate inclusion probabilities when the survey design
specifies that the selection type (seltype) is "proportional". Default is NULL.}

\item{legacy_var}{Character value for name of column for legacy site variable.
Default is NULL.}

\item{mindis}{Numeric value for the minimum distance required between elements
in the sample. If design is stratified, then mindis applies only within each stratum.
Units must be the same units as in sf geometry. Default is NULL.}

\item{DesignID}{Name for the design, which is used to create a site ID for
each site.  Default is "Site".}

\item{SiteBegin}{Number to use for first site in the design.  Default is 1.}

\item{maxtry}{Number of maximum attempts to ensure minimum distance (mindis) between sites.
Default is 10.}
}
\value{
Nothing is returned. If errors are found they are collected and written out.
  One or more errors will cause the call to grts to stop.
}
\description{
Check the input associated with the survey design list object and the sample frame
}
\author{
Tony Olsen email{olsen.tony@epa.gov}
}
