% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorprnt.R
\name{errorprnt}
\alias{errorprnt}
\title{Internal Function: Print the Error Messages Vector}
\usage{
errorprnt(error_vec = get("error_vec", envir = .GlobalEnv))
}
\arguments{
\item{error_vec}{Data frame that contains error messages.  The default is
"error_vec", which is the name given to the error messagess vector created
by functions in the spsurvey package.}
}
\value{
Invisible return.  Prints errors.
}
\description{
This function prints the error messages vector.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
