% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorprnt.R
\name{errorprnt}
\alias{errorprnt}
\title{Print errors from analysis functions}
\usage{
errorprnt(error_vec = get("error_vec", envir = .GlobalEnv))
}
\arguments{
\item{error_vec}{Data frame that contains error messages.  The default is
\code{"error_vec"}, which is the name given to the error messages vector created
by functions in the spsurvey package.}
}
\value{
Printed errors.
}
\description{
This function prints the error messages vector in the analysis functions.
}
\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
