% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Great_Lakes}
\alias{Great_Lakes}
\title{Great Lakes}
\format{
An \code{sf} multipolygon object with 4113 rows and 5 variables:
\describe{
  \item{COUNTRY}{Country.}
  \item{WTBDY_NM}{Water body name.}
  \item{RSRC_CLASS}{Bay class.}
  \item{PSTL_CODE}{Postal code.}
  \item{AREA_SQKM}{Area in square kilometers}
  \item{geometry}{simple features geometry list-column}
}
}
\usage{
Great_Lakes
}
\description{
An \code{sf} multipolygon object of 4113 polygons consisting
of shoreline segments in the Great Lakes region of North America.
}
\keyword{datasets}
