% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grts.R
\name{grts}
\alias{grts}
\title{Select a generalized random tessellation stratified (GRTS) sample}
\usage{
grts(
  sframe,
  n_base,
  stratum_var = NULL,
  seltype = NULL,
  caty_var = NULL,
  caty_n = NULL,
  aux_var = NULL,
  legacy_var = NULL,
  legacy_sites = NULL,
  legacy_stratum_var = NULL,
  mindis = NULL,
  maxtry = 10,
  n_over = NULL,
  n_near = NULL,
  wgt_units = NULL,
  pt_density = NULL,
  DesignID = "Site",
  SiteBegin = 1
)
}
\arguments{
\item{sframe}{The sample frame as an \code{sf} object. The coordinate
system for \code{sframe} must be one where distance for coordinates is meaningful.}

\item{n_base}{The base sample size required. If the design is unstratified,
this is a single numeric value. If the design is stratified, this is a named
vector whose names represent each stratum and whose values represent each
stratum's sample size. These names must match the values of the stratification
variable represented by \code{stratum_var}.}

\item{stratum_var}{A character string containing the name of the column from
\code{sframe} that identifies stratum membership for each element in the frame.
If stratum equals \code{NULL}, the design is unstratified and all elements in \code{sframe}
are eligible to be selected in the sample. The default is \code{NULL}.}

\item{seltype}{A character string or vector indicating the inclusion probability type,
which must be one of following: \code{"equal"} for equal inclusion probabilities;
\code{"unequal"} for unequal inclusion probabilities according to a categorical
variable specified by \code{caty_var}; and \code{"proportional"} for inclusion
probabilities proportional to a positive auxiliary variable. If the design is
unstratified, \code{seltype} is a single character vector. If the design is stratified, \code{seltype} is a named vector
whose names represent each stratum and whose values represent each stratum's
inclusion probability type. \code{seltype}'s default value tries to match the
intended inclusion probability type: If \code{caty_var} and \code{aux_var} are
not specified, \code{seltype} is \code{"equal"}; if \code{caty_var} is specified,
\code{seltype} is \code{"unequal"}; and if \code{aux_var} is specified, \code{seltype}
is \code{"proportional"}.}

\item{caty_var}{A character string containing the name of the column from
\code{sframe} that represents the unequal probability variable.}

\item{caty_n}{A character vector indicating the expected sample size for each
level of \code{caty_var}, the unequal probability variable. If the design
is unstratified, \code{caty_n} is a named vector whose names represent each
level of \code{caty_var} and whose values represent each level's expected
sample size. The sum of \code{caty_n} must equal \code{n_base}. If the design
is stratified and the expected sample sizes are the same among strata, \code{caty_n} is
a named vector whose names represent represent each
level of \code{caty_var} and whose values represent each level's expected
sample size -- these expected sample sizes are applied to all strata. The sum of
\code{caty_n} must equal each stratum's value in \code{n_base}.
If the design is stratified and the expected sample sizes differ among strata,
\code{caty_n} is a list where each element  is named as a stratum in \code{n_base}.
Each stratum's list element is a named vector whose
names represent each level of \code{caty_var} and whose values represent each
level's expected sample size (within the stratum). The sum of the values in each stratum's
list element must equal that stratum's value in \code{n_base}.}

\item{aux_var}{A character string containing the name of the column from
\code{sframe} that represents the proportional (to size) inclusion probability
variable (auxiliary variable). This auxiliary variable must be positive, and the resulting
inclusion probabilities are proportional to the values of the auxiliary variable.
Larger values of the auxiliary variable result in higher inclusion probabilities.}

\item{legacy_var}{If \code{sframe} is a \code{POINT} or \code{MULTIPOINT} geometry (a finite sample frame),
\code{legacy_var} is a character string containing the name of the column
from \code{sframe} that represents the legacy site variable. For legacy sites, the values of the
\code{legacy_var} column in \code{sframe} must contain character strings that
act as a legacy site identifier. For non-legacy sites, the values of the
\code{legacy_var} column in \code{sframe} must be \code{NA}.}

\item{legacy_sites}{If \code{sframe} is a \code{LINESTRING}, \code{MULTILINESTRING},
\code{POLYGON}, or \code{MULTIPOLYGON} geometry (an infinite sample frame),
\code{legacy_sites} is an sf object with a \code{POINT} geometry representing the
legacy sites.}

\item{legacy_stratum_var}{A character string containing the name of the column from
\code{legacy_sites} that identifies stratum membership for each element of \code{legacy_sites}.
This argument is required when the design is stratified and its levels
must be contained in the levels of \code{stratum_var}. The default value of \code{legacy_stratum_var}
is \code{stratum_var}, so \code{legacy_stratum_var} need only be specified explicitly when
the name of the stratification variable in \code{legacy_sites} differs from \code{stratum_var}.}

\item{mindis}{A numeric value indicating the desired minimum distance between sampled
sites. If design is stratified, then mindis is applied separately for each stratum.
The units of \code{mindis} must match the units in \code{sframe}.}

\item{maxtry}{The number of maximum attempts to apply the minimum distance algorithm to obtain
the desired minimum distance between sites. Each iteration takes roughly as long as the
standard GRTS algorithm. Successive iterations will always contain at least as many
sites satisfying the minimum distance requirement as the previous iteration. The algorithm stops
when the minimum distance requirement is met or there are \code{maxtry} iterations.
The default number of maximum iterations is \code{10}.}

\item{n_over}{If the design is unstratified and \code{seltype} is \code{"equal"} or \code{"proportional"},
\code{n_over} is an integer specifying the number of reverse hierarchically
ordered (rho) replacement sites desired.  If the design is unstratified and \code{seltype} is \code{"unequal"},
then \code{n_over} is a named character vector whose names match the levels of
\code{caty_var} and whose values are the expected rho replacement sample sizes for each
level. If the design is stratified and \code{seltype} is \code{"equal"} or \code{"proportional"},
\code{n_over} is a named vector whose names match the names of \code{n_base} and whose values
indicate the number of rho replacement sites for each stratum. If the design is stratified and
\code{seltype} is \code{"unequal"}, \code{n_over} changes based on whether the expected
rho replacement sample sizes change among strata. If \code{n_over} does not change among strata, \code{n_over}
is a named vector whose names match the names of \code{caty_var} and whose values are the expected rho replacement
sample sizes to be used for each stratum. If \code{n_over} changes among strata, \code{n_over} is a
list where each element represents a stratum in \code{n_base}.
The order of the strata in this list must match the order of the strata in \code{n_base}.
Each stratum's list element is a named vector whose names represent each level of \code{caty_var} and whose values represent each
level's expected rho replacement sample sizes (within the stratum).}

\item{n_near}{An integer from \code{1} to \code{10} specifying the number of
nearest neighbor replacement sites to be selected for each base site.}

\item{wgt_units}{The units used to compute the survey design weights. These
units must be standard units as defined by the \code{set_units()} function in
the units package. The default units match the units of the sf object.}

\item{pt_density}{A numeric value controlling the density of the GRTS approximation
for infinite sample frames. The GRTS approximation for infinite sample
frames vastly improves computational efficiency by generating many finite points and
selecting a sample from the points. \code{pt_density} represents the density
of finite points per unit to use in the approximation (and the units match
the units of the sample frame. The default is a density
such that the number of finite points used in the approximation equals 10
times the sample size requested.}

\item{DesignID}{A character string indicating the naming structure for each
site's identifier selected in the sample, which is included as a variable in the
sf object in the function's output.}

\item{SiteBegin}{A character string indicating the first number to use to match
with \code{DesignID} while creating each site's identifier selected in the sample. Successive
sites are given successive integers. The default starting number is \code{1}.}
}
\value{
A list with five elements:
  \itemize{
    \item \code{sites_legacy} An sf object containing legacy sites. This is
      \code{NULL} if legacy sites were not included in the sample.
    \item \code{sites_base} An sf object containing the base sites.
    \item \code{sites_over} An sf object containing the reverse hierarchically
      ordered replacement sites. This is \code{NULL} if no reverse hierarchically
      ordered replacement sites were included in the sample.
    \item \code{sites_near} An sf object containing the nearest neighbor
      replacement sites. This is \code{NULL} if no nearest neighbor replacement
      sites were included in the sample.
    \item \code{design} A list documenting the specifications of this design.
      This can be checked to verify your design ran as intended.
      \itemize{
        \item \code{Call} The original function call.
        \item \code{stratum} The unique strata. This equals \code{"None"} if
          the design was unstratified.
        \item \code{n_base} The base sample size per stratum.
        \item \code{seltype} The selection type per stratum.
        \item \code{caty_n} The expected sample sizes for each level of the
          unequal probability grouping variable per stratum. This equals
          \code{NULL} when \code{seltype} is not \code{"unequal"}.
        \item \code{legacy} A logical variable indicating whether legacy sites
          were included in the sample.
        \item \code{mindis} The minimum distance requirement desired. This
          equals \code{NULL} if there was no minimum distance requirement.
        \item \code{n_over} The reverse hierarchically ordered replacement
          site sample sizes per stratum. If \code{seltype} is \code{unequal},
          this represents the expected sample sizes. This is \code{NULL}
          if no reverse hierarchically ordered replacement sites were included
          in the sample.
        \item \code{n_near} The number of nearest neighbor replacement sites
          desired. This is \code{NULL} if no nearest neighbor replacement
          sites were included in the sample.
      }
  }
  When non-\code{NULL}, the \code{sites_legacy}, \code{sites_base},
  \code{sites_over}, and \code{sites_near} objects contain the original columns
  in \code{sframe} and include a few additional columns. These additional columns
  are
  \itemize{
    \item \code{siteID} A site identifier (as named using the \code{DesignID}
      and \code{SiteBegin} arguments to \code{grts()}).
    \item \code{siteuse} Whether the site is a legacy site (\code{Legacy}), base
      site (\code{Base}), reverse hierarchically ordered replacement site
      (\code{Over}), or nearest neighbor replacement site (\code{Near}).
    \item \code{replsite} The replacement site ordering. \code{replsite} is
      \code{None} if the site is not a replacement site, \code{Next} if it is
      the next reverse hierarchically ordered replacement site to use, or
      \code{Near_}, where the word following \code{_} indicates the ordering of sites closest to
      the originally sampled site.
    \item \code{lon_WGS84} Longitude coordinates using the WGS84 coordinate
      system (EPSG:4326).
    \item \code{lat_WGS84} Latitude coordinates using the WGS84 coordinate
      system (EPSG:4326).
    \item \code{stratum} A stratum indicator. \code{stratum} is \code{None}
      if the design was unstratified. If the design was \code{stratified},
      \code{stratum} indicates the stratum.
    \item \code{wgt} The survey design weight.
    \item \code{ip} The site's original inclusion probability (the reciprocal)
      of (\code{wgt}).
    \item \code{caty} An unequal probability grouping indicator. \code{caty}
      is \code{None} if the design did not use unequal inclusion probabilities.
      If the design did use unequal inclusion probabilities, \code{caty}
      indicates the unequal probability level.
    \item \code{aux} The auxiliary proportional probability variable. This
      column is only returned if \code{seltype} was \code{proportional} in the
      original design.
  }
}
\description{
Select a spatially balanced sample from a point (finite), linear (infinite),
or polygon / areal (infinite) sample frame using the Generalized Random Tessellation
Stratified (GRTS) algorithm. The GRTS algorithm accommodates unstratified and
stratified designs and allows for equal inclusion probabilities, unequal
inclusion probabilities according to a categorical variable, and inclusion
probabilities proportional to a positive auxiliary variable. Several additional
sampling options are included, such as including legacy (historical) sites,
requiring a minimum distance between sites, and selecting replacement sites.
}
\examples{
sample <- grts(NE_Lakes, n_base = 100)
strata_n <- c(low = 25, high = 30)
sample_strat <- grts(NE_Lakes, n_base = strata_n, stratum_var = "ELEV_CAT")
sample_over <- grts(NE_Lakes, n_base = 30, n_over = 5)
}
\references{
Stevens Jr., Don L. and Olsen, Anthony R. (2004). Spatially balanced sampling
of natural resources. \emph{Journal of the american Statistical association}, 99(465), 262-278.
}
\seealso{
\describe{
    \item{\code{\link{irs}}}{ to select a sample that is not spatially balanced}
 }
}
\author{
Tony Olsen \email{olsen.tony@epa.gov}
}
\keyword{design}
\keyword{survey}
