% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grts_stratum.R
\name{grts_stratum}
\alias{grts_stratum}
\title{For a single stratum, select a spatially balanced sample using generalized random
tessalation stratified algorithm. For a point sample frame, the selection is a one-step
process. For linear and area sample frames, a two-step process is used. First a systematic
point sample is selected from either linear or area sample frame to create a dense point
sample. Then the grts algorithm is applied to the systematic point sample. The selection
in all cases is based on the survey design specification for the stratum.}
\usage{
grts_stratum(
  stratum,
  dsgn,
  sframe,
  sf_type,
  pt_density = NULL,
  legacy_option = NULL,
  legacy_sites = NULL,
  maxtry = 10,
  warn_ind = FALSE,
  warn_df = NULL
)
}
\arguments{
\item{stratum}{Character value for the stratum name.}

\item{dsgn}{List of componenents that specify the survey design. Includes all strata.
See grts for contents of dsgn.}

\item{sframe}{Sample frame as an sf object. If the design is stratified,
unequal probability, proportional probability or has legacy sites, then sample frame 
must include variables that identify the stratum; category, auxillary and legacy variables
for unequal selection; or that identify elements that are legacy sites. 
The coordinate system for sframe must be one where distance for coordinates is meaningful.}

\item{sf_type}{The sample frame geometry type: point, linear or area}

\item{pt_density}{For linear and area sample frame, the point density for the systematic
sample. Must be in units of the sframe sf.object. Default is NULL.}

\item{legacy_option}{Logical variable that when TRUE legacy sites are to be included 
in the survey design. Default is FALSE}

\item{legacy_sites}{An sf object of legacy sites to be included in the survey design.}

\item{maxtry}{Number of maximum attempts to ensure minimum distance between sites.
Default is 10.}

\item{warn_ind}{A logical value where TRUE indicates a warning message.
Used for internal collection of messages only.}

\item{warn_df}{A data frame containing messages warning of potential issues.
Used for internal collection of messages only.}
}
\value{
rslts A list consisting of an sf object for base sites, an sf object of 
  n_over sites (NULL if none) an sf object of n_near sites (NULL if none) where
  the sf objects containing the sites selected that meet the survey design requirements 
  and warn_ind - logical value for warning indicator and warn_df - a data.frame
  for warning messages.
}
\description{
For a single stratum, select a spatially balanced sample using generalized random
tessalation stratified algorithm. For a point sample frame, the selection is a one-step
process. For linear and area sample frames, a two-step process is used. First a systematic
point sample is selected from either linear or area sample frame to create a dense point
sample. Then the grts algorithm is applied to the systematic point sample. The selection
in all cases is based on the survey design specification for the stratum.
}
\section{Other functions required}{

  \describe{
    \item{\code{\link{grtspts_ip}}}{Calculates inclusion probabilities}
    \item{\code{\link{grtspts_ipleg}}}{Calculates inclusion probabilities with legacy points}
    \item{\code{\link{get_address}}}{Creates hierarchical order for points}
    \item{\code{\link{UPpivotal}}}{Selects sample using pivotal method}
    \item{\code{\link{rho}}}{Orders sites in reverse hierarchical order}
    \item{\code{\link{grtspts_mindis}}}{selects a grts sample with minimum distance between
      points}
    }
}

\examples{
\dontrun{
  test_sample <- grts(stratum, dsgn=test_design, sframe = "test_sf", sf_type = "point")
}

}
\author{
Tony Olsen email{olsen.tony@epa.gov}
}
\keyword{survey}
