% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Illinois_River}
\alias{Illinois_River}
\title{Illinois River}
\format{
A data frame with 9087 rows and 3 variables:
\describe{
  \item{STATE_NAME}{.....}
  \item{SO_ID}{.....}
  \item{geometry}{.....}
}
}
\usage{
Illinois_River
}
\description{
A simple features (sf) multilinestring object of 9087 river segments created
from document..... The variables are:
}
\keyword{datasets}
