% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Illinois_River}
\alias{Illinois_River}
\title{Illinois River data}
\format{
244 rows and 2 variables:
\describe{
  \item{\code{STATE_NAME}}{ State name.}
  \item{\code{geometry}}{ MULTILINESTRING geometry using the NAD83 / Conus Albers
    coordinate reference system (EPSG: 5070).}
}
}
\usage{
Illinois_River
}
\description{
An (\code{sf}) MULTILINESTRING object of 244 segments of the
Illinois River in Arkansas and Oklahoma.
}
\keyword{datasets}
