% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Illinois_River}
\alias{Illinois_River}
\title{Illinois River data}
\format{
An \code{sf} multilinestring object object with 9087 rows and 3
variables:
\describe{
  \item{\code{STATE_NAME}}{ State name.}
  \item{\code{geometry}}{ Multilinestring geometry using a NAD83 / Conus Albers
    projection (EPSG: 5070).}
}
}
\usage{
Illinois_River
}
\description{
A simple features (\code{sf}) multilinestring object of 9087 segments of the
Illinois River in Arkansas and Oklahoma.
}
\keyword{datasets}
