% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Illinois_River}
\alias{Illinois_River}
\title{Illinois River}
\format{
An \code{sf} multilinestring object object with 9087 rows and 3
variables:
\itemize{
  \item{STATE_NAME}{State name.}
  \item{geometry}{Multilinestring geometry using a USA Contiguous Albers
    Equal Area Conic projection.}
}
}
\usage{
Illinois_River
}
\description{
A simple features (\code{sf}) multilinestring object of 9087 segments of the
Illinois River in Arkansas and Oklahoma.
}
\keyword{datasets}
