% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Illinois_River_Legacy}
\alias{Illinois_River_Legacy}
\title{Illinois River Legacy}
\format{
An \code{sf} point object object with 5 rows and 1
variable:
\describe{
  \item{geometry}{Point geometry using a USA Contiguous Albers
    Equal Area Conic projection.}
}
}
\usage{
Illinois_River_Legacy
}
\description{
A simple features (\code{sf}) point object of legacy sites for  the Illinois
River data.
}
\keyword{datasets}
