% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_check.R
\name{input_check}
\alias{input_check}
\title{Check Input Values for Analytical Functions}
\usage{
input_check(
  dframe,
  design_names,
  vars_cat,
  vars_cont,
  vars_stressor,
  vars_nondetect,
  subpops,
  sizeweight,
  popcorrect,
  popsize,
  vartype,
  jointprob,
  conf,
  cdfval = NULL,
  pctval = NULL,
  sigma = NULL,
  var_sigma = NULL,
  error_ind,
  error_vec
)
}
\arguments{
\item{dframe}{Data frame containing survey design variables, response
variables, and subpopulation (domain) variables.}

\item{design_names}{Vector composed of character values that identify the
names of the following survey design variables in the dframe argument:
siteID, weight, xcoord, ycoord, stratumID, clusterID, weight1, xcoord1,
ycoord1, sweight, sweight1, fpcsize, Ncluster, stage1size.}

\item{vars_cat}{Vector composed of character values that identify the
names of categorical response variables in the dframe data frame.}

\item{vars_cont}{Vector composed of character values that identify the
names of continuous response variables in the dframe data frame.}

\item{vars_stressor}{Vector composed of character values that identify the
names of stressor variables in the dframe argument.}

\item{vars_nondetect}{Vector composed of character values that identify the
names of logical variables in the dframe data frame specifying the presence
of not detected (nondetect) values for response variables.}

\item{subpops}{Vector composed of character values that identify the
names of subpopulation variables in the dframe argument.}

\item{sizeweight}{Logical value that indicates whether size weights should be
used during estimation.}

\item{popcorrect}{Logical value that indicates whether the finite population
correction factor is used during variance estimation.}

\item{popsize}{List providing known size of the resource.}

\item{vartype}{Character value providing choice of the variance estimator,
where "Local" = the local mean estimator, "SRS" = the simple random
sampling estimator, "HT" = the Horvitz-Thompson estimator, and "YG" = the
Yates-Grundy estimator.}

\item{jointprob}{Character value providing choice of joint inclusion
probability approximation for use with Horvitz-Thompson and Yates-Grundy
variance estimators, where "overton" indicates the Overton approximation,
"hr" indicates the Hartley_Rao approximation, and "brewer" equals the
Brewer approximation.}

\item{conf}{Numeric value providing the confidence level.}

\item{cdfval}{Vector of the set of values at which the CDF is estimated.}

\item{pctval}{Vector of the set of values at which percentiles are
estimated.}

\item{sigma}{Numeric value providing the measurement error variance.}

\item{var_sigma}{Numeric value providing the variance of the measurement
error variance.}

\item{error_ind}{Logical value that indicates whether error messages were
generated.}

\item{error_vec}{Vector for storing error messages.}
}
\value{
A list consisting of dframe, vars_cat, vars_cont, vars_stressor,
  subpops, popsize, vartype, jointprob, error_ind, and error_vec.
}
\description{
This function checks input values for errors, consistency, and compatibility
with analytical functions.
}
\section{Other Functions Required}{

  \describe{
    \item{\code{\link{vecprint}}}{takes an input vector and outputs a
      character string with line breaks inserted}
  }
}

\author{
Tom Kincaid \email{Kincaid.Tom@epa.gov}
}
