% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lake_Ontario}
\alias{Lake_Ontario}
\title{Lake Ontario data}
\format{
An \code{sf} multipolygon object with 187 rows and 5 variables:
\describe{
  \item{\code{COUNTRY}}{ Country.}
  \item{\code{RSRC_CLASS}}{ Bay class.}
  \item{\code{PSTL_CODE}}{ Postal code.}
  \item{\code{AREA_SQKM}}{ Area in square kilometers}
  \item{\code{geometry}}{ Multipolygon geometry using the NAD83 / Conus Albers
    coordinate reference system (EPSG: 5070).}
}
}
\usage{
Lake_Ontario
}
\description{
An \code{sf} multipolygon object of 187 polygons consisting
of shoreline segments in Lake Ontario.
}
\keyword{datasets}
