% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCCA_Gulf}
\alias{NCCA_Gulf}
\title{NCCA Gulf}
\format{
A data frame with 851 rows and 11 variables:
\describe{
  \item{SITE_ID}{A unique site identifier.}
  \item{YEAR}{Year of design cycle.}
  \item{REVISIT}{Site revisit status.}
  \item{WEIGHT}{Sampling design weights.}
  \item{SIZE}{Size category.}
  \item{STATE}{State name.}
  \item{B_MMI}{Benthic MMI value.}
  \item{BMMI_COND}{Benthic MMI categories.}
  \item{FT_COND}{Fish tissue condition categories.}
  \item{PHOS_COND}{Phosphorus condition categories.}
  \item{geometry}{Point geometry using a USA Contiguous Albers Equal Area
    Conic projection.}
}
}
\usage{
NCCA_Gulf
}
\description{
An sf object data frame consisting of gulf observations from a subset of the
Environmental Protection Agency's "National Coastal Condition
Assessment" in the Southern United States duing the years 2010 and 2015.
}
\keyword{datasets}
