% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NE_lakes.R
\docType{data}
\name{NE_lakes}
\alias{NE_lakes}
\title{Lakes in the Southern New England Region of the U.S.}
\format{An object of class SpatialPointsDataFrame.\cr
The data slot contains 6,121 rows and 4 variables:
\describe{
  \item{xcoord}{Albers projection x-coordinate of the lake centroid.}
  \item{ycoord}{Albers projection y-coordinate of the lake centroid.}
  \item{State}{state code for the lake.}
  \item{Area_Cat}{lake surface area category in hectares.}
}}
\usage{
NE_lakes
}
\description{
A dataset containing attributes for Lakes in the Southern New England Region
of the U.S.
}
