% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NE_Lakes}
\alias{NE_Lakes}
\title{New England Lakes}
\format{
A data frame with 5328 rows and 7 variables:
\describe{
  \item{AREA_HA}{.....}
  \item{AREA_HA_CAT}{.....}
  \item{ELEVATION}{.....}
  \item{ELEVATION_CAT}{.....}
  \item{LAT_DD_N83}{.....}
  \item{LON_DD_N83}{.....}
  \item{geometry}{.....}
}
}
\usage{
NE_Lakes
}
\description{
A simple features (sf) point object of 5328 lakes created from document.....
The variables are:
}
\keyword{datasets}
