% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NE_Lakes}
\alias{NE_Lakes}
\title{New England Lakes data}
\format{
An \code{sf} object with 200 rows and 6 variables:
\describe{
  \item{\code{AREA}}{ Lake area in hectares.}
  \item{\code{AREA_CAT}}{ Lake area categories based on a hectare cutoff.}
  \item{\code{ELEV}}{ Elevation in meters.}
  \item{\code{ELEV_CAT}}{ Elevation categories based on a meter cutoff.}
  \item{\code{LEGACY}}{ Legacy site indicator.}
  \item{\code{geometry}}{ Point geometry using the NAD83 / Conus Albers
    coordinate reference system (EPSG: 5070).}
}
}
\usage{
NE_Lakes
}
\description{
An \code{sf} point object of 200 lakes in the Northeastern
United States.
}
\keyword{datasets}
