% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NE_Lakes_df}
\alias{NE_Lakes_df}
\title{New England Lakes data (as a data frame)}
\format{
195 rows and 6 variables:
\describe{
  \item{\code{AREA}}{ Lake area in hectares.}
  \item{\code{AREA_CAT}}{ Lake area categories based on a hectare cutoff.}
  \item{\code{ELEV}}{ Elevation in meters.}
  \item{\code{ELEV_CAT}}{ Elevation categories based on a meter cutoff.}
  \item{\code{XCOORD}}{ x-coordinate using the WGS 84 coordinate reference system
    (EPSG: 4326)}
  \item{\code{YCOORD}}{ y-coordinate using WGS 84 coordinate reference system
    (EPSG: 4326)}
}
}
\usage{
NE_Lakes_df
}
\description{
An data frame of 195 lakes in the Northeastern
United States.
}
\keyword{datasets}
