% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NLA_PNW}
\alias{NLA_PNW}
\title{NLA PNW}
\format{
An \code{sf} point object with 851 rows and 9 variables:
\itemize{
  \item{SITE_ID}{A unique lake identifier.}
  \item{WEIGHT}{The sampling design weight.}
  \item{URBAN}{Urban category.}
  \item{STATE}{State name.}
  \item{BMMI}{Benthic MMI value.}
  \item{BMMI_COND}{Benthic MMI condition categories.}
  \item{PHOS_COND}{Phosphorus condition categories.}
  \item{NITR_COND}{Nitrogen condition categories.}
  \item{geometry}{Point geometry using a USA Contiguous Albers Equal Area
    Conic projection.}
}
}
\usage{
NLA_PNW
}
\description{
A data frame consisting lake observations from a subset of the Environmental
Protection Agency's "National Lakes Assessment" in the Western United
States during the year 2017.
}
\keyword{datasets}
