% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NRSA_EPA7}
\alias{NRSA_EPA7}
\title{NRSA EPA7 data}
\format{
A data frame with 682 rows and 10 variables:
\describe{
  \item{\code{SITE_ID}}{ A unique site identifier.}
  \item{\code{YEAR}}{ Year of design cycle.}
  \item{\code{WEIGHT}}{ Sampling design weights.}
  \item{\code{ECOREGION}}{ Ecoregion.}
  \item{\code{STATE}}{ State name.}
  \item{\code{BMMI}}{ Benthic MMI value.}
  \item{\code{BMMI_COND}}{ Benthic MMI categories.}
  \item{\code{PHOS_COND}}{ Phosphorus condition categories.}
  \item{\code{NITR_COND}}{ Nitrogen condition categories.}
  \item{\code{geometry}}{ Point geometry using a NAD83 / Conus Albers
    projection (EPSG: 5070).}
}
}
\usage{
NRSA_EPA7
}
\description{
An sf object consisting of stream observations from a subset of the Environmental
Protection Agency's "National Rivers and Streams Assessment" in the Central
United States duing the years 2008 and 2013.
}
\keyword{datasets}
