% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NRSA_EPA7}
\alias{NRSA_EPA7}
\title{NRSA EPA7}
\format{
A data frame with 682 rows and 18 variables:
\describe{
  \item{UNIQUE_ID}{.....}
  \item{SITE_ID}{.....}
  \item{DSGN_CYCLE}{.....}
  \item{LAT_DD83}{.....}
  \item{LON_DD83}{.....}
  \item{XCOORD}{.....}
  \item{YCOORD}{.....}
  \item{WGT_TP_EXTENT}{.....}
  \item{WGT_SP_CORE}{.....}
  \item{WGT_TP_CORE}{.....}
  \item{EVAL_CAT}{.....}
  \item{TNT_CAT}{.....}
  \item{NA_LEVEL1_ECO_NM}{.....}
  \item{STATE_NM}{.....}
  \item{BENTHIC_MMI}{.....}
  \item{BENTHIC_MMI_COND}{.....}
  \item{PHOSPHORUS_COND}{.....}
  \item{NITROGEN_COND}{.....}
}
}
\usage{
NRSA_EPA7
}
\description{
A data frame of stream information from a "National Rivers and Streams Assessment"..... The variables:
}
\keyword{datasets}
