% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.sframe}
\alias{plot.design}
\title{Plot sample frame and design objects}
\usage{
\method{plot}{sframe}(
  x,
  y,
  formula = ~1,
  variable_args = NULL,
  level_args = NULL,
  geom = FALSE,
  onlyshow = NULL,
  fix_bbox = TRUE,
  ...
)

\method{plot}{design}(
  x,
  y = NULL,
  formula = ~sites,
  sites = NULL,
  variable_args = NULL,
  level_args = NULL,
  geom = FALSE,
  onlyshow = NULL,
  fix_bbox = TRUE,
  showlegacy = FALSE,
  ...
)
}
\arguments{
\item{x}{object of class "sframe" or "design"}

\item{y}{ignored if `x` has class "sframe"; an object of class "sframe"
if `x` has class "design"}

\item{formula}{A formula. Left hand side variables can be numeric or
categorical (or factor) and right hand side variables can be categorical
(or factor). Right hand side variables that are numeric will be coerced to
a categorical (or factor) variable. If an intercept is included in the right
hand side formula, the total will also be summarized.}

\item{variable_args}{A named list. The name of each list corresponds to a
right hand side variable in `formula`. Values in the list are composed of 
graphical arguments that are to be passed to \strong{every} level of the
variable.}

\item{level_args}{A named list. The name of each list corresponds to a
right hand side variable in `formula`. The first name of each sublist 
should be "levels" and contain all levels of the variable. Subsequent
names correspond to graphical arguments that are to be passed to
the specified levels (in order). Values must be specified
for every level of each graphical argument, but applicable `sf` defaults
will be matched by inputting the value `NA`}

\item{geom}{Should separate geometries for each level of the right hand
side `formula` variables be plotted? Defaults to `FALSE`}

\item{onlyshow}{A string indicating the level of the single right hand side
variable for which a summary is requested.}

\item{fix_bbox}{Should the geometry bounding box be fixed across plots? 
Defaults to `TRUE`. If `TRUE`, the bounding box will be the largest
possible given `x` and `y`.}

\item{...}{Additional arguments to pass to `plot.sf()`.}
}
\value{

}
\description{
Plot sample frames and design objects. This function is
largely built on `plot.sf()`, and all `spsurvey` plotting methods can
supply additional arguments to `plot.sf()`. For more information
regarding `plot.sf()`, run `?plot.sf()`.
}
