% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_near.R
\name{replace_near}
\alias{replace_near}
\title{Replacement sites generated via nearest neighbors}
\usage{
replace_near(over.near, sites, sframe)
}
\arguments{
\item{over.near}{number of nearby sites to be used as potential replacement(s)
if a site cannot be sampled for any reason. Must be integer from \code{1} to \code{10}.
Default is \code{NULL}.}

\item{sites}{sf point object of selected sites that require replacement sites}

\item{sframe}{sf point object that was used as sample frame to select sites.}
}
\description{
Returns 1 to 10 nearest neighbor replacement sites for each site in the 
  base sample
}
\author{
Michael Dumelle \email{Dumelle.Michael@epa.gov} and Tony Olsen \email{Olsen.Tony@epa.gov}
}
\keyword{design}
\keyword{survey}
