% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sframe.R
\name{sframe}
\alias{sframe}
\title{Create an sframe object}
\usage{
sframe(object)
}
\arguments{
\item{object}{An `sp` or `sf` object}
}
\value{
An `sframe` object
}
\description{
sframe gives `sp` and `sf` objects class `"sframe"` to be used
in `summary()` and `plot()` generics. `sp` objects are coerced to `sf`
objects prior to giving class `sframe`.
}
\examples{
\dontrun{
data("NE_lakes")
NE_lakes <- sframe(NE_lakes)
}
}
