% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sframe.R
\name{sframe}
\alias{sframe}
\title{Create an sframe object}
\usage{
sframe(object, ...)
}
\arguments{
\item{object}{An \code{sp} or \code{sf} object}

\item{...}{Additional arguments to \code{st_as_sf()} if converting an sp object to an sf and sframe object}
}
\value{
An \code{sframe} object
}
\description{
sframe gives \code{sp} and \code{sf} objects class \code{"sframe"} to be used
in \code{summary()} and \code{plot()} generics. \code{sp} objects are coerced to \code{sf}
objects prior to giving class \code{sframe}.
}
\examples{
\dontrun{
data("NE_lakes")
NE_lakes <- sframe(NE_lakes)
}
###############################################################################
}
