% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sframe.R
\name{sframe}
\alias{sframe}
\title{Create an sframe object}
\usage{
sframe(object, ...)
}
\arguments{
\item{object}{For \code{sframe}, an \code{sp} or \code{sf} object. For \code{unsframe()}, an
\code{sframe} object.}

\item{...}{Additional arguments to \code{st_as_sf()} when using \code{sframe()}
to convert an sp object to an sf and sframe object}
}
\value{
For \code{sframe()}, an \code{sframe} object. For \code{unsframe()}, an \code{sf} object.
}
\description{
\code{sframe()} gives \code{sp} and \code{sf} objects class \code{"sframe"} to be used
in spsurvey's \code{summary()} and \code{plot()} generics. \code{sp} objects are coerced to \code{sf}
objects prior to giving class \code{sframe}. \code{unsframe()} removes class \code{"sframe"},
which may be useful when calling sf generic functions.
}
\examples{
\dontrun{
data("NE_lakes")
NE_lakes <- sframe(NE_lakes)
}
}
\seealso{
sframe_to_sf
}
