% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbalance.R
\name{spbalance}
\alias{spbalance}
\title{Calculate spatial balance metrics}
\usage{
spbalance(
  object,
  sframe,
  sites = "sites_base",
  metrics = "pielou",
  extents = FALSE
)
}
\arguments{
\item{object}{A design object output from \code{grts()} or \code{irs()}
having class \code{"design"}.}

\item{sframe}{An \code{sframe} or \code{sf} object.}

\item{sites}{The sites in object for which spatial balance metrics
are desired. This argument must be \code{"sites_base"}, \code{"sites_over"}, or
\code{"sites_near"}. Defaults to \code{"sites_base"}.}

\item{metrics}{A character vector of spatial balance metrics:
\itemize{
  \item{pielou - }{pielou evenness index (the default)}
  \item{simpsons - }{simpsons evenness index}
  \item{chisq - }{chi-squared loss}
  \item{pielou - }{absolute error loss}
 }}

\item{extents}{Should the total extent within each dirichlet
tesselation be returned? Defaults to FALSE.}
}
\value{
A list having names equal to each strata.
}
\description{
This function computes the spatial balance of a design with respect
to the sample frame (\code{sframe}) using Dirichlet Tesselations, measuring the
extent to which a object is a miniature of \code{sframe}. This function is
applicable for unstratified or stratified designs with equal selection
probabilities.
}
\examples{
\dontrun{
sample <- grts(NE_Lakes, 30)
spbalance(sample, NE_Lakes)
}
###############################################################################
}
