% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbalance.R
\name{spbalance}
\alias{spbalance}
\title{Calculate spatial balance metrics}
\usage{
spbalance(
  object,
  sframe,
  stratum_var = NULL,
  ip = NULL,
  metrics = "pielou",
  extents = FALSE
)
}
\arguments{
\item{object}{An \code{sf} object containig the sampled sites.}

\item{sframe}{An \code{sframe} or \code{sf} object containing the sample frame.}

\item{stratum_var}{The name of the stratum variable in \code{object}
and \code{sframe}. If \code{NULL} (the default), no strata is assumed.
If a single character vector is provided, it is assumed this is the
name of the stratum variable in \code{object} and \code{sframe}. If
a two-dimensional character vector is provided, one element must be
named "object" and corresponds to the name of the stratum variable
in \code{object}, while the other element must be named "sframe" and
correponds to the name of the stratum variable in \code{sframe}.}

\item{ip}{Inclusion probabilities associated with each row of \code{sframe}.
If these are not provided, an equal probability design is assumed (within
strata).}

\item{metrics}{A character vector of spatial balance metrics:
\itemize{
  \item{\code{pielou}:}{pielou evenness index (the default)}
  \item{\code{simpsons}:}{simpsons evenness index}
  \item{\code{rmse}:}{root-mean-squared error}
  \item{\code{mse}:}{mean-squared error}
  \item{\code{mae}:}{median-absolute error}
  \item{\code{medae}:}{mean-absolute error}
  \item{\code{chisq}:}{chi-squared loss}
 }
  All spatial balance metrics have a lower bound of zero, which indicates perfect
  spatial balance. As the metric value increases, the spatial balance decreases.}

\item{extents}{Should the total extent within each dirichlet
tesselation be returned? Defaults to FALSE.}
}
\value{
A list having names equal to each strata.
}
\description{
This function computes the spatial balance of a design with respect
to the sample frame (\code{sframe}) using Dirichlet Tesselations, measuring the
extent to which a object is a miniature of \code{sframe}. This function is
applicable for unstratified or stratified designs with equal selection
probabilities.
}
\examples{
\dontrun{
sample <- grts(NE_Lakes, 30)
spbalance(sample, NE_Lakes)
}
###############################################################################
}
