% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spbalance.R
\name{spbalance}
\alias{spbalance}
\title{Spatial Balance}
\usage{
spbalance(
  object,
  sframe,
  stratum_object = "stratum",
  stratum_sframe,
  metrics = "pielou",
  return_extent = FALSE
)
}
\arguments{
\item{object}{}

\item{sframe}{}

\item{stratum_object}{}

\item{stratum_sframe}{}

\item{metrics}{}

\item{return_extent}{}
}
\value{

}
\description{
This function computes the spatial balance of a design with respect 
to the sframe using Dirichlet Tesselations, measuring the
extent to which a object is a miniature of the sframe.
}
