% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprbind.R
\name{sprbind}
\alias{sprbind}
\title{Combine rows from GRTS or IRS samples.}
\usage{
sprbind(object, siteuse = NULL)
}
\arguments{
\item{object}{A sample from \code{grts()} or \code{irs()}}

\item{siteuse}{A character vector of site types to return. Can contain
\code{"Legacy"} (for legacy sites), \code{"Base"} (for base sites),
\code{"Over"} (for \code{n_over} replacement sites), and \code{"Near"}
(for \code{n_near} replacement sites). The default is \code{NULL}, which
returns all non-\code{NULL} output from \code{object$sites_legacy},
\code{object$sites_base}, \code{object$sites_over}, and \code{object$sites_near}.}
}
\value{
A single \code{sf} object containing rows for the requested sites.
}
\description{
This function row binds the \code{sites_legacy}, \code{sites_base},
\code{sites_over}, and \code{sites_near} objects from a GRTS or IRS sample
into a single \code{sf} object. This function is most useful when a single
\code{sf} object is desired; e.g. writing the sites to a single shape file the
\code{st_write} function in the sf package.
}
\examples{
sample <- grts(NE_Lakes, 50, n_over = 10)
sample <- sprbind(sample)
\dontrun{st_write(sample, "mypath/sample.shp")}
}
\author{
Michael Dumelle \email{Dumelle.Michael@epa.gov}
}
