% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spsurvey.R
\docType{package}
\name{spsurvey-package}
\alias{spsurvey}
\alias{spsurvey-package}
\title{spsurvey: Spatial Survey Design and Analysis}
\description{
The spsurvey package provides functions for selecting and analyzing
probability samples. To select spatially balanced probability samples,
spsurvey implements the Generalized Random Tessellation Stratified (GRTS)
algorithm.  The GRTS algorithm can be applied to finite resources (point
geometries) And infinite resources (linear and polygon / areal geometries)
and flexibly accommodates a diverse set of sampling design features,
including stratification, unequal inclusion probabilities, proportional
(to size) inclusion probabilities, legacy (historical) sites, a minimum
distance between sites, and two options for replacement sites. To analyze
probability samples, spsurvey implements a wide range of analysis functions
that cover categorical variable analysis, continuous variable analysis,
relative risk analysis, attributable risk analysis, risk difference
analysis, change analysis, and trend analysis. spsurvey can also be used to
plot and summarize data, select samples that are not spatially balanced,
compute spatial balance metrics, select panel samples, and more.
}
\author{
\strong{Maintainer}: Michael Dumelle \email{Dumelle.Michael@epa.gov} (\href{https://orcid.org/0000-0002-3393-5529}{ORCID})

Authors:
\itemize{
  \item Tom Kincaid \email{Kincaid.Tom@epa.gov}
  \item Tony Olsen \email{Olsen.Tony@epa.gov}
  \item Marc Weber \email{Weber.Marc@epa.gov}
}

Other contributors:
\itemize{
  \item Don Stevens [contributor]
  \item Denis White [contributor]
}

}
