% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.sframe}
\alias{summary.dframe}
\alias{summary.spdesign}
\title{Summarize sample frames, samples, and design frames.}
\usage{
\method{summary}{sframe}(object, formula, onlyshow = NULL, ...)

\method{summary}{dframe}(object, formula, onlyshow = NULL, ...)

\method{summary}{spdesign}(object, formula = ~siteuse, onlyshow = NULL, siteuse = NULL, ...)
}
\arguments{
\item{object}{A sample frame object having class \code{sframe} or a design object
output from \code{grts()} or \code{irs()} having class \code{design}.}

\item{formula}{A formula. Left-hand side variables can be numeric or
categorical (or factor) and right-hand side variables can be categorical
(or factor). Right-hand side variables that are numeric will be coerced
to a categorical (or factor) variable. If an intercept is included in the
right-hand side formula, the total will also be summarized.}

\item{onlyshow}{A string indicating the single level of the single right-hand side
variable for which a summary is requested. This argument is only used when
a single right-hand side variable is provided.}

\item{...}{Additional arguments to pass to \code{summary()}. If the left-hand
side of the formula is empty, the appropriate generic arguments are passed
to \code{summary.data.frame}. If the left-hand side of the formula is provided,
the appropriate generic arguments are passed to \code{summary.default}.}

\item{siteuse}{A character vector indicating the \code{design} sites
for which summaries are requested in \code{object}. Defaults to computing summaries for
each non-\code{NULL} \code{site_*} list in \code{object}.}
}
\value{
If the left-hand side of the formula is empty, a named list
containing summaries of the count distribution for each right-hand side
variable is returned. If the left-hand side of the formula contains a
variable, a named list containing five number
summaries (numeric left-hand side) or tables (categorical or factor left
hand side) is returned for each right-hand side variable.
}
\description{
\code{summary()} summarizes sample frames, samples or design objects, depending on
which is provided. The right-hand of the
formula specifies the variables (or factors) to
summarize by. If the left-hand side of the formula is empty, the
summary will be of the distributions of the right-hand side variables. If the left-hand side
of the formula contains a variable, the summary will be of the left-hand size variable
for each level of each right-hand side variable.
}
\examples{
NE_Lakes <- sframe(NE_Lakes)
summary(NE_Lakes, ELEV ~ 1)
summary(NE_Lakes, ~ ELEV_CAT * AREA_CAT)
sample <- grts(NE_Lakes, 100)
summary(sample, ~ ELEV_CAT * AREA_CAT)
}
\author{
Michael Dumelle \email{Dumelle.Michael@epa.gov}
}
