% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPpivotal.R
\name{UPpivotal}
\alias{UPpivotal}
\title{This function is from the R library "sampling". It selects an unequal probability
sample from a finite population.}
\usage{
UPpivotal(pik, eps = 1e-06)
}
\arguments{
\item{pik}{Vector of inclusion probabilities for the sample frame}
}
\value{
A vector of "0" and "1" that indicates if point in the sample frame is
  in the sample.
}
\description{
This function is from the R library "sampling". It selects an unequal probability
sample from a finite population.
}
\author{
From R package sampling
}
\keyword{survey}
